### 2025-06-18 #8425 @editorialbot
[PRE REVIEW]: pyTMD: Python-based tidal prediction software
**Submitting author:** @tsutterley (<a href="http://orcid.org/0000-0002-6964-1194">Tyler Sutterley</a>)
**Repository:** https://github.com/pyTMD/pyTMD
**Branch with paper.md** (empty if default branch): paper
**Version:** 2.2.4
