### 2025-06-19 2337939
feat: added function to calculate x and y coordinates of ellipse
