+++ doc/source/_assets/pytmd-refs.bib
@@ -608,7 +608,6 @@ title = {The Rotation of the Earth: A Geophysical Discussion},
 year = {1960},
 isbn = {9780521104067},
 url = {http://www.cambridge.org/9780521104067},
-doi = {10.1017/CB9780521104067},
 publisher = {Cambridge University Press},
 address = {New York}
 }
