### 2025-06-19 ad89494
test: add ellipse xy to math tests
