+++ pyTMD/ellipse.py
@@ -13,10 +13,10 @@ INPUTS:
     v: meridional current (NS)
 
 OUTPUTS:
-    umajor: amplitude of the semimajor semi-axis
-    uminor: amplitude of the semiminor semi-axis
-    uincl: angle of inclination of the northern semimajor semi-axis
-    uphase: phase lag of the maximum current behind the maximum tidal potential
+    major: amplitude of the semimajor semi-axis
+    minor: amplitude of the semiminor semi-axis
+    incl: angle of inclination of the northern semimajor semi-axis
+    phase: phase lag of the maximum current behind the maximum tidal potential
         of the individual constituent
 
 REFERENCE:
@@ -57,13 +57,13 @@ def ellipse(u: np.ndarray, v: np.ndarray):
 
     Returns
     -------
-    umajor: np.ndarray
+    major: np.ndarray
         amplitude of the semi-major axis
-    uminor: np.ndarray
+    minor: np.ndarray
         amplitude of the semi-minor axis
-    uincl: np.ndarray
+    incl: np.ndarray
         angle of inclination of the northern semi-major axis
-    uphase: np.ndarray
+    phase: np.ndarray
         phase lag of the maximum current behind the maximum tidal potential
     """
     # validate inputs
@@ -80,24 +80,24 @@ def ellipse(u: np.ndarray, v: np.ndarray):
     em = np.angle(wm, deg=True)
     # determine the amplitudes of the semimajor and semiminor axes
     # using Foreman's formula
-    umajor = (ap + am)
-    uminor = (ap - am)
+    major = (ap + am)
+    minor = (ap - am)
     # determine the inclination and phase using Foreman's formula
-    uincl = (em + ep)/2.0
-    uphase = (em - ep)/2.0
+    incl = (em + ep)/2.0
+    phase = (em - ep)/2.0
     # adjust orientation of ellipse
-    k = (uincl//180.0)
-    uincl -= 180.0*k
-    uphase += 180.0*k
-    uphase = np.mod(uphase, 360.0)
+    k = (incl//180.0)
+    incl -= 180.0*k
+    phase += 180.0*k
+    phase = np.mod(phase, 360.0)
     # return values
-    return (umajor, uminor, uincl, uphase)
+    return (major, minor, incl, phase)
 
 def inverse(
-        umajor: np.ndarray,
-        uminor: np.ndarray,
-        uincl: np.ndarray,
-        uphase: np.ndarray
+        major: np.ndarray,
+        minor: np.ndarray,
+        incl: np.ndarray,
+        phase: np.ndarray
     ):
     """
     Calculates currents u, v using the four tidal ellipse
@@ -105,13 +105,13 @@ def inverse(
 
     Parameters
     ----------
-    umajor: np.ndarray
+    major: np.ndarray
         amplitude of the semi-major axis
-    uminor: np.ndarray
+    minor: np.ndarray
         amplitude of the semi-minor axis
-    uincl: np.ndarray
+    incl: np.ndarray
         angle of inclination of the northern semi-major axis
-    uphase: np.ndarray
+    phase: np.ndarray
         phase lag of the maximum current behind the maximum tidal potential
 
     Returns
@@ -122,17 +122,17 @@ def inverse(
         meridional current (NS)
     """
     # validate inputs
-    umajor = np.atleast_1d(umajor)
-    uminor = np.atleast_1d(uminor)
+    major = np.atleast_1d(major)
+    minor = np.atleast_1d(minor)
     # convert inclination and phase to radians
-    uincl = np.atleast_1d(uincl)*np.pi/180.0
[…]
