### 2025-06-20 04af7f1
Bug fix to correctly reference local variable for currents
