### 2025-06-20 3d19461
Prevent `extra_databases` from overwriting main database, support dictionary databases (#419)
* Prevent `extra_databases` from overwriting main database
* Add support for dictionary-format `extra_databases`
* Add sample extra database to tests
* Add test for correct reading of extra databases
* Update test to include dictionary extra databases
