### 2025-06-20 8cd3daa
fix: pass **kwargs to load_database
fix: verify all height and time variables are finite

Update .gitignore
