### 2025-06-20 9bddaf2
Minor Bug fixes: loading FES currents and nodal_modulation documentation (#420)
* Bug fix to correctly reference local variable for currents
* Correct documentation for nodal_modulation return values
