### 2025-06-20 d821008
Update model.py
