### 2025-06-20 9042a77
fix: pass **kwargs to load_database (#421)
* fix: verify all height and time variables are finite
* chore: update .gitignore
