+++ pyTMD/io/model.py
@@ -12,6 +12,8 @@ PYTHON DEPENDENCIES:
 
 UPDATE HISTORY:
     Updated 06/2025: add function for reducing list of model files
+        fix extra_databases to not overwrite the default database
+        add capability to use a dictionary to expand the model database
     Updated 02/2025: fixed missing grid kwarg for reading from TMD3 models 
     Updated 11/2024: use Love numbers for long-period tides in node equilibrium
     Updated 10/2024: add wrapper functions to read and interpolate constants
@@ -125,18 +127,26 @@ def load_database(extra_databases: list = []):
     # extract JSON data
     with database.open(mode='r', encoding='utf-8') as fid:
         parameters = json.load(fid)
+    # verify that extra_databases is iterable
+    if isinstance(extra_databases, (str, pathlib.Path, dict)):
+        extra_databases = [extra_databases]
     # load any additional databases
     for db in extra_databases:
         # use database parameters directly if a dictionary
         if isinstance(db, dict):
-            extra = db
+            extra_database = copy.copy(db)
         # otherwise load parameters from JSON file path
         else:
-            with open(db, 'r', encoding='utf-8') as fid:
-                extra = json.load(fid)
+            # verify that extra database file exists
+            db = pathlib.Path(db).expanduser().absolute()
+            if not db.exists():
+                raise FileNotFoundError(db)
+            # extract JSON data
+            with db.open(mode='r', encoding='utf-8') as fid:
+                extra_database = json.load(fid)
         # Add additional models to database, accounting
         # for top level elevation and current dict keys
-        for key, val in extra.items():
+        for key, val in extra_database.items():
             parameters[key].update(val)
     return DataBase(**parameters)
 
@@ -167,6 +177,8 @@ class model:
         HDF5 ``description`` attribute string for output tide heights
     directory: str, pathlib.Path or None, default None
         Working data directory for tide models
+    extra_databases: list, default []
+        Additional databases for model parameters
     file_format: str
         File format for model
     flexure: bool
@@ -215,6 +227,7 @@ class model:
         # set default keyword arguments
         kwargs.setdefault('compressed', False)
         kwargs.setdefault('verify', True)
+        kwargs.setdefault('extra_databases', [])
         # set initial attributes
         self.compressed = copy.copy(kwargs['compressed'])
         self.constituents = None
@@ -223,6 +236,8 @@ class model:
         self.directory = None
         if directory is not None:
             self.directory = pathlib.Path(directory).expanduser()
+        # set any extra databases
+        self.extra_databases = copy.copy(kwargs['extra_databases'])
         self.flexure = False
         self.format = None
         self.grid_file = None
@@ -236,7 +251,7 @@ class model:
         self.verify = copy.copy(kwargs['verify'])
         self.version = None
 
-    def elevation(self, m: str, extra_databases: list = []):
+    def elevation(self, m: str):
         """
         Create a model object from known tidal elevation models
 
@@ -244,14 +259,12 @@ class model:
         ----------
         m: str
             model name
-        extra_databases: list, default []
-            Additional databases to load
         """
         # set working data directory if unset
         if self.directory is None:
             self.directory = pathlib.Path().absolute()
         # select between known tide models
-        parameters = load_database(extra_databases=extra_databases)
+        parameters = load_database(extra_databases=self.extra_databases)
         # try to extract parameters for model
         try:
             self.from_dict(parameters['elevation'][m])
@@ -269,7 +282,7 @@ class model:
         self.validate_format()
         return self
 
-    def current(self, m: str, extra_databases: list = []):
+    def current(self, m: str):
         """
         Create a model object from known tidal current models
 
@@ -277,14 +290,12 @@ class model:
         ----------
         m: str
             model name
[…]
