### 2025-06-23 bccd8be
feat: add function to get the x/y of a tidal ellipse
ci: add paths for running workflows
