+++ doc/source/release_notes/release-v2.1.1.rst
@@ -2,7 +2,7 @@
 `Release v2.1.1`__
 ##################
 
-* ``refactor``: made the inferrence of minor constituents an option (`#272 <https://github.com/pyTMD/pyTMD/pull/272>`_)
+* ``refactor``: made the inference of minor constituents an option (`#272 <https://github.com/pyTMD/pyTMD/pull/272>`_)
 * ``refactor``: 1-liners in ``compute.py`` (`#272 <https://github.com/pyTMD/pyTMD/pull/272>`_)
 * ``refactor``: lunisolar ephemerides functions (`#272 <https://github.com/pyTMD/pyTMD/pull/272>`_)
 * ``feat``: added more constituent parameters for OTIS/ATLAS predictions (`#272 <https://github.com/pyTMD/pyTMD/pull/272>`_)
