### 2025-06-23 12128f4
feat: add function to get the x/y of a tidal ellipse (#422)
ci: add paths for running workflows
