+++ .github/workflows/python-request.yml
@@ -5,6 +5,11 @@ name: Python on pull request
 
 on:
   pull_request:
+    paths:
+      - pyTMD
+      - scripts
+      - test
+      - .github/workflows/python-request.yml
   schedule:
     # * is a special character in YAML so you have to quote this string
     - cron:  '0 0 1 * *'
