### 2025-06-23 f363394
chore: bump version to `2.2.5`
