+++ doc/source/release_notes/release-v2.2.5.rst
@@ -1,0 +1,21 @@
+##################
+`Release v2.2.5`__
+##################
+
+* ``feat``: add option to include higher order polynomials in solution `(#411) <https://github.com/pyTMD/pyTMD/pull/411>`_
+* ``feat``: add option to reduce to a set of constituents for `#412 <https://github.com/pyTMD/pyTMD/issues/412>`_ `(#413) <https://github.com/pyTMD/pyTMD/pull/413>`_
+* ``fix``: Prevent ``extra_databases`` from overwriting main database `(#419) <https://github.com/pyTMD/pyTMD/pull/419>`_
+* ``feat``: Add support for dictionary-format ``extra_databases`` `(#419) <https://github.com/pyTMD/pyTMD/pull/419>`_
+* ``test``: Add sample extra database to tests `(#419) <https://github.com/pyTMD/pyTMD/pull/419>`_
+* ``test``: Add test for correct reading of extra databases `(#419) <https://github.com/pyTMD/pyTMD/pull/419>`_
+* ``test``: Update test to include dictionary extra databases `(#419) <https://github.com/pyTMD/pyTMD/pull/419>`_
+* ``fix``: loading FES currents and ``nodal_modulation`` documentation `(#420) <https://github.com/pyTMD/pyTMD/pull/420>`_
+* ``fix``: correctly reference local variable for FES currents `(#420) <https://github.com/pyTMD/pyTMD/pull/420>`_
+* ``fix``: Correct documentation for ``nodal_modulation`` return values `(#420) <https://github.com/pyTMD/pyTMD/pull/420>`_
+* ``fix``: pass ``**kwargs`` to load_database `(#421) <https://github.com/pyTMD/pyTMD/pull/421>`_
+* ``fix``: verify all height and time variables are finite `(#421) <https://github.com/pyTMD/pyTMD/pull/421>`_
+* ``chore``: update ``.gitignore`` `(#421) <https://github.com/pyTMD/pyTMD/pull/421>`_
+* ``feat``: add function to get the x/y of a tidal ellipse `(#422) <https://github.com/pyTMD/pyTMD/pull/422>`_
+* ``ci``: add paths for running workflows `(#422) <https://github.com/pyTMD/pyTMD/pull/422>`_
+
+.. __: https://github.com/pyTMD/pyTMD/releases/tag/2.2.5
