### 2025-06-27 c239fdd
docs: add doi for Kantha and Clayson
