### 2025-06-27 f2dcf98
docs: add doi for Kantha and Clayson
