+++ doc/source/_assets/pytmd-refs.bib
@@ -452,6 +452,8 @@ title = {{Numerical models of oceans and oceanic processes}},
 year = {2000},
 volume = {66},
 isbn = {0124340687},
+url = {https://doi.org/10.1016/s0074-6142(00)x8001-1},
+doi = {10.1016/s0074-6142(00)x8001-1},
 publisher = {Academic Press},
 address = {San Diego, CA},
 }
