### 2025-07-10 ad6a792
codespell (#423)
Thanks for the typo fixing 😁
