### 2025-07-10 7532cb1
fix: revert free-to-mean conversion to 2023 version
