### 2025-07-10 2d1bfdd
ci: fix paths
