+++ .github/workflows/update-provider-database.yml
@@ -5,8 +5,8 @@ name: Update provider database on pull request
 on:
   pull_request:
     paths:
-      - pyTMD
-      - providers
+      - pyTMD/**
+      - providers/**
       - .github/workflows/update-provider-database.yml
 
 jobs:
