### 2025-07-16 cd86c6f
docs: autogenerate model table
docs: include constituent table code for #425
