### 2025-07-16 f72cbe4
docs: autogenerate model table (#426)
docs: include constituent table code for #425
