+++ doc/source/getting_started/Getting-Started.rst
@@ -38,66 +38,9 @@ Directories
 This structure was chosen based on the different formats of each tide model.
 Presently, the following models and their directories are parameterized within ``pyTMD``.
 
-- Circum-Antarctic Tidal Simulations :cite:p:`Padman:2008ec`
-
-    * CATS0201: ``<path_to_tide_models>/cats0201_tmd/``
-    * `CATS2008 <https://doi.org/10.15784/601235>`_: ``<path_to_tide_models>/CATS2008/``
-    * CATS2008_load: ``<path_to_tide_models>/CATS2008a_SPOTL_Load/``
-    * `CATS2008-v2023 <https://doi.org/10.15784/601772>`_: ``<path_to_tide_models>/CATS2008_v2023/``
-
-- Arctic Ocean and Greenland Coast Tidal Simulations :cite:p:`Padman:2004hv`
-
-    * `AODTM-5 <https://arcticdata.io/catalog/view/doi:10.18739/A2901ZG3N>`_: ``<path_to_tide_models>/aodtm5_tmd/``
-    * `AOTIM-5 <https://arcticdata.io/catalog/view/doi:10.18739/A2S17SS80>`_: ``<path_to_tide_models>/aotim5_tmd/``
-    * `AOTIM-5-2018 <https://arcticdata.io/catalog/view/doi:10.18739/A21R6N14K>`_: ``<path_to_tide_models>/Arc5km2018/``
-    * `Arc2kmTM <https://arcticdata.io/catalog/view/doi:10.18739/A2D21RK6K>`_: ``<path_to_tide_models>/Arc2kmTM/``
-    * Gr1km-v2: ``<path_to_tide_models>/greenlandTMD_v2/``
-
-- TOPEX/POSEIDON global tide models :cite:p:`Egbert:2002ge`
-
-    * TPXO7.2: ``<path_to_tide_models>/TPXO7.2_tmd/``
-    * TPXO7.2_load: ``<path_to_tide_models>/TPXO7.2_load/``
-    * `TPXO8-atlas <https://www.tpxo.net/tpxo-products-and-registration>`_: ``<path_to_tide_models>/tpxo8_atlas/``
-    * `TPXO9.1 <https://www.tpxo.net/tpxo-products-and-registration>`_: ``<path_to_tide_models>/TPXO9.1/DATA/``
-    * `TPXO9-atlas <https://www.tpxo.net/tpxo-products-and-registration>`_: ``<path_to_tide_models>/TPXO9_atlas/``
-    * `TPXO9-atlas-v2 <https://www.tpxo.net/tpxo-products-and-registration>`_: ``<path_to_tide_models>/TPXO9_atlas_v2/``
-    * `TPXO9-atlas-v3 <https://www.tpxo.net/tpxo-products-and-registration>`_: ``<path_to_tide_models>/TPXO9_atlas_v3/``
-    * `TPXO9-atlas-v4 <https://www.tpxo.net/tpxo-products-and-registration>`_: ``<path_to_tide_models>/TPXO9_atlas_v4/``
-    * `TPXO9-atlas-v5 <https://www.tpxo.net/tpxo-products-and-registration>`_: ``<path_to_tide_models>/TPXO9_atlas_v5/``
-    * `TPXO10-atlas-v2 <https://www.tpxo.net/tpxo-products-and-registration>`_: ``<path_to_tide_models>/TPXO10_atlas_v2/``
-
-- Global Ocean Tide models :cite:p:`Ray:1999vm`
-
-    * GOT4.7: ``<path_to_tide_models>/GOT4.7/grids_oceantide/``
-    * GOT4.7_load: ``<path_to_tide_models>/GOT4.7/grids_loadtide/``
[…]
