### 2025-07-16 f73ef8d
docs: add back citations to models
