### 2025-07-17 cb3d318
docs: add back citations to models (#427)
