### 2025-07-21 pyTMD/pyTMD#428 @rcaneill
[JOSS review] error in notebook "Check-Tide-Map.ipynb"
Hi,

in the process of review for your JOSS submission, I tried to execute the following notebook, but there is an error:

```
Cell In[16], line 10
      8 import pyTMD.crs
      9 import pyTMD.io
---> 10 import pyTMD.time
     11 import pyTMD.tools
     12 # autoreload

ModuleNotFoundError: No module named 'pyTMD.time'
```

https://github.com/pyTMD/pyTMD/blob/cb3d318f71e587ab508f53824400e6c832a54d3e/doc/source/notebooks/Check-Tide-Map.ipynb#L67

I commented the line and it worked until the last cell where I got another error:

```
RuntimeError                              Traceback (most recent call last)
File [~/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/matplotlib/backends/registry.py:407](http://localhost:8889/home/romain/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/matplotlib/backends/registry.py#line=406), in BackendRegistry.resolve_gui_or_backend(self, gui_or_backend)
    406 try:
--> 407     return self.resolve_backend(gui_or_backend)
    408 except Exception:  # KeyError ?

File [~/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/matplotlib/backends/registry.py:369](http://localhost:8889/home/romain/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/matplotlib/backends/registry.py#line=368), in BackendRegistry.resolve_backend(self, backend)
    368 if gui is None:
--> 369     raise RuntimeError(f"'{backend}' is not a recognised backend name")
    371 return backend, gui if gui != "headless" else None

RuntimeError: 'widget' is not a recognised backend name

RuntimeError                              Traceback (most recent call last)
Cell In[15], line 1
----> 1 get_ipython().run_line_magic('matplotlib', 'widget')
      2 # check coordinates on tide grid
      3 m.figure,ax = plt.subplots(num=1, figsize=(8.25,5.25))

File [~/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/IPython/core/interactiveshell.py:2504](http://localhost:8889/home/romain/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/IPython/core/interactiveshell.py#line=2503), in InteractiveShell.run_line_magic(self, magic_name, line, _stack_depth)
   2502     kwargs['local_ns'] = self.get_local_scope(stack_depth)
   2503 with self.builtin_trap:
-> 2504     result = fn(*args, **kwargs)
   2506 # The code below prevents the output from being displayed
   2507 # when using magics with decorator @output_can_be_silenced
   2508 # when the last Python token in the expression is a ';'.
   2509 if getattr(fn, magic.MAGIC_OUTPUT_CAN_BE_SILENCED, False):

File [~/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/IPython/core/magics/pylab.py:103](http://localhost:8889/home/romain/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/IPython/core/magics/pylab.py#line=102), in PylabMagics.matplotlib(self, line)
     98     print(
     99         "Available matplotlib backends: %s"
    100         % _list_matplotlib_backends_and_gui_loops()
    101     )
    102 else:
--> 103     gui, backend = self.shell.enable_matplotlib(args.gui)
    104     self._show_matplotlib_backend(args.gui, backend)

File [~/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/IPython/core/interactiveshell.py:3787](http://localhost:8889/home/romain/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/IPython/core/interactiveshell.py#line=3786), in InteractiveShell.enable_matplotlib(self, gui)
   3784     import matplotlib_inline.backend_inline
   3786 from IPython.core import pylabtools as pt
-> 3787 gui, backend = pt.find_gui_and_backend(gui, self.pylab_gui_select)
   3789 if gui != None:
   3790     # If we have our first gui selection, store it
   3791     if self.pylab_gui_select is None:

File [~/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/IPython/core/pylabtools.py:349](http://localhost:8889/home/romain/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/IPython/core/pylabtools.py#line=348), in find_gui_and_backend(gui, gui_select)
    347 else:
    348     gui = _convert_gui_to_matplotlib(gui)
--> 349     backend, gui = backend_registry.resolve_gui_or_backend(gui)
    351 gui = _convert_gui_from_matplotlib(gui)
    352 return gui, backend

File [~/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/matplotlib/backends/registry.py:409](http://localhost:8889/home/romain/.cache/pypoetry/virtualenvs/review-pytmd-pD7IGWct-py3.12/lib/python3.12/site-packages/matplotlib/backends/registry.py#line=408), in BackendRegistry.resolve_gui_or_backend(self, gui_or_backend)
    407     return self.resolve_backend(gui_or_backend)
    408 except Exception:  # KeyError ?
--> 409     raise RuntimeError(
    410         f"'{gui_or_backend}' is not a recognised GUI loop or backend name")

RuntimeError: 'widget' is not a recognised GUI loop or backend name
```
