### 2025-07-21 #8566 @rcaneill
Hi @tsutterley, I continue on the review. Here are some points that should be fixed:

* [ ] **Functionality/Functionality** and **Documentation/automated tests**
I ran the test suite, but there seems to be an error with the credentials (from what I read in the github action workflow they are probably stored in the "secrets", but being able to run the tests locally should be possible). A paragraph on how to run the test locally could be added in the documentation
* [ ] Could you explain why the notebooks are not rendered in the documentation? It would 1) allow the user to see what the package is doing and 2) test the notebooks at the same time. I understand that it is not reasonable to test all the 50 models for every notebook, but maybe select 1 light model that'd be downloaded at build time is a good option.
* [ ] Following, I opened an issue as one of the notebook does not work (I did not test all of them), cf https://github.com/pyTMD/pyTMD/issues/428
* [ ] **Documentation/Statement of need** / **Software paper/statement of need**
I went through both the documentation and the paper and one point is not clear: what do the models compute? And what does pyTMD does to these data?
The way it is written now makes it look like it is a wrapper that simply open the data from all these models.
I suggest to add this clarification into the "Statement of need" section
* [ ] **Documentation/example usage**
The provided notebooks are a kind of example usage. However, as they are not rendered and quite hard to follow (due to the fact that they are very generic and handle every model possibility), the usage is not clear.
* [ ] **Software paper/summary**
The provided summary is clear and concise. However, 1 or 2 sentences could be added to be a little more precise (cf also my comment on statement of need).
* [ ] **Software paper/State of the field**
Except a brief mention in the acknowledgement section, this is not described. Could you please add it?

I will carefully check the last points later, from a 1st look both "quality" and "references" look good. I also take the opportunity to congratulate you on the docstrings of the functions.

I will go through a 2nd round of review when you’ll have answered to/fixed my comments
