### 2025-07-21 b134199
chore: removed functions that were moved to timescale
docs: add table of tidal current directories
ci: remove earthdata credentials
