+++ pyTMD/utilities.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 utilities.py
-Written by Tyler Sutterley (01/2025)
+Written by Tyler Sutterley (07/2025)
 Download and management utilities for syncing time and auxiliary files
 
 PYTHON DEPENDENCIES:
@@ -9,6 +9,7 @@ PYTHON DEPENDENCIES:
         https://pypi.python.org/pypi/lxml
 
 UPDATE HISTORY:
+    Updated 07/2025: removed (now) unused functions that were moved to timescale
     Updated 01/2025: added function to list a directory from the UHSLC
     Updated 08/2024: generalize hash function to use any available algorithm
     Updated 07/2024: added function to parse JSON responses from https
@@ -51,17 +52,13 @@ import re
 import io
 import ssl
 import json
-import netrc
 import ftplib
 import shutil
-import base64
 import socket
-import getpass
 import inspect
 import hashlib
 import logging
 import pathlib
-import builtins
 import warnings
 import importlib
 import posixpath
@@ -70,12 +67,8 @@ import lxml.etree
 import calendar, time
 import dateutil.parser
 if sys.version_info[0] == 2:
-    from urllib import quote_plus
-    from cookielib import CookieJar
     import urllib2
 else:
-    from urllib.parse import quote_plus
-    from http.cookiejar import CookieJar
     import urllib.request as urllib2
 
 __all__ = [
@@ -88,11 +81,8 @@ __all__ = [
     "url_split",
     "convert_arg_line_to_args",
     "build_logger",
-    "roman_to_int",
     "get_unix_time",
-    "isoformat",
     "even",
-    "ceil",
     "copy",
     "check_ftp_connection",
     "ftp_list",
@@ -104,14 +94,6 @@ __all__ = [
     "http_list",
     "from_http",
     "from_json",
-    "attempt_login",
-    "build_opener",
-    "get_token",
-    "list_tokens",
-    "revoke_token",
-    "check_credentials",
-    "cddis_list",
-    "from_cddis",
     "iers_list",
     "from_jpl_ssd",
     "uhslc_list"
@@ -358,30 +340,6 @@ def build_logger(name: str, **kwargs):
         logger.addHandler(handler)
     return logger
 
-# PURPOSE: convert Roman numerals to (Arabic) integers
-def roman_to_int(roman: str):
-    """
-    Converts a string from Roman numerals into an integer (Arabic)
-
-    Parameters
-    ----------
-    roman: str
-        Roman numeral string
-    """
-    # mapping between Roman and Arabic numerals
-    roman_map = {'i':1, 'v':5, 'x':10, 'l':50, 'c':100, 'd':500, 'm':1000}
-    # verify case
-    roman = roman.lower()
-    output = 0
-    # iterate through roman numerals in string and calculate total
-    for i,s in enumerate(roman):
-        if (i > 0) and (roman_map[s] > roman_map[roman[i-1]]):
-            output += roman_map[s] - 2*roman_map[roman[i-1]]
-        else:
-            output += roman_map[s]
-    # return the integer value
-    return output
-
 # PURPOSE: returns the Unix timestamp value for a formatted date string
 def get_unix_time(
         time_string: str,
@@ -411,24 +369,6 @@ def get_unix_time(
     else:
         return parsed_time.timestamp()
 
-# PURPOSE: output a time string in isoformat
-def isoformat(time_string: str):
-    """
-    Reformat a date string to ISO formatting
-
-    Parameters
-    ----------
-    time_string: str
-        formatted time string to parse
-    """
-    # try parsing with dateutil
-    try:
-        parsed_time = dateutil.parser.parse(time_string.rstrip())
-    except (TypeError, ValueError):
[…]
