+++ doc/source/background/Reference-Systems.rst
@@ -9,7 +9,7 @@ The :math:`xy` plane is also called the equatorial plane.
 
 Geodetic coordinates (longitude :math:`\lambda`, latitude :math:`\varphi`, and height :math:`h`) are used to describe the position of a point on the Earth with respect to a defined ellipsoid.
 Changing the terrestrial reference system can involve both translations and rotations of the reference system :cite:p:`Urban:2013vl`.
-This involves converting from a geographic coordinate system into a Cartesian coordinate system, and then performing matrix transformations [:ref:`Equation 2.1 <eq:2.1>`].
+One method of transformation involves converting from a geographic coordinate system into a Cartesian coordinate system, and then performing matrix transformations [:ref:`Equation 2.1 <eq:2.1>`].
 
 .. math::
     :label: 2.1

[…]
