### 2025-07-21 f498cd6
Update utilities.py
