### 2025-07-21 11b9130
chore: removed functions that were moved to timescale (#429)
docs: add table of tidal current directories
ci: remove earthdata credentials
