+++ doc/source/getting_started/Current-Table.rst
@@ -1,0 +1,10 @@
+:orphan:
+
+Tidal Current Directories
+#########################
+
+.. _tab-currents:
+
+.. csv-table:: Tidal Current Directories
+   :file: ../_assets/current-models.csv
+   :header-rows: 1
