### 2025-07-24 9a8fe00
fix: revert load pole tide to IERS 1996 convention definitions (#430)
https://ilrs.gsfc.nasa.gov/docs/1996/iers_1996_conventions.pdf
