### 2025-07-24 d81c676
fix: mask mean pole values prior to valid epoch of convention
