+++ pyTMD/compute.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 compute.py
-Written by Tyler Sutterley (05/2025)
+Written by Tyler Sutterley (07/2025)
 Calculates tidal elevations for correcting elevation or imagery data
 Calculates tidal currents at locations and times
 
@@ -62,6 +62,7 @@ PROGRAM DEPENDENCIES:
     interpolate.py: interpolation routines for spatial data
 
 UPDATE HISTORY:
+    Updated 07/2025: mask mean pole values prior to valid epoch of convention
     Updated 05/2025: added option to select constituents to read from model
     Updated 12/2024: moved check points function as compute.tide_masks
     Updated 11/2024: expose buffer distance for cropping tide model data
@@ -1064,9 +1065,10 @@ def LPT_displacements(
             )
             # calculate components of load pole tides
             S = np.einsum('ti...,tji...->tj...', dxi, R)
+            smask = np.reshape(np.any(dxi.mask, axis=1), (ny,nx))
             # reshape to output dimensions
             Srad.data[:,:,i] = np.reshape(S[:,2], (ny,nx))
-            Srad.mask[:,:,i] = np.isnan(Srad.data[:,:,i])
+            Srad.mask[:,:,i] = np.isnan(Srad.data[:,:,i]) | smask
     elif (TYPE == 'drift'):
         # calculate load pole tides in cartesian coordinates
         XYZ = np.c_[X, Y, Z]
@@ -1080,10 +1082,11 @@ def LPT_displacements(
         )
         # calculate components of load pole tides
         S = np.einsum('ti...,tji...->tj...', dxi, R)
+        smask = np.any(dxi.mask, axis=1)
         # reshape to output dimensions
         Srad = np.ma.zeros((nt), fill_value=FILL_VALUE)
         Srad.data[:] = S[:,2].copy()
-        Srad.mask = np.isnan(Srad.data)
+        Srad.mask = np.isnan(Srad.data) | smask
     elif (TYPE == 'time series'):
         nstation = len(x)
         Srad = np.ma.zeros((nstation,nt), fill_value=FILL_VALUE)
@@ -1102,9 +1105,10 @@ def LPT_displacements(
             )
             # calculate components of load pole tides
             S = np.einsum('ti...,ji...->tj...', dxi, R[s,:,:])
+            smask = np.any(dxi.mask, axis=1)
             # reshape to output dimensions
             Srad.data[s,:] = S[:,2].copy()
-            Srad.mask[s,:] = np.isnan(Srad.data[s,:])
+            Srad.mask[s,:] = np.isnan(Srad.data[s,:]) | smask
 
     # replace invalid data with fill values
     Srad.data[Srad.mask] = Srad.fill_value
@@ -1288,9 +1292,10 @@ def OPT_displacements(
             )
             # calculate components of ocean pole tides
             U = np.einsum('ti...,tji...->tj...', dxi, Rinv)
+            umask = np.reshape(np.any(dxi.mask, axis=1), (ny,nx))
             # reshape to output dimensions
             Urad.data[:,:,i] = np.reshape(U[:,2], (ny,nx))
-            Urad.mask[:,:,i] = np.isnan(Urad.data[:,:,i])
+            Urad.mask[:,:,i] = np.isnan(Urad.data[:,:,i]) | umask
     elif (TYPE == 'drift'):
         # calculate ocean pole tides in cartesian coordinates
         XYZ = np.c_[X, Y, Z]
@@ -1306,10 +1311,11 @@ def OPT_displacements(
         )
         # calculate components of ocean pole tides
[…]
