### 2025-07-24 18b0268
fix: mask mean pole values prior to valid epoch of convention (#431)
