### 2025-07-24 5196a35
chore: bump version to `2.2.6`
