+++ doc/source/release_notes/release-v2.2.6.rst
@@ -1,0 +1,18 @@
+##################
+`Release v2.2.6`__
+##################
+
+* ``fix``: ``codespell`` typo fixing `(#423) <https://github.com/pyTMD/pyTMD/pull/423>`_
+* ``feat``: added extrapolation option for FES2014 download `(#424) <https://github.com/pyTMD/pyTMD/pull/424>`_
+* ``fix``: revert SET free-to-mean conversion to 2023 version `(#424) <https://github.com/pyTMD/pyTMD/pull/424>`_
+* ``ci``: fix paths to check if needing to run tests `(#424) <https://github.com/pyTMD/pyTMD/pull/424>`_
+* ``docs``: autogenerate model table `(#426) <https://github.com/pyTMD/pyTMD/pull/426>`_
+* ``docs``: include constituent table code for `#425 <https://github.com/pyTMD/pyTMD/issues/425>`_ `(#426) <https://github.com/pyTMD/pyTMD/pull/426>`_
+* ``docs``: add back citations to models `(#427) <https://github.com/pyTMD/pyTMD/pull/427>`_
+* ``chore``: removed functions that were moved to timescale `(#429) <https://github.com/pyTMD/pyTMD/pull/429>`_
+* ``docs``: add table of tidal current directories `(#429) <https://github.com/pyTMD/pyTMD/pull/429>`_
+* ``ci``: remove earthdata credentials `(#429) <https://github.com/pyTMD/pyTMD/pull/429>`_
+* ``fix``: revert load pole tide to `IERS 1996 convention <https://ilrs.gsfc.nasa.gov/docs/1996/iers_1996_conventions.pdf>`_ definitions `(#430) <https://github.com/pyTMD/pyTMD/pull/430>`_
+* ``fix``: mask mean pole values prior to valid epoch of convention `(#431) <https://github.com/pyTMD/pyTMD/pull/431>`_
+
+.. __: https://github.com/pyTMD/pyTMD/releases/tag/2.2.6
