### 2025-07-28 861e627
docs: try rendering notebooks
