+++ doc/environment.yml
@@ -24,4 +24,4 @@ dependencies:
   - sphinx_rtd_theme
   - timescale>=0.0.8
   - pip:
-    - ..
+    - ..[all]
