### 2025-07-28 ce75cfe
Update environment.yml
