### 2025-07-29 0735abf
docs: filter warnings
