### 2025-07-29 0f186d6
docs: published rendered versions of notebooks for #394 (#432)
