### 2025-07-29 c403807
refactor: move NOAA query functions to module
