+++ pyTMD/io/__init__.py
@@ -6,6 +6,7 @@ from .FES import *
 from .GOT import *
 from .OTIS import *
 from .IERS import *
+from .NOAA import *
 from .constituents import constituents
 from .model import (
     model,
