### 2025-07-29 81aa7ac
test: add NOAA tests
