### 2025-07-29 020662f
add wrapper functions for NOAA queries
