### 2025-07-30 ed990b3
refactor: move NOAA webservices API query functions to module (#433)
