+++ pyTMD/tools.py
@@ -12,12 +12,10 @@ PYTHON DEPENDENCIES:
         https://ipywidgets.readthedocs.io/en/latest/
     ipyleaflet: Jupyter / Leaflet bridge enabling interactive maps
         https://github.com/jupyter-widgets/ipyleaflet
-    matplotlib: Python 2D plotting library
-        http://matplotlib.org/
-        https://github.com/matplotlib/matplotlib
 
 UPDATE HISTORY:
     Updated 07/2025: add a default directory for tide models
+        add full screen control to ipyleaflet maps
     Updated 09/2024: removed widget for ATLAS following database update
         added widget for setting constituent to plot in a cotidal chart
     Updated 07/2024: renamed format for netcdf to ATLAS-netcdf
@@ -34,11 +32,7 @@ UPDATE HISTORY:
     Updated 02/2022: add leaflet map projections
     Written 09/2021
 """
-import os
-import io
 import copy
-import base64
-import logging
 import datetime
 import numpy as np
 import pyTMD.io.model
@@ -49,15 +43,6 @@ IPython = pyTMD.utilities.import_dependency('IPython')
 ipyleaflet = pyTMD.utilities.import_dependency('ipyleaflet')
 ipywidgets = pyTMD.utilities.import_dependency('ipywidgets')
 pyproj = pyTMD.utilities.import_dependency('pyproj')
-matplotlib = pyTMD.utilities.import_dependency('matplotlib')
-plt = pyTMD.utilities.import_dependency('matplotlib.pyplot')
-cm = pyTMD.utilities.import_dependency('matplotlib.cm')
-colors = pyTMD.utilities.import_dependency('matplotlib.colors')
-
-try:
-    matplotlib.rcParams['axes.linewidth'] = 2.0
-except (AttributeError, ImportError, ModuleNotFoundError) as exc:
-    logging.debug("matplotlib not available")
 
 class widgets:
     def __init__(self, **kwargs):
@@ -70,10 +55,12 @@ class widgets:
         self.VBox = ipywidgets.VBox
 
         # default working data directory for tide models
-        d = pyTMD.utilities.compressuser(pyTMD.utilities.get_data_path('data'))
+        default_directory = pyTMD.utilities.compressuser(
+            pyTMD.utilities.get_data_path('data')
+        )
         # set the directory with tide models
         self.directory = ipywidgets.Text(
-            value=str(d),
+            value=str(default_directory),
             description='Directory:',
             disabled=False
         )
@@ -90,7 +77,9 @@ class widgets:
         )
 
         # dropdown menu for setting model constituents
-        constituents_list = ['q1','o1','p1','k1','n2','m2','s2','k2','s1','m4']
+        constituents_list = [
+            'q1','o1','p1','k1','n2','m2','s2','k2','s1','m4'
+        ]
         self.constituents = ipywidgets.Dropdown(
             options=constituents_list,
             value='m2',
@@ -115,81 +104,13 @@ class widgets:
             style=self.style,
         )
 
-# define projections for ipyleaflet tiles
-projections = dict(
-    # Alaska Polar Stereographic (WGS84)
-    EPSG5936 = dict(
-        Basemap = dict(
-            name='EPSG:5936',
-            custom=True,
-            proj4def="""+proj=stere +lat_0=90 +lat_ts=90 +lon_0=-150 +k=0.994
-                +x_0=2000000 +y_0=2000000 +datum=WGS84 +units=m +no_defs""",
-            origin=[-2.8567784109255e+07, 3.2567784109255e+07],
-            resolutions=[
-                238810.813354,
-                119405.406677,
-                59702.7033384999,
-                29851.3516692501,
-                14925.675834625,
-                7462.83791731252,
-                3731.41895865639,
[…]
