+++ doc/source/api_reference/io/io.rst
@@ -14,3 +14,4 @@ Input/output functions for reading and writing tidal data
     constituents.rst
     model.rst
     IERS.rst
+    NOAA.rst

[…]
