### 2025-08-02 367281d
[WIP] add simplified body tides from Cartwright (1971)
used for the solid earth tides with most altimeters: including TOPEX and ICESat (GLAS)
