+++ pyTMD/predict.py
@@ -1065,10 +1065,11 @@ def equilibrium_tide(
     P20 = 0.5*(3.0*np.sin(lat*np.pi/180.0)**2 - 1.0)
     # calculate long-period equilibrium tide and convert to meters
     # Multiply by gamma_2 * normalization * P20(lat)
+    norm = np.sqrt((4.0 + 1.0)/(4.0*np.pi))
     if (nlat != nt):
-        lpet = gamma_2*np.sqrt((4.0+1.0)/(4.0*np.pi))*np.outer(P20,Z/100.0)
+        lpet = gamma_2*norm*np.outer(P20,Z/100.0)
     else:
-        lpet = gamma_2*np.sqrt((4.0+1.0)/(4.0*np.pi))*P20*(Z/100.0)
+        lpet = gamma_2*norm*P20*(Z/100.0)
     # return the long-period equilibrium tides
     return lpet
 
@@ -1324,6 +1325,18 @@ def solid_earth_tide(
 
             - ``'tide_free'``: no permanent direct and indirect tidal potentials
             - ``'mean_tide'``: permanent tidal potentials (direct and indirect)
+    h2: float, default 0.6078
+        Degree-2 Love number of vertical displacement
+    l2: float, default 0.0847
+        Degree-2 Love (Shida) number of horizontal displacement
+    h3: float, default 0.292
+        Degree-3 Love number of vertical displacement
+    l3: float, default 0.015
+        Degree-3 Love (Shida) number of horizontal displacement
+    mass_ratio_solar: float, default 332946.0482
+        Mass ratio between the Earth and the Sun
+    mass_ratio_lunar: float, default 0.0123000371
+        Mass ratio between the Earth and the Moon
 
     Returns
     -------
@@ -1770,3 +1783,163 @@ def _free_to_mean(
     DZ = -dr*sinphi - dn*cosphi
     # return the corrections
     return np.c_[DX, DY, DZ]
+
+def body_tide(
+        t: np.ndarray,
+        lon: np.ndarray,
+        lat: np.ndarray,
+        deltat: float | np.ndarray = 0.0,
+        gamma_0: float = 9.80665,
+        tide_system: str = 'tide_free',
+        **kwargs
+    ):
+    """
+    Compute the solid Earth tides due to the gravitational
+    attraction of the moon and sun using the approach of
+    :cite:t:`Cartwright:1971iz,Cartwright:1973em`
+    
+    Parameters
+    ----------
+    t: np.ndarray
+        Time (days relative to January 1, 1992)
+    lon: np.ndarray
+        longitude (degrees north)
+    lat: np.ndarray
+        latitude (degrees north)
+    deltat: float or np.ndarray, default 0.0
+        time correction for converting to Ephemeris Time (days)
+    gamma_0: float, default 9.80665
+        Normal gravity (m/s^2)
+    tide_system: str, default 'tide_free'
+        Permanent tide system for the output solid Earth tide
+
+            - ``'tide_free'``: no permanent direct and indirect tidal potentials
+            - ``'mean_tide'``: permanent tidal potentials (direct and indirect)
+    h2: float, default 0.609
+        Degree-2 Love number of vertical displacement
+    l2: float, default 0.0852
+        Degree-2 Love (Shida) number of horizontal displacement
+    h3: float, default 0.291
+        Degree-3 Love number of vertical displacement
+    l3: float, default 0.015
+        Degree-3 Love (Shida) number of horizontal displacement
+
+    Returns
+    -------
+    zeta: np.ndarray
+        Solid Earth tide in meters
+    """
+    # set default keyword arguments
[…]

[…]
