### 2025-08-03 pyTMD/pyTMD#434 @castelao
Link README to `Contribution Guidelines`
@tsutterley , you have a nice `Contribution Guidelines` in the documentation. I suggest adding a link from your `README` in the root level. Probably in the [Contributing](https://github.com/pyTMD/pyTMD?tab=readme-ov-file#contributing) section.

This is not a blocker for openjournals/joss-reviews#8566 .
