### 2025-08-03 1591933
Basic pixi setup
