### 2025-08-04 0d05d0e
ci: pin to `macos-14` for now
