+++ doc/source/background/spectra.py
@@ -29,8 +29,7 @@ frange.append([1.75, 2.10])
 l = 2
 # parse the Cartwright and Edden (1973) table for degree-2 constituents
 CTE = pyTMD.arguments._parse_tide_potential_table(
-    pyTMD.arguments._ce1973_table_1,
-    format='Cartwright')
+    pyTMD.arguments._ce1973_table_1)
 # for each spectral line
 for i, line in enumerate(CTE):
     # spherical harmonic dependence (order)
@@ -74,8 +73,7 @@ for i, line in enumerate(CTE):
 l = 3
 # parse the Cartwright and Tayler (1971) table for degree-3 constituents
 CTE = pyTMD.arguments._parse_tide_potential_table(
-    pyTMD.arguments._ct1971_table_5,
-    format='Cartwright')
+    pyTMD.arguments._ct1971_table_5)
 # for each line in the table
 for i, line in enumerate(CTE):
     # spherical harmonic dependence (order)
