+++ pyTMD/arguments.py
@@ -1766,7 +1766,12 @@ def _frequency(
     coef: list or np.ndarray
         Doodson coefficients (Cartwright numbers) for constituents
     method: str, default 'Cartwright'
-        method for calculating astronomical arguments
+        Method for computing the mean longitudes
+
+            - ``'Cartwright'``
+            - ``'Meeus'``
+            - ``'ASTRO5'`` 
+            - ``'IERS'``
 
     Returns
     -------
@@ -1895,18 +1900,19 @@ def _love_numbers(
     return (h2, k2, l2)
 
 # Doodson (1921) table with values missing from Cartwright tables
+# Hs1: amplitude for epoch span 1 (1900 epoch)
 _d1921_table = get_data_path(['data','d1921_tab.txt'])
 # Cartwright and Tayler (1971) table with 3rd-degree values
+# Hs1: amplitude for epoch span 1 (1861-09-21 to 1879-09-22)
+# Hs2: amplitude for epoch span 2 (1915-05-16 to 1933-05-22)
+# Hs3: amplitude for epoch span 2 (1951-05-23 to 1969-05-22)
 _ct1971_table_5 = get_data_path(['data','ct1971_tab5.txt'])
 # Cartwright and Edden (1973) table with updated values
 _ce1973_table_1 = get_data_path(['data','ce1973_tab1.txt'])
-# Woodworth (1990) tables with updated and 3rd-degree values
-_w1990_table_1 = get_data_path(['data','w1990_tab1.txt'])
-_w1990_table_2 = get_data_path(['data','w1990_tab2.txt'])
 
 def _parse_tide_potential_table(
         table: str | pathlib.Path,
-        format: str = 'Cartwright',
+        columns: int = 3,
     ):
     """Parse tables of tide-generating potential
 
@@ -1914,20 +1920,14 @@ def _parse_tide_potential_table(
     ----------
     table: str or pathlib.Path
         table of tide-generating potentials
-    format: str, default 'Cartwright'
-        format for tide-generating potential table
-            - 'Cartwright' (default)
-            - 'Doodson' (1921)
-            - 'Woodworth' (1990)
+    columns: int, default 3
+        number of amplitude columns in the table
 
     Returns
     -------
     CTE: float
         Cartwright-Tayler-Edden table values
     """
-    # set default keyword arguments
-    if format not in ('Cartwright', 'Doodson', 'Woodworth'):
-        raise ValueError(f'Unknown format: {format}')
     # verify table path
     table = pathlib.Path(table).expanduser().absolute()
     with table.open(mode='r', encoding='utf8') as f:
@@ -1940,25 +1940,15 @@ def _parse_tide_potential_table(
     # p: coefficient for mean longitude of lunar perigee
     # n: coefficient for mean longitude of ascending lunar node
     # pp: coefficient for mean longitude of solar perigee
-    if (format == 'Cartwright'):
-        # Hs1: amplitude for epoch span 1 (1861-09-21 to 1879-09-22)
-        # Hs2: amplitude for epoch span 2 (1915-05-16 to 1933-05-22)
-        # Hs3: amplitude for epoch span 2 (1951-05-23 to 1969-05-22)
-        # DO: Doodson number for coefficient
-        # Hs0: Doodson scaled amplitude for 1900
-        names = ('tau','s','h','p','n','pp','Hs1','Hs2','Hs3','DO','Hs0')
-        formats = ('i','i','i','i','i','i','f8','f8','f8','U7','f8')
-    elif (format == 'Doodson'):
-        # Hs1: amplitude for epoch span 1 (1900 epoch)
-        # DO: Doodson number for coefficient
-        # Hs0: Doodson scaled amplitude for 1900
[…]
