+++ .github/workflows/python-request.yml
@@ -21,7 +21,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-14, windows-latest]
-        python-version: [3.11]
+        python-version: [3.13]
     env:
       OS: ${{ matrix.os }}
       PYTHON: ${{ matrix.python-version }}
