### 2025-08-05 59367ff
bump scipy version
