+++ doc/environment.yml
@@ -22,7 +22,7 @@ dependencies:
   - pip
   - python>=3.6
   - python-dateutil
-  - scipy>=1.10.1
+  - scipy>=1.16.0
   - setuptools_scm
   - sphinx
   - sphinx-argparse>=0.4
