### 2025-08-05 bb43876
ci: try bumping back up to macos-latest
