### 2025-08-05 0330575
feat: add simplified body tides from Cartwright (1971) (#436)
* ci: bump python version to `3.13`
* chore: bump scipy version to `1.16.0`

this solid earth model is used as a correction for many altimetry datasets: including TOPEX and ICESat (GLAS)
