+++ test/test_solid_earth.py
@@ -201,3 +201,27 @@ def test_solid_earth_radial(EPHEMERIDES):
     assert np.isclose(-tide_earth_free2mean, predicted, atol=5e-4).all()
     assert np.isclose(tide_mean-tide_free, predicted, atol=5e-4).all()
 
+def test_body_tides():
+    """Test simplified solid tides using predictions from ICESat-2
+    """
+    times = np.array(['2018-10-14 00:21:48','2018-10-14 00:21:48',
+        '2018-10-14 00:21:48','2018-10-14 00:21:48',
+        '2022-07-23 13:53:08','2022-07-23 13:53:08',
+        '2022-07-23 13:53:08','2022-07-23 13:53:08'], dtype=np.datetime64)
+    longitudes = np.array([-136.79534534,-136.79545175,
+        -136.79548250,-136.79549453,-71.77356870,-71.77374742,
+        -71.77392700,-71.77410705])
+    latitudes = np.array([68.95910366,68.95941755,
+        68.95950895,68.95954490,-79.00591611,-79.00609103,
+        -79.00626593,-79.00644081])
+    # expected results (tide-free)
+    tide_earth = np.array([-0.14320290,-0.14320324,
+        -0.14320339,-0.14320345,-0.11887791,-0.11887763,
+        -0.11887724,-0.11887683])
+    # predict tides using simplified body tides
+    # using tide potentials from Cartwright and Tayler (1971)
+    ts = timescale.from_datetime(times)
+    tide_free = pyTMD.predict.body_tide(ts.tide, longitudes, latitudes,
+        deltat=ts.tt_ut1, tide_system='tide_free')
+    # since we are using simplified body tides: assert within 2 mm
+    assert np.isclose(tide_earth, tide_free[:,2], atol=2e-3).all()
