+++ pyTMD/arguments.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 arguments.py
-Written by Tyler Sutterley (04/2025)
+Written by Tyler Sutterley (08/2025)
 Calculates the nodal corrections for tidal constituents
 Modification of ARGUMENTS fortran subroutine by Richard Ray 03/1999
 
@@ -39,6 +39,8 @@ REFERENCES:
         Ocean Tides", Journal of Atmospheric and Oceanic Technology, (2002).
 
 UPDATE HISTORY:
+    Updated 08/2025: make frequency function a wrapper around a function
+        to calculate using Doodson coefficients (Cartwright numbers)
     Updated 04/2025: convert longitudes p and n to radians within nodal function
         use schureman_arguments function to get nodal variables for FES models
         added Schureman to list of M1 options in nodal arguments
@@ -113,6 +115,7 @@ __all__ = [
     "_arguments_table",
     "_minor_table",
     "_constituent_parameters",
+    "_frequency",
     "_love_numbers",
     "_parse_tide_potential_table",
     "_to_constituent_id",
@@ -1521,8 +1524,8 @@ def frequency(
         **kwargs
     ):
     """
-    Calculates the angular frequency for tidal constituents
-    :cite:p:`Ray:1999vm`
+    Wrapper function for calculating the angular frequency
+    for tidal constituents :cite:p:`Ray:1999vm`
 
     Parameters
     ----------
@@ -1552,28 +1555,9 @@ def frequency(
         method = 'Cartwright'
     else:
         method = 'ASTRO5'
-    # Modified Julian Dates at J2000
-    MJD = np.array([51544.5, 51544.55])
-    # time interval in seconds
-    deltat = 86400.0*(MJD[1] - MJD[0])
-    # calculate the mean longitudes of the sun and moon
-    s, h, p, n, pp = pyTMD.astro.mean_longitudes(MJD, method=method)
-
-    # number of temporal values
-    nt = len(np.atleast_1d(MJD))
-    # initial time conversions
-    hour = 24.0*np.mod(MJD, 1)
-    # convert from hours solar time into mean lunar time in degrees
-    tau = 15.0*hour - s + h
-    # variable for multiples of 90 degrees (Ray technical note 2017)
-    k = 90.0 + np.zeros((nt))
-
-    # determine equilibrium arguments
-    fargs = np.c_[tau, s, h, p, n, pp, k]
-    rates = (fargs[1,:] - fargs[0,:])/deltat
-    fd = np.dot(rates, coefficients_table(constituents, **kwargs))
-    # convert to radians per second
-    omega = 2.0*np.pi*fd/360.0
+    # get Doodson coefficients
+    coef = coefficients_table(constituents, **kwargs)
+    omega = _frequency(coef, method=method)
     return omega
 
 def aliasing_period(
@@ -1769,6 +1753,58 @@ def _constituent_parameters(c: str, **kwargs):
     # return the values for the constituent
     return (amplitude, phase, omega, alpha, species)
 
+def _frequency(
+        coef: np.ndarray,
+        **kwargs
+    ):
+    """
+    Calculates the angular frequency for Doodson coefficients
+    (Cartwright numbers) :cite:p:`Ray:1999vm`
+
+    Parameters
+    ----------
+    coef: list or np.ndarray
+        Doodson coefficients (Cartwright numbers) for constituents
+    method: str, default 'Cartwright'
+        Method for computing the mean longitudes
+
+            - ``'Cartwright'``
+            - ``'Meeus'``
+            - ``'ASTRO5'`` 
+            - ``'IERS'``
+
+    Returns
+    -------
+    omega: np.ndarray
[…]
