### 2025-08-05 c204356
docs: improve comments on simplified body tides
