### 2025-08-05 851ebf0
docs: add link to contribution guidelines for #434
docs: update tidal spectra to add degree-3 tides
