### 2025-08-05 7bf3505
docs: add link to contribution guidelines for #434 (#437)
docs: update tidal spectra to add degree-3 tides
