### 2025-08-05 44d85c8
refactor: use numpy deg2rad
