### 2025-08-05 27e77bf
refactor: use numpy deg2rad
