### 2025-08-05 814c7a1
docs: set images to latest `ubuntu` and `python`
