### 2025-08-05 7d78db8
docs: set images to latest `ubuntu` and `python` (#438)
