### 2025-08-06 9f9a4ab
beef up the statement of need
scrub some portions to keep near word count limits

mention PERTH is a fortran software
