+++ doc/source/_assets/pytmd-refs.bib
@@ -531,6 +531,12 @@ url = {https://doi.org/10.5194/os-17-615-2021},
 doi = {10.5194/os-17-615-2021},
 pages = {615--649},
 }
+@article{Lyard:2025tr,
+author = {Lyard, F and Carrere, L and Fouchet, E and Cancet, M and Greenberg, D and Dibarboure, G and Picot, N},
+title = {{FES2022: a step towards a SWOT-compliant tidal correction}},
+journal = {Journal of Geophysical Research: Oceans},
+year = {in review}
+}
 @article{Mathews:1991kv,
 author = {Mathews, P M and Buffett, B A and Herring, T A and Shapiro, I I},
 title = {{Forced nutations of the Earth: Influence of inner core dynamics: 1. Theory}},
