### 2025-08-06 f9c797a
beef up the statement of need
scrub some portions to keep near word count limits

mention PERTH is a fortran software
