+++ doc/source/_assets/paper.md
@@ -32,7 +32,7 @@ affiliations:
  - name: Department of Geophysics, Colorado School of Mines, United States
    index: 3
 
-date: 05 June 2024
+date: 06 August 2025
 bibliography: pytmd-refs.bib
 ---
 
@@ -42,42 +42,31 @@ bibliography: pytmd-refs.bib
 
 # Statement of need
 
-`pyTMD` is a generalized tide program that allows users to calculate both tide deflections and currents. As a tide model driver, `pyTMD` can read from a broad suite of models, and use different physics schemes in the internal calculations. Over 50 different models are presently supported within `pyTMD`, and additional models can be defined with a JSON file. The online documentation contains background information for tide modeling and multiple `Jupyter` notebooks that demonstrate core functionality.
+There are several ocean tide prediction software options available. The OSU Tidal Inversion Software (`OTIS`) and OSU Tidal Prediction Software (`OTPS`) were developed by Lana Erofeeva and Gary Egbert. The Tidal Model Driver ([`TMD`](https://github.com/EarthAndSpaceResearch/TMD_Matlab_Toolbox_v2.5)) MATLAB Toolbox was developed by Laurie Padman, Lana Erofeeva and Susan Howard. An updated version of the MATLAB Toolbox ([`TMD3`](https://github.com/chadagreene/tide-model-driver)) was developed by Chad Greene. The NASA GSFC PREdict Tidal Heights (`PERTH3`) software was developed by Richard Ray and Remko Scharroo. An updated and more versatile version of the NASA GSFC Fortran software ([`PERTH5`](https://codeberg.org/rray/perth5)) was developed by Richard Ray. [`pyFES`](https://cnes.github.io/aviso-fes/) was produced by LEGOS, NOVELTIS and CLS Ocean and Climate Division, and funded by CNES [@Lyard:2025tr]. These software options are typically are created by or for the model providers, and support their specific model formats.
 
-`pyTMD` was designed to be used by beginners and scientific researchers alike. It has been used in a number of scientific publications for modeling regional tides [@Freer:2023bt; @Millan:2023ju; @Sutterley:2019cl], modeling global tides [@Gregg:2024ky; @Paprotny:2024et], and creating several Earth observation datasets [@ENVEO:2021jh; @Smith:2024fd; @Smith:2024cv]. It has also been leveraged within larger earth-observation software packages [@Fitzpatrick:2024iv; @BishopTaylor:2025dc]. 
+`pyTMD` is a generalized tide program that allows users to calculate both tide deflections and currents. As a tide model driver, `pyTMD` can read from a broad suite of models, and use different physics schemes in the internal calculations. Over 50 different models are presently supported within `pyTMD`, and additional models can be defined with a JSON file. 
+
+`pyTMD` was designed to be used by beginners and scientific researchers alike. The online documentation contains background information for both tidal modeling and prediction. The software has been used in a number of scientific publications for modeling regional tides [@Freer:2023bt; @Millan:2023ju; @Sutterley:2019cl], modeling global tides [@Gregg:2024ky; @Paprotny:2024et], and creating several Earth observation datasets [@ENVEO:2021jh; @Smith:2024fd; @Smith:2024cv]. It has also been leveraged within larger earth-observation software packages [@Fitzpatrick:2024iv; @BishopTaylor:2025dc]. 
 
 # Functionality
 
 ## Ocean and Load Tides
 
[…]
