### 2025-08-08 54f206d
feat: add anelastic correction for long-period body tides
