### 2025-08-08 82713b7
test: add check for long-period complex love numbers
