### 2025-08-08 73eb8bf
fix: replace invalid NOAA water level values (#439)
* feat: add anelastic correction for long-period body tides

* test: add check for long-period complex love numbers
