+++ pyTMD/arguments.py
@@ -42,6 +42,7 @@ UPDATE HISTORY:
     Updated 08/2025: add Cartwright and Tayler table with radiational tides
         make frequency function a wrapper around one that calculates using
             Doodson coefficients (Cartwright numbers)
+        add complex love numbers function for correcting out-of-phase effects
     Updated 04/2025: convert longitudes p and n to radians within nodal function
         use schureman_arguments function to get nodal variables for FES models
         added Schureman to list of M1 options in nodal arguments
@@ -118,6 +119,7 @@ __all__ = [
     "_constituent_parameters",
     "_frequency",
     "_love_numbers",
+    "_complex_love_numbers",
     "_parse_tide_potential_table",
     "_to_constituent_id",
     "_to_doodson_number",
@@ -1553,12 +1555,13 @@ def frequency(
     # set function for astronomical longitudes
     # use ASTRO5 routines if not using an OTIS type model
     if kwargs['corrections'] in ('OTIS','ATLAS','TMD3','netcdf'):
-        method = 'Cartwright'
+        kwargs.setdefault('method', 'Cartwright')
     else:
-        method = 'ASTRO5'
+        kwargs.setdefault('method', 'ASTRO5')
     # get Doodson coefficients
     coef = coefficients_table(constituents, **kwargs)
-    omega = _frequency(coef, method=method)
+    # calculate the angular frequency
+    omega = _frequency(coef, **kwargs)
     return omega
 
 def aliasing_period(
@@ -1808,7 +1811,8 @@ def _frequency(
 
 def _love_numbers(
         omega: np.ndarray,
-        model: str = 'PREM'
+        model: str = 'PREM',
+        **kwargs
     ):
     """
     Compute the body tide Love/Shida numbers for a given frequency
@@ -1827,6 +1831,8 @@ def _love_numbers(
             - 'PEM-C'
             - 'C2'
             - 'PREM'
+    astype: np.dtype, default np.float64
+        data type for the output Love numbers
 
     Returns
     -------
@@ -1837,6 +1843,8 @@ def _love_numbers(
     l2: float
         Degree-2 Love (Shida) number of horizontal displacement
     """
+    # set default keyword arguments
+    kwargs.setdefault('astype', np.float64)
     # free core nutation frequencies (cycles per sidereal day) and
     # Love number parameters from Wahr (1981) table 6
     # and Mathews et al. (1995) table 3
@@ -1887,7 +1895,7 @@ def _love_numbers(
         # calculate love numbers following J. Wahr (1979)
         # use resonance formula for tides in the diurnal band
         # frequency of the o1 tides (radians/second)
-        omega_o1, = frequency('o1')
+        omega_o1, = frequency('o1', **kwargs)
         # convert frequency from cycles per sidereal day
         # frequency of free core nutation (radians/second)
         omega_fcn = lambda_fcn*7292115e-11
@@ -1898,7 +1906,76 @@ def _love_numbers(
         k2 = k0 + k1*ratio
         l2 = l0 + l1*ratio
     # return the Love numbers for frequency
-    return (h2, k2, l2)
+    return np.array([h2, k2, l2], dtype=kwargs['astype'])
+
+def _complex_love_numbers(
+        omega: np.ndarray,
+        **kwargs
+    ):
+    """
+    Compute the complex body tide Love/Shida numbers with in-phase
+    and out-of-phase components for a given frequency
+    :cite:p:`Mathews:1997js,Petit:2010tp`
+
+    Parameters
[…]
