+++ test/test_solid_earth.py
@@ -201,7 +201,9 @@ def test_solid_earth_radial(EPHEMERIDES):
     assert np.isclose(-tide_earth_free2mean, predicted, atol=5e-4).all()
     assert np.isclose(tide_mean-tide_free, predicted, atol=5e-4).all()
 
-def test_body_tides():
+# parameterize method
+@pytest.mark.parametrize("METHOD", ['ASTRO5','IERS'])
+def test_body_tides(METHOD):
     """Test simplified solid tides using predictions from ICESat-2
     """
     times = np.array(['2018-10-14 00:21:48','2018-10-14 00:21:48',
@@ -222,6 +224,6 @@ def test_body_tides():
     # using tide potentials from Cartwright and Tayler (1971)
     ts = timescale.from_datetime(times)
     tide_free = pyTMD.predict.body_tide(ts.tide, longitudes, latitudes,
-        deltat=ts.tt_ut1, tide_system='tide_free')
+        deltat=ts.tt_ut1, tide_system='tide_free', method=METHOD)
     # since we are using simplified body tides: assert within 2 mm
     assert np.isclose(tide_earth, tide_free[:,2], atol=2e-3).all()
