### 2025-08-08 64ee19b
refactor: unpin dependencies in pyproject.toml
    will be pinned in the lock file

docs: add pixi to install instructions
