### 2025-08-08 5b66ecf
refactor: unpin dependencies in pyproject.toml (#440)
will be pinned in the lock file

docs: add pixi to install instructions
