### 2025-08-08 ada76fb
refactor: drop `setuptools-scm` dependency
refactor: move scripts to be entry points
