### 2025-08-11 43ca442
refactor: drop `setuptools-scm` dependency (#441)
refactor: move scripts to be entry points
