+++ README.rst
@@ -98,7 +98,6 @@ Dependencies
 - `numpy: Scientific Computing Tools For Python <https://www.numpy.org>`_
 - `pyproj: Python interface to PROJ library <https://pypi.org/project/pyproj/>`_
 - `scipy: Scientific Tools for Python <https://www.scipy.org/>`_
-- `setuptools_scm: manager for python package versions using scm metadata <https://pypi.org/project/setuptools-scm>`_
 - `timescale: Python tools for time and astronomical calculations <https://pypi.org/project/timescale/>`_
 
 References
