+++ environment.yml
@@ -17,5 +17,4 @@ dependencies:
   - python>=3.6
   - python-dateutil
   - scipy>=1.16.0
-  - setuptools_scm
   - timescale>=0.0.8
