### 2025-08-11 1ec6ed5
docs: fix urls for moved scripts
