+++ doc/source/api_reference/scripts/reduce_OTIS_files.rst
@@ -6,7 +6,7 @@ reduce_OTIS_files.py
 
 `Source code`__
 
-.. __: https://github.com/pyTMD/pyTMD/blob/main/scripts/reduce_OTIS_files.py
+.. __: https://github.com/pyTMD/pyTMD/blob/main/pyTMD/scripts/reduce_OTIS_files.py
 
 Calling Sequence
 ################
