+++ doc/source/api_reference/scripts/gsfc_got_tides.rst
@@ -13,7 +13,7 @@ gsfc_got_tides.py
 
 `Source code`__
 
-.. __: https://github.com/pyTMD/pyTMD/blob/main/scripts/gsfc_got_tides.py
+.. __: https://github.com/pyTMD/pyTMD/blob/main/pyTMD/scripts/gsfc_got_tides.py
 
 Calling Sequence
 ################
