### 2025-08-11 e2e3ca2
fix: argparse of entry points
