### 2025-08-11 987e82a
docs: fix urls for moved scripts (#442)
* docs: fix urls for moved scripts

* fix: argparse of entry points
