+++ doc/source/index.rst
@@ -138,11 +138,11 @@ Project Details
     :hidden:
     :caption: Utilities
 
-    api_reference/arcticdata_tides.rst
-    api_reference/aviso_fes_tides.rst
-    api_reference/gsfc_got_tides.rst
-    api_reference/reduce_OTIS_files.rst
-    api_reference/verify_box_tpxo.rst
+    api_reference/scripts/arcticdata_tides.rst
+    api_reference/scripts/aviso_fes_tides.rst
+    api_reference/scripts/gsfc_got_tides.rst
+    api_reference/scripts/reduce_OTIS_files.rst
+    api_reference/scripts/verify_box_tpxo.rst
 
 .. toctree::
     :maxdepth: 1
