### 2025-08-11 51f57d4
docs: pixi as subsection on `README.rst`
