### 2025-08-11 dbccccb
docs: pixi as subsection on `README.rst` (#443)
