### 2025-08-11 908493b
feat: update functions for diurnal complex love numbers
use values from Mathews et al. (2002)
