+++ pyTMD/arguments.py
@@ -43,6 +43,7 @@ UPDATE HISTORY:
         make frequency function a wrapper around one that calculates using
             Doodson coefficients (Cartwright numbers)
         add complex love numbers function for correcting out-of-phase effects
+        use Mathews et al. (2002) functions for diurnal complex love numbers
     Updated 04/2025: convert longitudes p and n to radians within nodal function
         use schureman_arguments function to get nodal variables for FES models
         added Schureman to list of M1 options in nodal arguments
@@ -1915,7 +1916,7 @@ def _complex_love_numbers(
     """
     Compute the complex body tide Love/Shida numbers with in-phase
     and out-of-phase components for a given frequency
-    :cite:p:`Mathews:1997js,Petit:2010tp`
+    :cite:p:`Mathews:1997js,Mathews:2002cr,Petit:2010tp`
 
     Parameters
     ----------
@@ -1933,47 +1934,68 @@ def _complex_love_numbers(
     l2: complex
         Degree-2 Love (Shida) number of horizontal displacement
     """
-    # Doodson cefficients from table 7.3b of the 2010 IERS conventions
-    coefficients = np.array([
-        [0.0, 0.0, 0.0, 0.0, -1.0, 0.0, 0.0],
-        [0.0, 0.0, 2.0, 0.0, 0.0, 0.0, 0.0],
-        [0.0, 1.0, 0.0, -1.0, 0.0, 0.0, 0.0],
-        [0.0, 2.0, 0.0, 0.0, 0.0, 0.0, 0.0],
-        [0.0, 2.0, 0.0, 0.0, -1.0, 0.0, 0.0]
-    ])
-    # calculate angular frequencies for coefficients from IERS table
-    anelastic_omegas = [_frequency(c, **kwargs) for c in coefficients]
-    # love numbers with adjustments for the frequency dependence
-    # induced by mantle anelasticity at long-periods
-    anelastic_love_numbers = np.array([
-        [0.6344 + 0.0093j, 0.31537 - 0.00541j, 0.0936 + 0.0028j],
-        [0.6182 + 0.0054j, 0.30593 - 0.00315j, 0.0886 + 0.0016j],
-        [0.6126 + 0.0041j, 0.30270 - 0.00237j, 0.0870 + 0.0012j],
-        [0.6109 + 0.0037j, 0.30171 - 0.00213j, 0.0864 + 0.0011j],
-        [0.6109 + 0.0037j, 0.30171 - 0.00213j, 0.0864 + 0.0011j],
-    ])
-    # calculate the Love numbers for the frequency
-    h2, k2, l2 = _love_numbers(omega, **kwargs)
+    # (number of sidereal days per solar day)
+    sidereal_ratio = 1.002737909
+    # number of seconds in a sidereal day (approx 86164.1)
+    sidereal_day = 86400.0/sidereal_ratio
+    # frequency in cycles per sidereal day
+    f = omega*sidereal_day/(2.0*np.pi)
     # Love numbers for different frequency bands
     if (omega > 1e-4):
-        # add out-of-phase components for the semi-diurnal band
-        h2 += 0.0022j
-        k2 += 0.0013j
-        l2 += 0.0007j
+        # in-phase and out-of-phase components for the semi-diurnal band
+        h2 = 0.6078 - 0.0022j
+        k2 = 0.30102 - 0.0013j
+        l2 = 0.0847 - 0.0007j
     elif (omega < 2e-5):
         # compute in-phase and out-of-phase components for the long period band
-        # interpolate the anelastic Love numbers to the frequency
-        h2 = np.interp(omega, anelastic_omegas, anelastic_love_numbers[:, 0],
[…]
