### 2025-08-11 b1de48d
fix: use real-valued body tide love numbers for permanent tide
