### 2025-08-12 ef0e139
feat: include mantle anelastic effects when inferring long-period tides
