### 2025-08-12 baa5623
refactor: make mantle anelasticity effects optional
