### 2025-08-12 f248311
docs: update docstring for infer long period band
