### 2025-08-12 0757e3b
feat: add option for mantle anelasticity for LPET predict
