+++ pyTMD/predict.py
@@ -26,6 +26,7 @@ UPDATE HISTORY:
         add correction of anelastic effects for long-period body tides
         use sign convention from IERS for complex body tide Love numbers
         include mantle anelastic effects when inferring long-period tides
+        added option to include mantle anelastic effects for LPET predict
     Updated 07/2025: revert free-to-mean conversion to April 2023 version
         revert load pole tide to IERS 1996 convention definitions
         mask mean pole values prior to valid epoch of convention
@@ -888,6 +889,7 @@ def _infer_long_period(
                 # include variations largely due to mantle anelasticity
                 h2, k2, l2 = pyTMD.arguments._complex_love_numbers(omajor[i])
             else:
+                # Love numbers for long-period tides (Wahr, 1981)
                 h2, k2, l2 = pyTMD.arguments._love_numbers(omajor[i],
                     astype=np.complex128)
             # tilt factor: response with respect to the solid earth
@@ -948,6 +950,7 @@ def _infer_long_period(
             # include variations largely due to mantle anelasticity
             h2, k2, l2 = pyTMD.arguments._complex_love_numbers(omega[k])
         else:
+            # Love numbers for long-period tides (Wahr, 1981)
             h2, k2, l2 = pyTMD.arguments._love_numbers(omega[k],
                 astype=np.complex128)
         # tilt factor: response with respect to the solid earth
@@ -988,6 +991,8 @@ def equilibrium_tide(
         time correction for converting to Ephemeris Time (days)
     corrections: str, default 'OTIS'
         use nodal corrections from OTIS/ATLAS or GOT/FES models
+    include_anelasticity: bool, default False
+        compute Love numbers taking into account mantle anelasticity
     constituents: list
         long-period tidal constituent IDs
 
@@ -1002,11 +1007,14 @@ def equilibrium_tide(
         'mst', 'mt', '085.465']
     kwargs.setdefault('constituents', cindex)
     kwargs.setdefault('deltat', 0.0)
+    kwargs.setdefault('include_anelasticity', False)
     kwargs.setdefault('corrections', 'OTIS')
 
     # number of input points
     nt = len(np.atleast_1d(t))
     nlat = len(np.atleast_1d(lat))
+    # number of constituents
+    nc = 15
 
     # set function for astronomical longitudes
     # use ASTRO5 routines if not using an OTIS type model
@@ -1017,16 +1025,24 @@ def equilibrium_tide(
     # convert from Modified Julian Dates into Ephemeris Time
     MJD = t + _mjd_tide
     # compute principal mean longitudes
-    s, h, p, N, pp = pyTMD.astro.mean_longitudes(MJD + kwargs['deltat'],
+    s, h, p, n, pp = pyTMD.astro.mean_longitudes(MJD + kwargs['deltat'],
         method=method)
+    # initial time conversions
+    hour = 24.0*np.mod(MJD, 1)
+    # convert from hours solar time into mean lunar time in degrees
[…]
