### 2025-08-12 8f37cbd
feat: update functions for diurnal complex love numbers (#444)
* feat: include mantle anelastic effects when inferring long-period tides
* feat: add option for mantle anelasticity for LPET predictions
