+++ test/test_love_numbers.py
@@ -62,6 +62,14 @@ def test_complex_love_numbers():
     """
     # Doodson coefficients
     coefficients = {}
+    # diurnal species
+    coefficients['q1'] = [1.0, -2.0, 0.0, 1.0, 0.0, 0.0, -1.0]
+    coefficients['o1'] = [1.0, -1.0, 0.0, 0.0, 0.0, 0.0, -1.0]
+    coefficients['pi1'] = [1.0, 1.0, -3.0, 0.0, 0.0, 1.0, -1.0]
+    coefficients['p1'] = [1.0, 1.0, -2.0, 0.0, 0.0, 0.0, -1.0]
+    # semi-diurnal species
+    coefficients['m2'] = [2.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
+    # long-period species
     coefficients['055.565'] = [0.0, 0.0, 0.0, 0.0, -1.0, 0.0, 0.0]
     coefficients['ssa'] = [0.0, 0.0, 2.0, 0.0, 0.0, 0.0, 0.0]
     coefficients['mm'] = [0.0, 1.0, 0.0, -1.0, 0.0, 0.0, 0.0]
@@ -69,12 +77,19 @@ def test_complex_love_numbers():
     coefficients['075.565'] = [0.0, 2.0, 0.0, 0.0, -1.0, 0.0, 0.0]
     # expected values
     exp = {}
+    # diurnal species
+    exp['q1'] = (0.6036 - 0.0026j, 0.29785 - 0.00139j, 0.0846 - 0.0008j)
+    exp['o1'] = (0.6028 - 0.0025j, 0.29747 - 0.00137j, 0.0846 - 0.0008j)
+    exp['pi1'] = (0.5878 - 0.0015j, 0.28996 - 0.00086j, 0.0847 - 0.0007j)
+    exp['p1'] = (0.5817 - 0.0011j, 0.28692 - 0.00067j, 0.0853 - 0.0007j)
+    # semi-diurnal species
+    exp['m2'] = (0.6078 - 0.0022j, 0.30102 - 0.0013j, 0.0847 - 0.0007j)
     # long-period species
-    exp['055.565'] = (0.6344 + 0.0093j, 0.31537 - 0.00541j, 0.0936 + 0.0028j)
-    exp['ssa'] = (0.6182 + 0.0054j, 0.30593 - 0.00315j, 0.0886 + 0.0016j)
-    exp['mm'] = (0.6126 + 0.0041j, 0.30270 - 0.00237j, 0.0870 + 0.0012j)
-    exp['mf'] = (0.6109 + 0.0037j, 0.30171 - 0.00213j, 0.0864 + 0.0011j)
-    exp['075.565'] = (0.6109 + 0.0037j, 0.30171 - 0.00213j, 0.0864 + 0.0011j)
+    exp['055.565'] = (0.6344 - 0.0093j, 0.31537 - 0.00541j, 0.0936 - 0.0028j)
+    exp['ssa'] = (0.6182 - 0.0054j, 0.30593 - 0.00315j, 0.0886 - 0.0016j)
+    exp['mm'] = (0.6126 - 0.0041j, 0.30270 - 0.00237j, 0.0870 - 0.0012j)
+    exp['mf'] = (0.6109 - 0.0037j, 0.30171 - 0.00213j, 0.0864 - 0.0011j)
+    exp['075.565'] = (0.6109 - 0.0037j, 0.30171 - 0.00213j, 0.0864 - 0.0011j)
     # for each tidal constituent
     for c, v in exp.items():
         # calculate Love numbers
