+++ pyTMD/predict.py
@@ -24,6 +24,9 @@ PROGRAM DEPENDENCIES:
 UPDATE HISTORY:
     Updated 08/2025: add simplified solid earth tide prediction function
         add correction of anelastic effects for long-period body tides
+        use sign convention from IERS for complex body tide Love numbers
+        include mantle anelastic effects when inferring long-period tides
+        added option to include mantle anelastic effects for LPET predict
     Updated 07/2025: revert free-to-mean conversion to April 2023 version
         revert load pole tide to IERS 1996 convention definitions
         mask mean pole values prior to valid epoch of convention
@@ -823,8 +826,9 @@ def _infer_long_period(
     ):
     """
     Infer the tidal values for long-period minor constituents
-    using their relation with major constituents
-    :cite:p:`Ray:1999vm,Ray:2014fu,Cartwright:1973em`
+    using their relation with major constituents with option to
+    take into account variations due to mantle anelasticity
+    :cite:p:`Ray:1999vm,Ray:2014fu,Cartwright:1973em,Mathews:2002cr`
 
     Parameters
     ----------
@@ -838,6 +842,8 @@ def _infer_long_period(
         time correction for converting to Ephemeris Time (days)
     minor: list or None, default None
         tidal constituent IDs of minor constituents for inference
+    include_anelasticity: bool, default False
+        compute Love numbers taking into account mantle anelasticity
     raise_exception: bool, default False
         Raise a ``ValueError`` if major constituents are not found
 
@@ -849,6 +855,7 @@ def _infer_long_period(
     # set default keyword arguments
     kwargs.setdefault('deltat', 0.0)
     kwargs.setdefault('corrections', 'OTIS')
+    kwargs.setdefault('include_anelasticity', False)
     kwargs.setdefault('raise_exception', False)
     # list of minor constituents
     kwargs.setdefault('minor', None)
@@ -877,7 +884,19 @@ def _infer_long_period(
         j = [j for j,val in enumerate(constituents) if (val.lower() == c)]
         if j:
             j1, = j
-            z[:,i] = zmajor[:,j1]/amajor[i]
+            # complex Love numbers of degree 2 for long-period band
+            if kwargs['include_anelasticity']:
+                # include variations largely due to mantle anelasticity
+                h2, k2, l2 = pyTMD.arguments._complex_love_numbers(omajor[i])
+            else:
+                # Love numbers for long-period tides (Wahr, 1981)
+                h2, k2, l2 = pyTMD.arguments._love_numbers(omajor[i],
+                    astype=np.complex128)
+            # tilt factor: response with respect to the solid earth
+            # use real components from Mathews et al. (2002)
+            gamma_2 = (1.0 + k2.real - h2.real)
+            # "normalize" tide values
+            z[:,i] = zmajor[:,j1]/(amajor[i]*gamma_2)
             nz += 1
 
     # raise exception or log error
@@ -926,13 +945,24 @@ def _infer_long_period(
 
[…]
