+++ doc/source/_assets/pytmd-refs.bib
@@ -484,6 +484,19 @@ doi = {10.1029/97JB01515},
 pages = {20469--20477},
 publisher = {American Geophysical Union (AGU)},
 }
+@article{Mathews:2002cr,
+author = {Mathews, P M and Herring, T A and Buffett, B A},
+title = {{Modeling of nutation and precession: New nutation series for nonrigid Earth and insights into the Earth's interior}},
+journal = {Journal of Geophysical Research: Solid Earth},
+year = {2002},
+month = apr,
+volume = {107},
+number = {B4},
+issn = {0148-0227},
+url = {https://doi.org/10.1029/2001jb000390},
+doi = {10.1029/2001JB000390},
+publisher = {American Geophysical Union (AGU)},
+}
 @book{Meeus:1991vh,
 author = {Meeus, J H},
 title = {{Astronomical Algorithms}},

[…]
