### 2025-08-13 16e5f79
refactor: switch time decimal in pole tides to nominal years of 365.25 days
refactor: convert angles with numpy radians and degrees functions
feat: add arcs2rad and rad2arcs functions and use in conversions
