+++ pyTMD/spatial.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 spatial.py
-Written by Tyler Sutterley (03/2025)
+Written by Tyler Sutterley (08/2025)
 
 Spatial transformation routines
 
@@ -11,6 +11,7 @@ PYTHON DEPENDENCIES:
         https://numpy.org/doc/stable/user/numpy-for-matlab-users.html
 
 UPDATE HISTORY:
+    Updated 08/2025: convert angles with numpy radians and degrees functions
     Updated 03/2025: add more ellipsoidal parameters to datum class
     Updated 02/2025: major refactor to move io routines out of this module
     Updated 12/2024: add latitude and longitude as potential dimension names
@@ -881,14 +882,13 @@ def to_cartesian(
     lin_ecc = np.sqrt((2.0*flat - flat**2)*a_axis**2)
     ecc1 = lin_ecc/a_axis
     # convert from geodetic latitude to geocentric latitude
-    dtr = np.pi/180.0
     # geodetic latitude in radians
-    latitude_geodetic_rad = lat*dtr
+    latitude_geodetic_rad = np.radians(lat)
     # prime vertical radius of curvature
     N = a_axis/np.sqrt(1.0 - ecc1**2.0*np.sin(latitude_geodetic_rad)**2.0)
     # calculate X, Y and Z from geodetic latitude and longitude
-    X = (N + h) * np.cos(latitude_geodetic_rad) * np.cos(lon*dtr)
-    Y = (N + h) * np.cos(latitude_geodetic_rad) * np.sin(lon*dtr)
+    X = (N + h) * np.cos(latitude_geodetic_rad) * np.cos(np.radians(lon))
+    Y = (N + h) * np.cos(latitude_geodetic_rad) * np.sin(np.radians(lon))
     Z = (N * (1.0 - ecc1**2.0) + h) * np.sin(latitude_geodetic_rad)
     # return the cartesian coordinates
     # flattened to singular values if necessary
@@ -1038,10 +1038,8 @@ def _moritz_iterative(
     # Linear eccentricity and first numerical eccentricity
     lin_ecc = np.sqrt((2.0*flat - flat**2)*a_axis**2)
     ecc1 = lin_ecc/a_axis
-    # degrees to radians
-    dtr = np.pi/180.0
     # calculate longitude
-    lon = np.arctan2(y, x)/dtr
+    lon = np.degrees(np.arctan2(y, x))
     # set initial estimate of height to 0
     h = np.zeros_like(lon)
     h0 = np.inf*np.ones_like(lon)
@@ -1062,8 +1060,10 @@ def _moritz_iterative(
         phi = np.arctan(z/(p*(1.0 - ecc1**2*N/(N + h))))
         # add to iterator
         i += 1
+    # convert latitudes and fix values
+    lat = np.clip(np.degrees(phi), -90.0, 90.0)
     # return longitude, latitude and height
-    return (lon, phi/dtr, h)
+    return (lon, lat, h)
 
 def _bowring_iterative(
         x: np.ndarray,
@@ -1102,10 +1102,8 @@ def _bowring_iterative(
     # square of first and second numerical eccentricity
     e12 = lin_ecc**2/a_axis**2
     e22 = lin_ecc**2/b_axis**2
-    # degrees to radians
-    dtr = np.pi/180.0
     # calculate longitude
[…]
