### 2025-08-13 94e1bea
refactor: switch time decimal in pole tides to nominal years of 365.25 days (#445)
refactor: convert angles with numpy radians and degrees functions
feat: add arcs2rad and rad2arcs functions and use in conversions
