+++ pyTMD/astro.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 astro.py
-Written by Tyler Sutterley (05/2025)
+Written by Tyler Sutterley (08/2025)
 Astronomical and nutation routines
 
 PYTHON DEPENDENCIES:
@@ -18,6 +18,8 @@ REFERENCES:
     Oliver Montenbruck, Practical Ephemeris Calculations, 1989.
 
 UPDATE HISTORY:
+    Updated 08/2025: convert angles with numpy radians and degrees functions
+        convert arcseconds to radians with arcs2rad function in math.py
     Updated 05/2025: use Barycentric Dynamical Time (TDB) for JPL ephemerides
     Updated 04/2025: added schureman arguments function for FES models
         more outputs from schureman arguments function for M1 constituent
@@ -68,6 +70,7 @@ import timescale.time
 from pyTMD.math import (
     polynomial_sum,
     normalize_angle,
+    arcs2rad,
     rotate
 )
 from pyTMD.utilities import (
@@ -210,19 +213,17 @@ def mean_longitudes(
     elif (kwargs['method'].upper() == 'IERS'):
         # compute the Delaunay arguments (IERS conventions)
         l, lp, F, D, omega = delaunay_arguments(MJD)
-        # degrees to radians
-        dtr = np.pi/180.0
         # convert to Doodson arguments in degrees
         # mean longitude of moon
-        S = (F + omega)/dtr
+        S = np.degrees(F + omega)
         # mean longitude of sun
-        H = (F + omega - D)/dtr
+        H = np.degrees(F + omega - D)
         # longitude of lunar perigee
-        P = (F + omega - l)/dtr
+        P = np.degrees(F + omega - l)
         # longitude of ascending lunar node
-        N = omega/dtr
+        N = np.degrees(omega)
         # longitude of solar perigee
-        Ps = (-lp + F - D + omega)/dtr
+        Ps = np.degrees(-lp + F - D + omega)
     else:
         # Formulae for the period 1990--2010 derived by David Cartwright
         # convert from MJD to days relative to 2000-01-01T12:00:00
@@ -295,8 +296,6 @@ def doodson_arguments(
     Ps: np.ndarray
         mean longitude of solar perigee (radians)
     """
-    # degrees to radians
-    dtr = np.pi/180.0
     # convert from MJD to centuries relative to 2000-01-01T12:00:00
     T = (MJD - _mjd_j2000)/_century
     # hour of the day
@@ -335,12 +334,12 @@ def doodson_arguments(
     Ps = polynomial_sum(np.array([282.93734098, 1.71945766667,
         4.5688889e-4, -1.778e-8, -3.34e-9]), T)
     # take the modulus of each and convert to radians
-    S = dtr*normalize_angle(S)
-    H = dtr*normalize_angle(H)
-    P = dtr*normalize_angle(P)
-    TAU = dtr*normalize_angle(TAU)
-    Np = dtr*normalize_angle(Np)
-    Ps = dtr*normalize_angle(Ps)
+    S = np.radians(normalize_angle(S))
+    H = np.radians(normalize_angle(H))
+    P = np.radians(normalize_angle(P))
+    TAU = np.radians(normalize_angle(TAU))
+    Np = np.radians(normalize_angle(Np))
+    Ps = np.radians(normalize_angle(Ps))
     # return as tuple
     return (TAU, S, H, P, Np, Ps)
 
[…]

[…]
