### 2025-08-13 f134e1c
refactor: rename to asec2rad
matches nomenclature i used in `timescale`
