### 2025-08-13 fcfc937
refactor: rename to asec2rad (#446)
matches nomenclature i used in `timescale`
