### 2025-08-14 39eb979
fix: return numpy arrays if cannot infer minor constituents
fix: make long-period inference optional

see https://github.com/GeoscienceAustralia/eo-tides/issues/117
