### 2025-08-14 050d7df
fix: return numpy arrays if cannot infer minor constituents (#447)
fix: make long-period inference optional

see https://github.com/GeoscienceAustralia/eo-tides/issues/117
