### 2025-08-14 e3d24fc
refactor: convert microarcseconds to radians with masec2rad function in math.py
