+++ test/test_math.py
@@ -7,7 +7,7 @@ import numpy as np
 import pyTMD.ellipse
 import pyTMD.math
 
-def test_arcs2rad():
+def test_asec2rad():
     """
     Tests the conversion of arcseconds to radians
     """
@@ -24,21 +24,21 @@ def test_arcs2rad():
     test = angles * atr
     assert np.allclose(exp, test)
 
-def test_rad2arcs():
+def test_masec2rad():
     """
-    Tests the conversion of radians to arcseconds
+    Tests the conversion of microarcseconds to radians
     """
-    # test angles in radians
-    angles = np.array([-np.pi, -np.pi/2, 0, np.pi/2, np.pi, 3*np.pi/2, 2*np.pi])
-    # expected values in arcseconds
-    exp = np.array([-180, -90, 0, 90, 180, 270, 360])*3600.0
-    # test conversion to arcseconds
-    test = pyTMD.math.rad2asec(angles)
+    # test angles in microarcseconds
+    angles = np.array([-180, -90, 0, 90, 180, 270, 360])*3600.0*1e6
+    # expected values in radians
+    exp = np.array([-np.pi, -np.pi/2, 0, np.pi/2, np.pi, 3*np.pi/2, 2*np.pi])
+    # test conversion to radians
+    test = pyTMD.math.masec2rad(angles)
     assert np.allclose(exp, test)
-    # arcseconds to radians
-    atr = np.pi/648000.0
-    # test conversion from radians
-    test = angles / atr
+    # microarcseconds to radians
+    atr = np.pi/648e9
+    # test conversion to radians
+    test = angles * atr
     assert np.allclose(exp, test)
 
 def test_normalize_angle():
