### 2025-08-15 532aad2
refactor: more degree to radian conversions
