### 2025-08-15 46bbf0d
refactor: convert microarcseconds to radians with masec2rad function in math.py (#448)
