### 2025-08-15 7903070
chore: bump version to `2.2.7`
