### 2025-08-20 b3da047
docs: add some admonitions
docs: convert README to markdown
