+++ README.rst
@@ -1,158 +1,0 @@
-=====
-pyTMD
-=====
-
-|License|
-|Documentation Status|
-|PyPI|
-|conda-forge|
-|commits-since|
-|zenodo|
-
-.. |License| image:: https://img.shields.io/github/license/pyTMD/pyTMD
-   :target: https://github.com/pyTMD/pyTMD/blob/main/LICENSE
-
-.. |Documentation Status| image:: https://readthedocs.org/projects/pytmd/badge/?version=latest
-   :target: https://pytmd.readthedocs.io/en/latest/?badge=latest
-
-.. |PyPI| image:: https://img.shields.io/pypi/v/pyTMD.svg
-   :target: https://pypi.python.org/pypi/pyTMD/
-
-.. |conda-forge| image:: https://img.shields.io/conda/vn/conda-forge/pytmd
-   :target: https://anaconda.org/conda-forge/pytmd
-
-.. |commits-since| image:: https://img.shields.io/github/commits-since/pyTMD/pyTMD/latest
-   :target: https://github.com/pyTMD/pyTMD/releases/latest
-
-.. |zenodo| image:: https://zenodo.org/badge/DOI/10.5281/zenodo.5555395.svg
-   :target: https://doi.org/10.5281/zenodo.5555395
-
-Python-based tidal prediction software for estimating ocean, load, solid Earth and pole tides
-
-For more information: see the documentation at `pytmd.readthedocs.io <https://pytmd.readthedocs.io/>`_
-
-Installation
-############
-
-From PyPI:
-
-.. code-block:: bash
-
-   python3 -m pip install pyTMD
-
-To include all optional dependencies:
-
-.. code-block:: bash
-
-   python3 -m pip install pyTMD[all]
-
-Using ``conda`` or ``mamba`` from conda-forge:
-
-.. code-block:: bash
-
-   conda install -c conda-forge pytmd
-
-.. code-block:: bash
-
-   mamba install -c conda-forge pytmd
-
-Development version from GitHub:
-
-.. code-block:: bash
-
-   python3 -m pip install git+https://github.com/pyTMD/pyTMD.git
-
-Running with Pixi
------------------
-
-Alternatively, you can use `Pixi <https://pixi.sh/>`_ for a streamlined workspace environment:
-
-1. Install Pixi following the `installation instructions <https://pixi.sh/latest/#installation>`_
-
-2. Clone the project repository:
-
-.. code-block:: bash
-
-   git clone https://github.com/pyTMD/pyTMD.git
-
-3. Move into the ``pyTMD`` directory
-
-.. code-block:: bash
-
-   cd pyTMD
-
-4. Install dependencies and start JupyterLab:
-
-.. code-block:: bash
-
-   pixi run start
-
-This will automatically create the environment, install all dependencies, and launch JupyterLab in the `notebooks <./doc/source/notebooks/>`_ directory.
-
-Dependencies
-############
-
-- `dateutil: powerful extensions to datetime <https://dateutil.readthedocs.io/en/stable/>`_
-- `lxml: processing XML and HTML in Python <https://pypi.python.org/pypi/lxml>`_
-- `netCDF4: Python interface to the netCDF C library <https://unidata.github.io/netcdf4-python/>`_
-- `numpy: Scientific Computing Tools For Python <https://www.numpy.org>`_
-- `pyproj: Python interface to PROJ library <https://pypi.org/project/pyproj/>`_
-- `scipy: Scientific Tools for Python <https://www.scipy.org/>`_
-- `timescale: Python tools for time and astronomical calculations <https://pypi.org/project/timescale/>`_
-
-References
-##########
-
-    T. C. Sutterley, T. Markus, T. A. Neumann, M. R. van den Broeke, J. M. van Wessem, and S. R. M. Ligtenberg,
[…]
