+++ pyproject.toml
@@ -21,7 +21,7 @@ maintainers = [
     {name = "pyTMD contributors"}
 ]
 license = {file = "LICENSE"}
-readme = "README.rst"
+readme = {file = "README.md", content-type = "text/markdown"}
 
 requires-python = "~=3.6"
 dependencies = [

[…]
