### 2025-08-20 f4998b0
docs: convert README to markdown (#450)
* docs: update contribution guidelines
* docs: add some admonitions
