### 2025-08-21 f451a92
docs: add more notes and admonitions
