### 2025-08-21 e65cee1
docs: update contribution guidelines
