### 2025-08-21 d9df865
citation in admonition
