### 2025-08-21 ef7182f
Update Citations.rst
