### 2025-08-21 b089851
docs: add more notes and admonitions (#451)
* docs: update contribution guidelines
