+++ doc/source/api_reference/io/IERS.rst
@@ -4,7 +4,9 @@ IERS
 
 - Reads ocean pole load tide coefficients provided by IERS as computed by :cite:t:`Desai:2002ev` and :cite:t:`Desai:2015jr`
 - See `materials from Chapter 7 of the IERS Conventions <https://webtai.bipm.org/iers/convupdt/convupdt_c7.html>`_
-- `Ocean Pole Tide File <ftp://maia.usno.navy.mil/conventions/2010/2010_update/chapter7/additional_info/opoleloadcoefcmcor.txt.gz>`_
+
+.. tip::
+    `Ocean Pole Tide File <ftp://maia.usno.navy.mil/conventions/2010/2010_update/chapter7/additional_info/opoleloadcoefcmcor.txt.gz>`_
 
 Calling Sequence
 ----------------
