+++ doc/source/api_reference/scripts/aviso_fes_tides.rst
@@ -5,7 +5,9 @@ aviso_fes_tides.py
 - Downloads the FES (Finite Element Solution) global tide model from `AVISO <https://www.aviso.altimetry.fr/en/data/products/auxiliary-products/global-tide-fes.html>`_
 - Decompresses the model tar files into the constituent files and auxiliary files
 - Must have `data access to tide models from AVISO <https://www.aviso.altimetry.fr/en/data/data-access.html>`_
-- *FES outputs are licensed for scientific purposes only*
+
+.. note::
+    *FES outputs are licensed for scientific purposes only*
 
 `Source code`__
 
