### 2025-08-26 e923ddb
refactor:  use `timescale` shortcut wrapper function (#453)
