### 2025-08-26 1b2c217
feat: add functions for converting models to xarray
starting work for #452
