### 2025-08-26 c6eac6a
docs: add s3 recipes and examples
chore: add aws feature dependencies to pixi
chore: update pixi lock file
docs: add pixi line to create docs
