### 2025-08-27 dcd0d78
feat: use vectorized linear interpolator in `infer_minor`
feat: improve performance of bilinear interpolation
