+++ doc/source/user_guide/zarr-to-s3-obstore.py
@@ -7,13 +7,12 @@ s3_bucket = 'pytmd-scratch'
 tide_model = 'FES2022'
 # setup tide model
 m = pyTMD.io.model(compressed=False)
-
-# setup zarr store using obstore
-presigned_s3_url = f's3://{s3_bucket}/{tide_model}.zarr'
-s3_store = obstore.store.S3Store.from_url(presigned_s3_url, region="us-west-2")
-store = zarr.storage.ObjectStore(s3_store, read_only=False)
-
 # read tide model and convert to xarray DataTree
 dtree = m.to_datatree(tide_model)
+
+# setup zarr store using obstore
+presigned_s3_url = f's3://{s3_bucket}/{m.name}.zarr'
+s3_store = obstore.store.S3Store.from_url(presigned_s3_url, region="us-west-2")
+store = zarr.storage.ObjectStore(s3_store, read_only=False)
 # save to zarr store
 dtree.to_zarr(store, mode='w', zarr_format=3, consolidated=True)
