+++ pyTMD/interpolate.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 interpolate.py
-Written by Tyler Sutterley (09/2024)
+Written by Tyler Sutterley (08/2025)
 Interpolators for spatial data
 
 PYTHON DEPENDENCIES:
@@ -12,6 +12,8 @@ PYTHON DEPENDENCIES:
         https://docs.scipy.org/doc/
 
 UPDATE HISTORY:
+    Updated 08/2025: added vectorized 1D linear interpolation function
+        improve performance of bilinear interpolation and allow extrapolation
     Updated 09/2024: deprecation fix case where an array is output to scalars
     Updated 07/2024: changed projection flag in extrapolation to is_geographic
     Written 12/2022
@@ -24,12 +26,56 @@ import scipy.interpolate
 import pyTMD.spatial
 
 __all__ = [
+    "interp1d",
     "bilinear",
     "spline",
     "regulargrid",
     "extrapolate",
     "_distance"
 ]
+
+# PURPOSE: 1-dimensional linear interpolation on arrays
+def interp1d(
+        x: float,
+        xp: np.ndarray,
+        fp: np.ndarray,
+        extrapolate: str = 'linear'
+    ):
+    """
+    Vectorized one-dimensional linear interpolation
+
+    Parameters
+    ----------
+    x: np.ndarray
+        x-coordinates of the interpolated values
+    xp: np.ndarray
+        x-coordinates of the data points
+    fp: np.ndarray
+        y-coordinates of the data points
+    extrapolate: str, default = 'linear'
+        Method of extrapolation
+
+            - ``'linear'``
+            - ``'nearest'``    
+
+    Returns
+    -------
+    f: np.ndarray
+        Interpolated values at x
+    """
+    # clip coordinates to handle nearest-neighbor extrapolation
+    if (extrapolate == 'nearest'):
+        x = np.clip(x, a_min=xp.min(), a_max=xp.max())
+    # find indice where x could be inserted into xp
+    j = np.searchsorted(xp, x) - 1
+    # clip indices to handle linear extrapolation
+    if (extrapolate == 'linear'):
+        j = np.clip(j, a_min=0, a_max=len(xp) - 2)
+    # fractional distance between points
+    d = np.divide(x - xp[j], xp[j+1] - xp[j])
+    # calculate interpolated values
+    f = (1.0 - d)*fp[:,j] + d*fp[:,j+1]
+    return f
 
 # PURPOSE: bilinear interpolation of input data to output data
 def bilinear(
@@ -39,6 +85,7 @@ def bilinear(
         lon: np.ndarray,
         lat: np.ndarray,
         fill_value: float = np.nan,
+        extrapolate: bool = False,
         dtype: str | np.dtype = np.float64
     ):
     """
@@ -58,6 +105,8 @@ def bilinear(
         output longitude
     fill_value: float, default np.nan
         invalid value
+    extrapolate: bool, default False
+        allow linear extrapolation of points
     dtype: np.dtype, default np.float64
         output data type
 
@@ -70,9 +119,6 @@ def bilinear(
     if not isinstance(idata, np.ma.MaskedArray):
         idata = np.ma.array(idata)
         idata.mask = np.zeros_like(idata, dtype=bool)
-    # find valid points (within bounds)
-    valid, = np.nonzero((lon >= ilon.min()) & (lon <= ilon.max()) &
-        (lat > ilat.min()) & (lat < ilat.max()))
[…]
