+++ doc/source/user_guide/Recipes.rst
@@ -20,7 +20,7 @@ Use ``s3fs`` to setup a ``zarr`` store
 ######################################
 
 The above examples use ``obstore`` to access a ``zarr`` store on S3.
-The ``s3fs`` package is an alternative method that creates a ``fsspec`` based store.
+The ``s3fs`` package is an alternative method that uses a ``fsspec`` based store.
 
 .. include:: ./setup-s3fs.py
    :literal:
