### 2025-08-28 482fa0e
test: add xarray tests
ci: add xarray to environment
