### 2025-08-28 385cbcc
feat: add functions for converting models to `xarray` (#455)
docs: add s3 recipes and examples
chore: add aws feature dependencies to pixi
chore: update pixi lock file
docs: add pixi task that can create HTML documentation
feat: use vectorized linear interpolator in `infer_minor`
feat: improve performance of `bilinear` interpolation
test: add `xarray` tests
ci: add `xarray` to test environment
