+++ doc/source/api_reference/interpolate.rst
@@ -16,6 +16,8 @@ Calling Sequence
 
 .. __: https://github.com/pyTMD/pyTMD/blob/main/pyTMD/interpolate.py
 
+.. autofunction:: pyTMD.interpolate.interp1d
+
 .. autofunction:: pyTMD.interpolate.bilinear
 
 .. autofunction:: pyTMD.interpolate.spline

[…]
