### 2025-08-28 6bbc48e
feat: add basic accessor for tide model data
