+++ doc/source/api_reference/io/dataset.rst
@@ -1,0 +1,15 @@
+=======
+dataset
+=======
+
+An ``xarray.Dataset`` extension for tidal model data
+
+`Source code`__
+
+.. __: https://github.com/pyTMD/pyTMD/blob/main/pyTMD/io/dataset.py
+
+General Attributes and Methods
+==============================
+
+.. autoclass:: pyTMD.io.dataset
+   :members:
