### 2025-08-28 cfef555
test: add xarray accessor checks
