### 2025-08-28 0731130
feat: add basic xarray accessor for tide model data (#456)
* feat: add basic accessor for tide model data

* test: add xarray accessor checks
