+++ pyTMD/io/__init__.py
@@ -15,3 +15,9 @@ from .model import (
     read_constants,
     interpolate_constants
 )
+
+# Attempt to import the dataset module for xarray support
+try:
+    from .dataset import dataset
+except (ImportError, AttributeError):
+    dataset = None
