+++ pyTMD/interpolate.py
@@ -14,6 +14,8 @@ PYTHON DEPENDENCIES:
 UPDATE HISTORY:
     Updated 08/2025: added vectorized 1D linear interpolation function
         improve performance of bilinear interpolation and allow extrapolation
+        added a penalized least square inpainting function to gap fill data
+        standardized most variable names between interpolation functions
     Updated 09/2024: deprecation fix case where an array is output to scalars
     Updated 07/2024: changed projection flag in extrapolation to is_geographic
     Written 12/2022
@@ -21,12 +23,14 @@ UPDATE HISTORY:
 from __future__ import annotations
 
 import numpy as np
+import scipy.fftpack
 import scipy.spatial
 import scipy.interpolate
 import pyTMD.spatial
 
 __all__ = [
     "interp1d",
+    "inpaint",
     "bilinear",
     "spline",
     "regulargrid",
@@ -39,7 +43,8 @@ def interp1d(
         x: float,
         xp: np.ndarray,
         fp: np.ndarray,
-        extrapolate: str = 'linear'
+        extrapolate: str = 'linear',
+        **kwargs
     ):
     """
     Vectorized one-dimensional linear interpolation
@@ -77,32 +82,118 @@ def interp1d(
     f = (1.0 - d)*fp[:,j] + d*fp[:,j+1]
     return f
 
+def inpaint(
+        xs: np.ndarray,
+        ys: np.ndarray,
+        zs: np.ndarray,
+        N: int = 0,
+        s0: int = 3,
+        power: int = 2,
+        epsilon: float = 2.0,
+        **kwargs
+    ):
+    """
+    Inpaint over missing data in a two-dimensional array using a
+    penalized least square method based on discrete cosine transforms
+    :cite:p:`Garcia:2010hn,Wang:2012ei`
+
+    Parameters
+    ----------
+    xs: np.ndarray
+        input x-coordinates
+    ys: np.ndarray
+        input y-coordinates
+    zs: np.ndarray
+        input data
+    N: int, default 100
+        Number of iterations (0 for nearest neighbors)
+    s0: int, default 3
+        Smoothing
+    power: int, default 2
+        power for lambda function
+    epsilon: float, default 2.0
+        relaxation factor
+    """
+    # find masked values
+    if isinstance(zs, np.ma.MaskedArray):
+        W = np.logical_not(zs.mask)
+    else:
+        W = np.isfinite(zs)
+    # no valid values can be found
+    if not np.any(W):
+        raise ValueError('No valid values found')
+
+    # dimensions of input grid
+    ny, nx = np.shape(zs)
+    # calculate lambda function
+    L = np.zeros((ny, nx))
+    L += np.broadcast_to(np.cos(np.pi*np.arange(ny)/ny)[:, None], (ny, nx))
+    L += np.broadcast_to(np.cos(np.pi*np.arange(nx)/nx)[None, :], (ny, nx))
+    LAMBDA = np.power(2.0*(2.0 - L), power)
+
+    # calculate initial values using nearest neighbors
+    # computation of distance Matrix
+    # use scipy spatial KDTree routines
+    xgrid, ygrid = np.meshgrid(xs, ys)
+    tree = scipy.spatial.cKDTree(np.c_[xgrid[W], ygrid[W]])
+    # find nearest neighbors
+    masked = np.logical_not(W)
+    _, ii = tree.query(np.c_[xgrid[masked], ygrid[masked]], k=1)
+    # copy valid original values
+    z0 = np.zeros_like(zs)
+    z0[W] = np.copy(zs[W])
+    # copy nearest neighbors
[…]
