### 2025-08-29 a4644ce
docs: add gap fill notebook to examples
fix: suppress crs warnings about proj conversion
