+++ pyTMD/data/w1990_tab2.txt
@@ -1,0 +1,99 @@
+ 0  0  0  1  0  0 -0.00021 055.655
+ 0  0  2 -1  0  0 -0.00004 057.455
+ 0  1 -2  0  0  0  0.00004 063.555
+ 0  1  0  0 -1  0  0.00019 065.545
+ 0  1  0  0  0  0 -0.00375 065.555
+ 0  1  0  0  1  0 -0.00059 065.565
+ 0  1  0  0  2  0  0.00005 065.575
+ 0  2 -2  1  0  0 -0.00012 073.655
+ 0  2  0 -1  0  0 -0.00061 075.455
+ 0  2  0 -1  1  0 -0.00010 075.465
+ 0  3 -2  0  0  0 -0.00010 083.555
+ 0  3  0 -2  0  0 -0.00007 085.355
+ 0  3  0  0  0  0 -0.00030 085.555
+ 0  3  0  0  1  0 -0.00019 085.565
+ 0  3  0  0  2  0 -0.00004 085.575
+ 0  4  0 -1  0  0 -0.00008 095.455
+ 0  4  0 -1  1  0 -0.00005 095.465
+ 1 -4  0  2  0  0 -0.00006 115.755
+ 1 -4  2  0  0  0 -0.00006 117.555
+ 1 -3  0  1 -1  0 -0.00014 125.645
+ 1 -3  0  1  0  0 -0.00035 125.655
+ 1 -3  2 -1  0  0 -0.00007 127.455
+ 1 -2  0  0 -2  0 -0.00004 135.535
+ 1 -2  0  0 -1  0 -0.00050 135.545
+ 1 -2  0  0  0  0 -0.00128 135.555
+ 1 -2  0  2  0  0 -0.00008 135.755
+ 1 -2  2  0  0  0 -0.00011 137.555
+ 1 -1  0 -1  0  0  0.00007 145.455
+ 1 -1  0  1 -1  0  0.00010 145.645
+ 1 -1  0  1  0  0 -0.00065 145.655
+ 1 -1  0  1  0  0  0.00009 145.655
+ 1 -1  2 -1  0  0 -0.00013 147.455
+ 1  0  0  0 -1  0  0.00059 155.545
+ 1  0  0  0  0  0 -0.00399 155.555
+ 1  0  0  0  1  0  0.00052 155.565
+ 1  1 -2  1  0  0 -0.00004 163.655
+ 1  1  0 -1 -1  0  0.00003 165.445
+ 1  1  0 -1  0  0 -0.00022 165.455
+ 1  1  0 -1  1  0  0.00003 165.465
+ 1  1  0  1  0  0 -0.00008 165.655
+ 1  1  0  1  1  0 -0.00003 165.665
+ 1  2 -2  0  0  0 -0.00005 173.555
+ 1  2  0  0 -1  0  0.00005 175.545
+ 1  2  0  0  0  0 -0.00146 175.555
+ 1  2  0  0  1  0 -0.00059 175.565
+ 1  2  0  0  2  0 -0.00005 175.575
+ 1  3 -2  1  0  0 -0.00005 183.655
+ 1  3  0 -1  0  0 -0.00024 185.455
+ 1  3  0 -1  1  0 -0.00010 185.465
+ 1  4 -2  0  0  0 -0.00004 193.555
+ 1  4  0  0  0  0 -0.00005 195.555
+ 1  4  0  0  1  0 -0.00005 195.565
+ 2 -4  2  1  0  0 -0.00006 217.655
+ 2 -3  0  2  0  0 -0.00018 225.755
+ 2 -3  2  0 -1  0 -0.00003 227.545
+ 2 -3  2  0  0  0 -0.00018 227.555
[…]
