+++ pyTMD/data/w1990_tab1.txt
@@ -1,0 +1,387 @@
+ 0  0  0  0  0  0 -0.31458 055.555
+ 0  0  0  0  1  0  0.02793 055.565
+ 0  0  0  0  2  0 -0.00028 055.575
+ 0  0  0  2  1  0  0.00004 055.765
+ 0  0  1  0 -1 -1 -0.00004 056.544
+ 0  0  1  0  0 -1 -0.00492 056.554
+ 0  0  1  0  0  1  0.00026 056.556
+ 0  0  1  0  1 -1  0.00005 056.564
+ 0  0  2 -2 -1  0  0.00002 057.345
+ 0  0  2 -2  0  0 -0.00032 057.355
+ 0  0  2  0  0  0 -0.03098 057.555
+ 0  0  2  0  0 -2 -0.00012 057.553
+ 0  0  2  0  1  0  0.00077 057.565
+ 0  0  2  0  2  0  0.00017 057.575
+ 0  0  3  0  0 -1 -0.00181 058.554
+ 0  0  3  0  1 -1  0.00003 058.564
+ 0  0  4  0  0 -2 -0.00007 059.553
+ 0  1 -3  1 -1  1  0.00002 062.646
+ 0  1 -3  1  0  1 -0.00029 062.656
+ 0  1 -3  1  1  1  0.00000 062.666
+ 0  1 -2 -1 -2  0  0.00002 063.435
+ 0  1 -2 -1 -1  0  0.00007 063.445
+ 0  1 -2  1 -1  0  0.00048 063.645
+ 0  1 -2  1  0  0 -0.00673 063.655
+ 0  1 -2  1  1  0  0.00044 063.665
+ 0  1 -1 -1 -1  1  0.00000 064.446
+ 0  1 -1 -1  0  1 -0.00022 064.456
+ 0  1 -1  0  0  0  0.00020 064.555
+ 0  1 -1  1  0 -1  0.00005 064.654
+ 0  1  0 -1 -2  0 -0.00003 065.435
+ 0  1  0 -1 -1  0  0.00231 065.445
+ 0  1  0 -1  0  0 -0.03518 065.455
+ 0  1  0 -1  1  0  0.00229 065.465
+ 0  1  0  1  0  0  0.00188 065.655
+ 0  1  0  1  1  0  0.00077 065.665
+ 0  1  0  1  2  0  0.00021 065.675
+ 0  1  1 -1  0 -1  0.00018 066.454
+ 0  1  2 -1  0  0  0.00049 067.455
+ 0  1  2 -1  1  0  0.00024 067.465
+ 0  1  2 -1  2  0  0.00004 067.475
+ 0  1  3 -1  0 -1  0.00002 068.454
+ 0  2 -4  2  0  0 -0.00011 071.755
+ 0  2 -3  0  0  1 -0.00038 072.556
+ 0  2 -3  0  1  1  0.00002 072.566
+ 0  2 -2  0 -1  0 -0.00042 073.545
+ 0  2 -2  0  0  0 -0.00583 073.555
+ 0  2 -2  0  1  0  0.00038 073.565
+ 0  2 -2  2  0  0  0.00004 073.755
+ 0  2 -1 -2  0  1 -0.00004 074.356
+ 0  2 -1 -1  0  0  0.00003 074.455
+ 0  2 -1  0  0 -1  0.00006 074.554
+ 0  2 -1  0  0  1 -0.00020 074.556
+ 0  2 -1  0  1  1 -0.00004 074.566
+ 0  2  0 -2 -1  0  0.00015 075.345
+ 0  2  0 -2  0  0 -0.00288 075.355
[…]
