### 2025-09-03 9069a27
test: add check if running on GitHub Actions
https://github.com/openjournals/joss-reviews/issues/8566#issuecomment-3097482521
