+++ test/test_download_and_read.py
@@ -21,6 +21,7 @@ PYTHON DEPENDENCIES:
         https://pypi.org/project/timescale/
 
 UPDATE HISTORY:
+    Updated 09/2025: added check if running on GitHub Actions or locally
     Updated 08/2025: added xarray tests to verify implementation
     Updated 06/2025: subset to specific constituents when reading model
     Updated 12/2024: create test files from matlab program for comparison
@@ -49,6 +50,7 @@ UPDATE HISTORY:
         will install octave and oct2py in development requirements
     Written 08/2020
 """
+import os
 import re
 import io
 import copy
@@ -57,7 +59,6 @@ import json
 import boto3
 import shutil
 import pytest
-import inspect
 import pathlib
 import zipfile
 import posixpath
@@ -76,9 +77,8 @@ import timescale.time
 pd = pyTMD.utilities.import_dependency('pandas')
 oct2py = pyTMD.utilities.import_dependency('oct2py')
 
-# current file path
-filename = inspect.getframeinfo(inspect.currentframe()).filename
-filepath = pathlib.Path(filename).absolute().parent
+# check if running on GitHub Actions CI
+GITHUB_ACTIONS = os.environ.get('GITHUB_ACTIONS', False)
 
 # PURPOSE: calculate the matlab serial date from calendar date
 # http://scienceworld.wolfram.com/astronomy/JulianDate.html
@@ -92,6 +92,10 @@ def convert_calendar_serial(year, month, day, hour=0.0, minute=0.0, second=0.0):
 
 # PURPOSE: Test and Verify CATS2008 model read and prediction programs
 class Test_CATS2008:
+    @pytest.fixture(autouse=True)
+    def init(self, directory):
+        self.directory = pathlib.Path(directory).expanduser().absolute()
+
     # PURPOSE: Download CATS2008 from US Antarctic Program
     @pytest.fixture(scope="class", autouse=False)
     def download_CATS2008(self):
@@ -108,7 +112,7 @@ class Test_CATS2008:
         # verify that model files are within downloaded zip file
         assert all(m)
         # output tide directory for model
-        modelpath = filepath.joinpath('CATS2008')
+        modelpath = self.directory.joinpath('CATS2008')
         # extract each member (model and configuration files)
         for member in m:
             # strip directories from member filename
@@ -118,7 +122,7 @@ class Test_CATS2008:
         # close the zipfile object
         zfile.close()
         # output control file for tide model
-        CFname = filepath.joinpath('Model_CATS2008')
+        CFname = self.directory.joinpath('Model_CATS2008')
         fid = CFname.open(mode='w', encoding='utf8')
         for model_file in ['hf.CATS2008.out','uv.CATS2008.out','grid_CATS2008']:
             print(modelpath.joinpath(model_file), file=fid)
@@ -134,7 +138,7 @@ class Test_CATS2008:
         CFname.unlink(missing_ok=True)
 
     # PURPOSE: Download CATS2008 from AWS S3 bucket
-    @pytest.fixture(scope="class", autouse=True)
+    @pytest.fixture(scope="class", autouse=GITHUB_ACTIONS)
     def AWS_CATS2008(self,
             aws_access_key_id,
             aws_secret_access_key,
@@ -149,11 +153,11 @@ class Test_CATS2008:
         s3 = session.resource('s3')
         bucket = s3.Bucket('pytmd')
         # model parameters for CATS2008
-        modelpath = filepath.joinpath('CATS2008')
+        modelpath = self.directory.joinpath('CATS2008')
         # recursively create model directory
         modelpath.mkdir(parents=True, exist_ok=True)
[…]
