### 2025-09-03 89e2472
Update test_download_and_read.py
