+++ test/test_download_and_read.py
@@ -98,7 +98,7 @@ class Test_CATS2008:
 
     # PURPOSE: Download CATS2008 from US Antarctic Program
     @pytest.fixture(scope="class", autouse=False)
-    def download_CATS2008(self):
+    def download_CATS2008(self, directory):
         # download CATS2008 zip file and read as virtual file object
         HOST = ['https://www.usap-dc.org','dataset','usap-dc','601235',
             '2019-12-19T23:26:43.6Z','CATS2008.zip?dataset_id=601235']
@@ -112,7 +112,7 @@ class Test_CATS2008:
         # verify that model files are within downloaded zip file
         assert all(m)
         # output tide directory for model
-        modelpath = self.directory.joinpath('CATS2008')
+        modelpath = directory.joinpath('CATS2008')
         # extract each member (model and configuration files)
         for member in m:
             # strip directories from member filename
@@ -122,7 +122,7 @@ class Test_CATS2008:
         # close the zipfile object
         zfile.close()
         # output control file for tide model
-        CFname = self.directory.joinpath('Model_CATS2008')
+        CFname = directory.joinpath('Model_CATS2008')
         fid = CFname.open(mode='w', encoding='utf8')
         for model_file in ['hf.CATS2008.out','uv.CATS2008.out','grid_CATS2008']:
             print(modelpath.joinpath(model_file), file=fid)
@@ -140,6 +140,7 @@ class Test_CATS2008:
     # PURPOSE: Download CATS2008 from AWS S3 bucket
     @pytest.fixture(scope="class", autouse=GITHUB_ACTIONS)
     def AWS_CATS2008(self,
+            directory,
             aws_access_key_id,
             aws_secret_access_key,
             aws_region_name
@@ -153,11 +154,11 @@ class Test_CATS2008:
         s3 = session.resource('s3')
         bucket = s3.Bucket('pytmd')
         # model parameters for CATS2008
-        modelpath = self.directory.joinpath('CATS2008')
+        modelpath = directory.joinpath('CATS2008')
         # recursively create model directory
         modelpath.mkdir(parents=True, exist_ok=True)
         # output control file for tide model
-        CFname = self.directory.joinpath('Model_CATS2008')
+        CFname = directory.joinpath('Model_CATS2008')
         fid = CFname.open(mode='w', encoding='utf8')
         # retrieve each model file from s3
         for model_file in ['hf.CATS2008.out','uv.CATS2008.out','grid_CATS2008']:
@@ -191,11 +192,11 @@ class Test_CATS2008:
 
     # PURPOSE: Download Antarctic Tide Gauge Database from US Antarctic Program
     @pytest.fixture(scope="class", autouse=False)
-    def download_AntTG(self):
+    def download_AntTG(self, directory):
         # download Tide Gauge Database text file
         HOST = ['https://www.usap-dc.org','dataset','usap-dc','601358',
             '2020-07-10T19:50:08.8Z','AntTG_ocean_height_v1.txt?dataset_id=601358']
-        local = self.directory.joinpath('AntTG_ocean_height_v1.txt')
+        local = directory.joinpath('AntTG_ocean_height_v1.txt')
         pyTMD.utilities.from_http(HOST, local=local)
         assert local.exists()
         # run tests
@@ -206,6 +207,7 @@ class Test_CATS2008:
     # PURPOSE: Download Antarctic Tide Gauge Database from AWS
     @pytest.fixture(scope="class", autouse=GITHUB_ACTIONS)
     def AWS_AntTG(self,
+            directory,
             aws_access_key_id,
             aws_secret_access_key,
             aws_region_name
@@ -221,7 +223,7 @@ class Test_CATS2008:
[…]
