### 2025-09-03 10261f1
test: add check if running on GitHub Actions (#460)
https://github.com/openjournals/joss-reviews/issues/8566#issuecomment-3097482521
