+++ test/test_perth3_read.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 u"""
-test_perth3_read.py (08/2025)
+test_perth3_read.py (09/2025)
 Tests that GOT4.7 data can be downloaded from AWS S3 bucket
 Tests the read program to verify that constituents are being extracted
 Tests that interpolated results are comparable to NASA PERTH3 program
@@ -17,6 +17,7 @@ PYTHON DEPENDENCIES:
         https://pypi.org/project/timescale/
 
 UPDATE HISTORY:
+    Updated 09/2025: added check if running on GitHub Actions or locally
     Updated 08/2025: added xarray tests to verify implementation
     Updated 06/2025: subset to specific constituents when reading model
     Updated 09/2024: drop support for the ascii definition file format
@@ -37,14 +38,13 @@ UPDATE HISTORY:
         replaced numpy bool/int to prevent deprecation warnings
     Written 08/2020
 """
+import os
 import io
 import gzip
 import json
 import boto3
 import shutil
 import pytest
-import inspect
-import pathlib
 import posixpath
 import numpy as np
 import xarray as xr
@@ -55,13 +55,16 @@ import pyTMD.compute
 import pyTMD.predict
 import timescale.time
 
-# current file path
-filename = inspect.getframeinfo(inspect.currentframe()).filename
-filepath = pathlib.Path(filename).absolute().parent
+# check if running on GitHub Actions CI
+GITHUB_ACTIONS = os.environ.get('GITHUB_ACTIONS', False)
 
 # PURPOSE: Download GOT4.7 constituents from AWS S3 bucket
-@pytest.fixture(scope="module", autouse=True)
-def download_model(aws_access_key_id,aws_secret_access_key,aws_region_name):
+@pytest.fixture(scope="module", autouse=GITHUB_ACTIONS)
+def download_model(directory,
+        aws_access_key_id,
+        aws_secret_access_key,
+        aws_region_name
+    ):
     # get aws session object
     session = boto3.Session(
         aws_access_key_id=aws_access_key_id,
@@ -72,7 +75,7 @@ def download_model(aws_access_key_id,aws_secret_access_key,aws_region_name):
     bucket = s3.Bucket('pytmd')
 
     # model parameters for GOT4.7
-    model = pyTMD.io.model(filepath,compressed=True,
+    model = pyTMD.io.model(directory, compressed=True,
         verify=False).elevation('GOT4.7')
     # recursively create model directory
     model_directory = model.model_file[0].parent
@@ -96,9 +99,9 @@ def download_model(aws_access_key_id,aws_secret_access_key,aws_region_name):
 @pytest.mark.parametrize("METHOD", ['spline','linear','bilinear'])
 @pytest.mark.parametrize("CROP", [False, True])
 # PURPOSE: Tests that interpolated results are comparable to PERTH3 program
-def test_verify_GOT47(METHOD, CROP):
+def test_verify_GOT47(directory, METHOD, CROP):
     # model parameters for GOT4.7
-    model = pyTMD.io.model(filepath,compressed=True).elevation('GOT4.7')
+    model = pyTMD.io.model(directory,compressed=True).elevation('GOT4.7')
     # perth3 test program infers m4 tidal constituent
     # constituent files included in test
     constituents = ['q1','o1','p1','k1','n2','m2','s2','k2','s1']
@@ -106,7 +109,7 @@ def test_verify_GOT47(METHOD, CROP):
     model.scale = 1.0
 
     # read validation dataset
[…]
