+++ MANIFEST.in
@@ -2,6 +2,7 @@ graft pyTMD/data
 prune .github*
 prune doc*
 prune notebooks*
+prune providers*
 prune run*
 prune test*
 exclude *.cfg

[…]
