### 2025-09-03 c6fe4bc
docs: add project testing readme
https://github.com/openjournals/joss-reviews/issues/8566#issuecomment-3097482521
