+++ doc/source/index.rst
@@ -152,6 +152,7 @@ Project Details
 
     project/Contributors.rst
     project/Licenses.rst
+    project/Testing.rst
     project/Citations.rst
 
 .. toctree::
