### 2025-09-03 8b67806
fix: no parallelization
