### 2025-09-03 42a5ce6
docs: add project testing readme (#461)
https://github.com/openjournals/joss-reviews/issues/8566#issuecomment-3097482521
