### 2025-09-04 fd695ff
feat: can choose different tide potential catalogs for body tides
refactor: remake tide potential catalogs to include degree
