### 2025-09-04 a1f3ad9
docs: add planetary_longitudes
