+++ pyTMD/astro.py
@@ -265,42 +265,42 @@ def planetary_longitudes(MJD: np.ndarray):
 
     Returns
     -------
-    LM: np.ndarray
+    LMe: np.ndarray
         mean longitude of Mercury (degrees)
-    LV: np.ndarray
+    LVe: np.ndarray
         mean longitude of Venus (degrees)
-    LA: np.ndarray
+    LMa: np.ndarray
         mean longitude of Mars (degrees)
-    LJ: np.ndarray
+    LJu: np.ndarray
         mean longitude of Jupiter (degrees)
-    LS: np.ndarray
+    LSa: np.ndarray
         mean longitude of Saturn (degrees)
     """
     # convert from MJD to centuries relative to 2000-01-01T12:00:00
     T = (MJD - _mjd_j2000)/_century
     # mean longitudes of Mercury
     mercury_longitude = np.array([252.250906, 149474.0722491, 3.035e-4, 1.8e-8])
-    LM = polynomial_sum(mercury_longitude, T)
+    LMe = polynomial_sum(mercury_longitude, T)
     # mean longitudes of Venus
     venus_longitude = np.array([181.9798001, 58519.2130302, 3.1014e-4, 1.5e-8])
-    LV = polynomial_sum(venus_longitude, T)
+    LVe = polynomial_sum(venus_longitude, T)
     # mean longitudes of Mars
     mars_longitude = np.array([355.433, 19141.6964471, 3.1052e-4, 1.e-8])
-    LA = polynomial_sum(mars_longitude, T)
+    LMa = polynomial_sum(mars_longitude, T)
     # mean longitudes of Jupiter
     jupiter_longitude = np.array([34.351519, 3036.3027748, 2.233e-4, 3.7e-8])
-    LJ = polynomial_sum(jupiter_longitude, T)
+    LJu = polynomial_sum(jupiter_longitude, T)
     # mean longitudes of Saturn
     saturn_longitude = np.array([50.077444, 1223.5110686, 5.1908-4, -3.0e-8])
-    LS = polynomial_sum(saturn_longitude, T)
+    LSa = polynomial_sum(saturn_longitude, T)
     # take the modulus of each
-    LM = normalize_angle(LM)
-    LV = normalize_angle(LV)
-    LA = normalize_angle(LA)
-    LJ = normalize_angle(LJ)
-    LS = normalize_angle(LS)
+    LMe = normalize_angle(LMe)
+    LVe = normalize_angle(LVe)
+    LMa = normalize_angle(LMa)
+    LJu = normalize_angle(LJu)
+    LSa = normalize_angle(LSa)
     # return as tuple
-    return (LM, LV, LA, LJ, LS)
+    return (LMe, LVe, LMa, LJu, LSa)
 
 # PURPOSE: computes the phase angles of astronomical means
 def phase_angles(MJD: np.ndarray):
