+++ doc/source/user_guide/zarr-to-s3-obstore.py
@@ -8,7 +8,7 @@ tide_model = 'FES2022'
 # setup tide model
 m = pyTMD.io.model(compressed=False)
 # read tide model and convert to xarray DataTree
-dtree = m.to_datatree(tide_model, gap_fill=True)
+dtree = m.to_datatree(tide_model, gap_fill=True, N=100)
 
 # setup zarr store using obstore
 presigned_s3_url = f's3://{s3_bucket}/{m.name}.zarr'
