### 2025-09-04 f5b5993
docs: add citations for tide potential catalogs
