### 2025-09-04 535dea8
feat: can choose different tide potential catalogs for body tides (#462)
refactor: remake tide potential catalogs to include degree
