+++ pyTMD/predict.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 predict.py
-Written by Tyler Sutterley (08/2025)
+Written by Tyler Sutterley (09/2025)
 Prediction routines for ocean, load, equilibrium and solid earth tides
 
 REFERENCES:
@@ -22,6 +22,8 @@ PROGRAM DEPENDENCIES:
     spatial.py: utilities for working with geospatial data
 
 UPDATE HISTORY:
+    Updated 09/2025: make permanent tide amplitude an input parameter
+        can choose different tide potential catalogs for body tides
     Updated 08/2025: add simplified solid earth tide prediction function
         add correction of anelastic effects for long-period body tides
         use sign convention from IERS for complex body tide Love numbers
@@ -1798,11 +1800,12 @@ def _frequency_dependence_long_period(
 def _free_to_mean(
         XYZ: np.ndarray,
         h2: float | np.ndarray,
-        l2: float | np.ndarray
+        l2: float | np.ndarray,
+        H0: float = -0.31460
     ):
     """
     Calculate offsets for converting the permanent tide from
-    a tide-free to a mean-tide state
+    a tide-free to a mean-tide state :cite:p:`Mathews:1997js`
 
     Parameters
     ----------
@@ -1812,6 +1815,8 @@ def _free_to_mean(
         Degree-2 Love number of vertical displacement
     l2: float or np.ndarray
         Degree-2 Love (Shida) number of horizontal displacement
+    H0: float, default -0.31460
+        Mean amplitude of the permanent tide (meters)
     """
     # Compute the normalized position vector of coordinates
     radius = np.sqrt(np.sum(XYZ**2, axis=1))
@@ -1819,8 +1824,6 @@ def _free_to_mean(
     cosphi = np.sqrt(XYZ[:,0]**2 + XYZ[:,1]**2)/radius
     sinla = XYZ[:,1]/cosphi/radius
     cosla = XYZ[:,0]/cosphi/radius
-    # time-independent constituent of amplitude (Mathews et al. 1997)
-    H0 = -0.31460
     # in Mathews et al. (1997): dR0=-0.1196 m with h2=0.6026
     dR0 = np.sqrt(5.0/(4.0*np.pi))*h2*H0
     # in Mathews et al. (1997): dN0=-0.0247 m with l2=0.0831
@@ -1835,6 +1838,18 @@ def _free_to_mean(
     # return the corrections
     return np.c_[DX, DY, DZ]
 
+# tide potential tables
+_tide_potential_table = {}
+# Cartwright and Tayler (1971) table with 3rd-degree values
+# Cartwright and Edden (1973) table with updated values
+_tide_potential_table['CTE1973'] = pyTMD.arguments._cte1973_table
+# Hartmann and Wenzel (1995) tidal potential catalog
+_tide_potential_table['HW1995'] = pyTMD.arguments._hw1995_table
+# Tamura (1987) tidal potential catalog
+_tide_potential_table['T1987'] = pyTMD.arguments._t1987_table
+# Woodworth (1990) tables with updated and 3rd-degree values
+_tide_potential_table['W1990'] = pyTMD.arguments._w1990_table
+
 # PURPOSE: estimate solid Earth tides due to gravitational attraction
 # using a simplified approach based on Cartwright and Tayler (1971)
 def body_tide(
@@ -1844,6 +1859,7 @@ def body_tide(
         deltat: float | np.ndarray = 0.0,
         method: str = 'ASTRO5',
         tide_system: str = 'tide_free',
+        catalog: str = 'CTE1973',
         **kwargs
     ):
     """
@@ -1874,6 +1890,15 @@ def body_tide(
 
[…]
