### 2025-09-04 ecc9a11
feat: include complex latitudinal dependence
