### 2025-09-05 465ce18
feat: include complex latitudinal dependence (#463)
