### 2025-09-05 5c27a73
feat: add wrapper for catalog-based SET
refactor: split IERS ephemeride method into a separate function
docs: update solid earth tide description
