### 2025-09-05 8cd480c
feat: add wrapper for catalog-based SET (#464)
refactor: split IERS ephemeride method into a separate function
docs: update solid earth tide description
