+++ pyTMD/compute.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 compute.py
-Written by Tyler Sutterley (08/2025)
+Written by Tyler Sutterley (09/2025)
 Calculates tidal elevations for correcting elevation or imagery data
 Calculates tidal currents at locations and times
 
@@ -28,10 +28,11 @@ Calculates radial ocean pole load tide displacements following IERS Convention
 (2010) guidelines for correcting elevation or imagery data
     https://iers-conventions.obspm.fr/chapter7.php
 
-Ocean Pole Tides (SET)
+Solid Earth Tides (SET)
 Calculates radial Solid Earth tide displacements following IERS Convention
 (2010) guidelines for correcting elevation or imagery data
     https://iers-conventions.obspm.fr/chapter7.php
+Or by using a tide potential catalog following Cartwright and Tayler (1971)
 
 PYTHON DEPENDENCIES:
     numpy: Scientific Computing Tools For Python
@@ -62,6 +63,8 @@ PROGRAM DEPENDENCIES:
     interpolate.py: interpolation routines for spatial data
 
 UPDATE HISTORY:
+    Updated 09/2025: added wrapper for calculating solid earth tides
+        using a tide potential catalog following Cartwright and Tayler (1971)
     Updated 08/2025: convert angles with numpy radians and degrees functions
         pass kwargs to computation of long-period equilibrium tides
         use timescale shortcut wrapper functions to create Timescale objects
@@ -1343,13 +1346,49 @@ def OPT_displacements(
 # PURPOSE: compute solid earth tidal elevations
 def SET_displacements(
         x: np.ndarray, y: np.ndarray, delta_time: np.ndarray,
+        METHOD: str = 'ephemerides',
+        **kwargs
+    ):
+    """
+    Compute solid earth tidal elevations (body tides) at points and times
+
+    Parameters
+    ----------
+    x: np.ndarray
+        x-coordinates in projection EPSG
+    y: np.ndarray
+        y-coordinates in projection EPSG
+    delta_time: np.ndarray
+        seconds since EPOCH or datetime array
+    METHOD: str, default 'IERS'
+        method for calculating solid earth tidal elevations
+
+            - ``'ephemerides'``: following :cite:t:`Petit:2010tp` guidelines
+            - ``'catalog'``: using tide potential catalogs
+    """
+    if (METHOD.lower() == 'ephemerides'):
+        return _ephemeride_SET(
+            x, y, delta_time,
+            **kwargs
+        )
+    elif (METHOD.lower() == 'catalog'):
+        return _catalog_SET(
+            x, y, delta_time,
+            **kwargs
+        )
+    else:
+        raise ValueError(f"Invalid METHOD: {METHOD}")
+
+# PURPOSE: compute solid earth tides following IERS conventions
+def _ephemeride_SET(
+        x: np.ndarray, y: np.ndarray, delta_time: np.ndarray,
         EPSG: str | int = 4326,
         EPOCH: list | tuple = (2000, 1, 1, 0, 0, 0),
         TYPE: str | None = 'drift',
         TIME: str = 'UTC',
         ELLIPSOID: str = 'WGS84',
-        TIDE_SYSTEM='tide_free',
-        EPHEMERIDES='approximate',
+        TIDE_SYSTEM: str = 'tide_free',
+        EPHEMERIDES: str = 'approximate',
         **kwargs
     ):
     """
@@ -1516,6 +1555,162 @@ def SET_displacements(
                 tide_system=TIDE_SYSTEM)
             # calculate components of solid earth tides
             SE = np.einsum('ti...,ji...->tj...', dxi, R[s,:,:])
[…]
