### 2025-09-05 bc68c83
chore: bump version to `2.2.8` (#465)
