+++ doc/source/release_notes/release-v2.2.8.rst
@@ -1,0 +1,36 @@
+##################
+`Release v2.2.8`__
+##################
+
+* ``docs``: convert ``README`` to markdown `(#450) <https://github.com/pyTMD/pyTMD/pull/450>`_
+* ``docs``: update contribution guidelines `(#450) <https://github.com/pyTMD/pyTMD/pull/450>`_
+* ``docs``: add some admonitions `(#450) <https://github.com/pyTMD/pyTMD/pull/450>`_
+* ``docs``: add more notes and admonitions `(#451) <https://github.com/pyTMD/pyTMD/pull/451>`_
+* ``docs``: update contribution guidelines `(#451) <https://github.com/pyTMD/pyTMD/pull/451>`_
+* ``refactor``:  use ``timescale`` shortcut wrapper function `(#453) <https://github.com/pyTMD/pyTMD/pull/453>`_
+* ``feat``: add functions for converting models to ``xarray`` `(#455) <https://github.com/pyTMD/pyTMD/pull/455>`_
+* ``docs``: add s3 recipes and examples `(#455) <https://github.com/pyTMD/pyTMD/pull/455>`_
+* ``chore``: add aws feature dependencies to pixi `(#455) <https://github.com/pyTMD/pyTMD/pull/455>`_
+* ``chore``: update pixi lock file `(#455) <https://github.com/pyTMD/pyTMD/pull/455>`_
+* ``docs``: add pixi task that can create HTML documentation `(#455) <https://github.com/pyTMD/pyTMD/pull/455>`_
+* ``feat``: use vectorized linear interpolator in ``infer_minor`` `(#455) <https://github.com/pyTMD/pyTMD/pull/455>`_
+* ``feat``: improve performance of ``bilinear`` interpolation `(#455) <https://github.com/pyTMD/pyTMD/pull/455>`_
+* ``test``: add ``xarray`` tests `(#455) <https://github.com/pyTMD/pyTMD/pull/455>`_
+* ``ci``: add ``xarray`` to test environment `(#455) <https://github.com/pyTMD/pyTMD/pull/455>`_
+* ``feat``: add basic ``xarray`` accessor for tide model data `(#456) <https://github.com/pyTMD/pyTMD/pull/456>`_
+* ``test``: add ``xarray`` accessor checks `(#456) <https://github.com/pyTMD/pyTMD/pull/456>`_
+* ``feat``: added option to gap fill constituent grids `(#457) <https://github.com/pyTMD/pyTMD/pull/457>`_
+* ``docs``: add gap fill notebook to examples `(#458) <https://github.com/pyTMD/pyTMD/pull/458>`_
+* ``fix``: suppress ``crs`` warnings about ``proj`` conversion `(#458) <https://github.com/pyTMD/pyTMD/pull/458>`_
+* ``test``: add inpaint gap filling test `(#459) <https://github.com/pyTMD/pyTMD/pull/459>`_
+* ``feat``: add Woodworth tide potential tables `(#459) <https://github.com/pyTMD/pyTMD/pull/459>`_
+* ``test``: add check if running tests on GitHub Actions `(#460) <https://github.com/pyTMD/pyTMD/pull/460>`_
+* ``docs``: add project testing readme `(#461) <https://github.com/pyTMD/pyTMD/pull/461>`_
+* ``feat``: can choose different tide potential catalogs for body tides `(#462) <https://github.com/pyTMD/pyTMD/pull/462>`_
+* ``refactor``: remake tide potential catalogs to include degree `(#462) <https://github.com/pyTMD/pyTMD/pull/462>`_
+* ``feat``: include complex latitudinal dependence `(#463) <https://github.com/pyTMD/pyTMD/pull/463>`_
[…]
