+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pyTMD"
-version = "2.2.7"
+version = "2.2.8"
 description = "Python-based tidal prediction software for estimating ocean, load, solid Earth and pole tides"
 keywords = [
     "Ocean Tides",
