### 2025-09-08 #8566 @rcaneill
I also feel that the beginning of the "state of need" part of the paper is quite hard to read as it a list of tools. I don’t believe that you need to add all the name of the developers, this would lighten the paragraph. Moreover, instead of starting as a list, you could start with something like:
> Tide models output fields that need to be post processed to compute e.g. height deflection or currents at particular location. Each software has been created and tailored for one specific model each and cannot be used with other models (e.g. OTIS, TMD, etc). pyTMD is a generalized tide program that can read all/almost all/50 model output [...]

This phrasing is only a proposition, but one way or another you should clarify the summary+statement of need part
