### 2025-09-08 #8566 @rcaneill
I let the comments unchecked on https://github.com/openjournals/joss-reviews/issues/8566#issuecomment-3097482521 as my 2 previous comments correspond to my previous checklist. I’ll check when my comments are answered
