### 2025-09-08 4490277
docs: add blurbs about the harmonic method
docs: add tide potential catalog table
