+++ doc/source/background/Tides.rst
@@ -14,8 +14,19 @@ Ocean tides can be observed using float gauges, GPS stations, gravimeters, tiltm
     Tide gauges measure the height of the ocean surface relative to the land upon which they are situated (*ocean tides only*).
     Satellite altimeters measure the height of the ocean surface relative to the center of mass of the Earth system (*combination of ocean and earth tides*).
 
-Tidal oscillations for both ocean and load tides can be decomposed into a series of tidal constituents (or partial tides) of particular frequencies that are associated with the relative positions of the sun, moon and Earth.
-These tidal constituents are typically classified into different "species" based on their approximate period: short-period, semi-diurnal, diurnal, and long-period [see :ref:`tab-constituents`].
+``pyTMD`` uses the harmonic method for tide prediction, which is based on harmonic analysis and decomposition.
+Here, tidal oscillations for both ocean and load tides are decomposed into a series of tidal constituents (or partial tides) of particular frequencies.
+These frequencies are associated with particular astronomical forcings, which are largely based on the relative positions of the sun, moon and Earth.
+The tide height (or current) at any time can be estimated through a summation of all available tidal constituents :cite:p:`Doodson:1921kt,Foreman:1989dt`:
+
+.. math::
+    :label: 1.1
+    :name: eq:1.1
+
+    h(t) = z_0 +\sum_{k=1}^{n} f_k(t) \left[A_k\cos{\left(G_k(t) + u_k(t) + \theta_k\right)} \right]
+
+where :math:`z_0` is the datum offset, :math:`k` is the constituent number, :math:`n` is the total number of considered constituents, :math:`A_k` and :math:`\theta_k` are the constituent amplitude and phase lag provided by the tide model, :math:`G_k` is the equilibrium phase [see :ref:`Equation 1.2 <eq:1.2>`], :math:`f_k(t)` and :math:`u_k(t)` are the nodal amplitude and phase modulations [see :term:`Nodal Corrections`].
+Tidal constituents are typically classified into different "species" based on their approximate period: short-period, semi-diurnal, diurnal, and long-period [see :ref:`tab-constituents`].
 
 .. plot:: ./background/spectra.py
     :show-source-link: False
@@ -47,12 +58,13 @@ This resonance affects the instantaneous elastic response of the solid Earth to 
 ``pyTMD`` uses the astronomical argument formalism outlined in :cite:t:`Doodson:1921kt` for the prediction of ocean and load tides. 
 For any given time, ``pyTMD.astro`` calculates the longitudes of the moon (:math:`S`), sun (:math:`H`), lunar perigree (:math:`P`), ascending lunar node (:math:`N`) and solar perigree (:math:`Ps`), which are used in combination with the lunar hour angle (:math:`\tau`) and the extended Doodson number (:math:`k`) in a seven-dimensional Fourier series :cite:p:`Doodson:1921kt,Dietrich:1980ua,Pugh:2014di`.
 Each constituent has a particular "Doodson number" describing the polynomial coefficients of each of these astronomical terms in the Fourier series :cite:p:`Doodson:1921kt`. 
+These can be summed together to estimate the equilibrium phase (:math:`G`).
 
 .. math::
-    :label: 1.1
-    :name: eq:1.1
+    :label: 1.2
+    :name: eq:1.2
 
-    \sigma(t) = d_1\tau + d_2 S + d_3 H + d_4 P + d_5 N + d_6 Ps + d_7 k
+    G(t) = d_1\tau + d_2 S + d_3 H + d_4 P + d_5 N + d_6 Ps + d_7 k
 
 .. tip::
 
@@ -71,10 +83,11 @@ The tidal deformation of the Earth is to a very high degree instantaneous, with 
[…]
