+++ doc/source/_assets/pytmd-refs.bib
@@ -78,6 +78,14 @@ doi = {10.1179/sre.1985.28.218.202},
 pages = {202--206},
 publisher = {Informa UK Limited},
 }
+@article{Bullesfeld:1985tj,
+author = {B{\"u}llesfeld, F-J},
+title = {{Ein Beitrag zur harmonischen Darstellung des gezeitenerzeugenden Potentials}},
+journal = {Deutsche Geodaetische Kommission},
+year = {1985},
+volume = {314},
+note = {Reihe C},
+}
 @article{Capitaine:2003fx,
 author = {Capitaine, N and Wallace, P T and Chapront, J},
 title = {{Expressions for IAU 2000 precession quantities}},
@@ -314,17 +322,6 @@ doi = {10.21105/joss.06683},
 pages = {6683},
 publisher = {The Open Journal},
 }
-@article{Freer:2023bt,
-author = {Freer, B I D and Marsh, O J and Hogg, A E and Fricker, H A and Padman, L},
-title = {{Modes of Antarctic tidal grounding line migration revealed by Ice, Cloud, and land Elevation Satellite-2 (ICESat-2) laser altimetry}},
-journal = {The Cryosphere},
-year = {2023},
-volume = {17},
-number = {9},
-url = {https://tc.copernicus.org/articles/17/4079/2023/},
-doi = {10.5194/tc-17-4079-2023},
-pages = {4079--4101},
-}
 @article{Foreman:1989dt,
 author = {Foreman, M G and Henry, R F},
 title = {{The harmonic analysis of tidal model time series}},
@@ -352,6 +349,17 @@ url = {https://doi.org/10.1175/2008jtecho615.1},
 doi = {10.1175/2008JTECHO615.1},
 pages = {806--817},
 publisher = {American Meteorological Society},
+}
+@article{Freer:2023bt,
+author = {Freer, B I D and Marsh, O J and Hogg, A E and Fricker, H A and Padman, L},
+title = {{Modes of Antarctic tidal grounding line migration revealed by Ice, Cloud, and land Elevation Satellite-2 (ICESat-2) laser altimetry}},
+journal = {The Cryosphere},
+year = {2023},
+volume = {17},
+number = {9},
+url = {https://tc.copernicus.org/articles/17/4079/2023/},
+doi = {10.5194/tc-17-4079-2023},
+pages = {4079--4101},
 }
 @article{Garcia:2010hn,
 author = {Garcia, D},
@@ -436,7 +444,7 @@ month = dec,
 volume = {22},
 number = {24},
 issn = {0094-8276},
-url = {http://dx.doi.org/10.1029/95gl03324},
+url = {https://doi.org/10.1029/95gl03324},
 doi = {10.1029/95GL03324},
 pages = {3553--3556},
 publisher = {American Geophysical Union (AGU)},
@@ -1085,6 +1093,19 @@ url = {https://doi.org/10.1016/j.envsoft.2011.10.015},
 doi = {10.1016/j.envsoft.2011.10.015},
 pages = {139--142},
 publisher = {Elsevier BV},
+}
+@inbook{Wenzel:1997kn,
+author = {Wenzel, H-G},
+title = {{Tide-generating potential for the earth}},
+year = {1997},
+isbn = {978-3-540-68700-9},
+url = {https://doi.org/10.1007/BFb0011455},
+doi = {10.1007/BFb0011455},
+pages = {9--26},
+booktitle = {Tidal Phenomena},
+publisher = {Springer Berlin Heidelberg},
+address = {Berlin, Heidelberg},
+editor = {Wilhelm, Helmut},
 }
 @inproceedings{Woolard:1953wp,
 author = {Woolard, E W},
