+++ doc/source/background/Tide-Catalog-Table.rst
@@ -1,0 +1,27 @@
+:orphan:
+
+Tide Potential Catalogs
+-----------------------
+
+``pyTMD`` currently include support for the following tide potential catalogs:
+
+.. _tab-catalogs:
+
+.. list-table:: Tide Potential Catalogs
+    :header-rows: 1
+
+    * - Short Name
+      - Constituents
+      - Reference(s)
+    * - ``CTE1973``
+      - 484
+      - :cite:t:`Cartwright:1971iz` and :cite:t:`Cartwright:1973em`
+    * - ``HW1995``
+      - 12935
+      - :cite:t:`Hartmann:1995jp`
+    * - ``T1987``
+      - 1200
+      - :cite:t:`Tamura:1987tp`
+    * - ``W1990``
+      - 486
+      - Woodworth updates to ``CTE1973``
