### 2025-09-10 108984c
docs: improve summary and statement of need
at 1177 words though 👎
