+++ doc/source/_assets/paper.md
@@ -32,17 +32,18 @@ affiliations:
  - name: Department of Geophysics, Colorado School of Mines, United States
    index: 3
 
-date: 06 August 2025
+date: 10 September 2025
 bibliography: pytmd-refs.bib
 ---
 
 # Summary
 
-`pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and Earth tides. It is designed to be able to handle a wide range of model formats and can incorporate different physics schemes. This flexibility allows `pyTMD` to be tailored to fit specific use cases, while still allowing ease of use.
+`pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and Earth tides. It is not an ocean or load tide model, but a tool for using tidal constants provided by tide models to predict the height deflections or currents at particular locations and times. It is designed to be able to handle a wide range of model formats and can incorporate different physics schemes. This flexibility allows `pyTMD` to be tailored to fit specific use cases, while still allowing ease of use.
 
 # Statement of need
 
-There are several ocean tide prediction software options available. The OSU Tidal Inversion Software (`OTIS`) and OSU Tidal Prediction Software (`OTPS`) were developed by Lana Erofeeva and Gary Egbert. The Tidal Model Driver ([`TMD`](https://github.com/EarthAndSpaceResearch/TMD_Matlab_Toolbox_v2.5)) MATLAB Toolbox was developed by Laurie Padman, Lana Erofeeva and Susan Howard. An updated version of the MATLAB Toolbox ([`TMD3`](https://github.com/chadagreene/tide-model-driver)) was developed by Chad Greene. The NASA GSFC PREdict Tidal Heights (`PERTH3`) software was developed by Richard Ray and Remko Scharroo. An updated and more versatile version of the NASA GSFC Fortran software ([`PERTH5`](https://codeberg.org/rray/perth5)) was developed by Richard Ray. [`pyFES`](https://cnes.github.io/aviso-fes/) was produced by LEGOS, NOVELTIS and CLS Ocean and Climate Division, and funded by CNES [@Lyard:2025tr]. These software options are typically created by or for the model providers, and support their specific model formats.
+There are several ocean tide prediction software options available. The OSU Tidal Inversion Software (`OTIS`) and OSU Tidal Prediction Software (`OTPS`) are Fortran programs developed by Lana Erofeeva and Gary Egbert for the [TPXO](https://www.tpxo.net/home) family of models. The Tidal Model Driver ([`TMD`](https://github.com/EarthAndSpaceResearch/TMD_Matlab_Toolbox_v2.5)) is a MATLAB Toolbox developed by Laurie Padman, Lana Erofeeva and Susan Howard for the same family of models. An updated version of the MATLAB Toolbox ([`TMD3`](https://github.com/chadagreene/tide-model-driver)) was developed by Chad Greene. [`pyFES`](https://cnes.github.io/aviso-fes/) is a Python library produced by LEGOS, NOVELTIS and CLS Ocean and Climate Division, and funded by CNES [@Lyard:2025tr] for the Finite Element Solution (FES) family of models. 
+The NASA GSFC PREdict Tidal Heights (`PERTH3`) software is a Fortran program developed by Richard Ray and Remko Scharroo for the Goddard Ocean Tide (GOT) family of models. An updated and more versatile version of the NASA GSFC Fortran software ([`PERTH5`](https://codeberg.org/rray/perth5)) was developed by Richard Ray. `PERTH5` is a more generalized program, and can read from multiple different tide model formats. These software options are typically created by or for the model providers, and, with the exception of `PERTH5`, singly support their specific model formats.  
 
[…]
