### 2025-09-10 7ac4fd4
ci: add environmental variable for paper path
