+++ .github/workflows/joss-draft-pdf.yml
@@ -6,6 +6,9 @@ on:
       - doc/source/_assets/pytmd-refs.bib
       - .github/workflows/joss-draft-pdf.yml
 
+env:
+  PAPER_PATH: doc/source/_assets
+
 jobs:
   paper:
     runs-on: ubuntu-latest
@@ -18,7 +21,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: doc/source/_assets/paper.md
+          paper-path: ${{ env.PAPER_PATH }}/paper.md
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
@@ -26,6 +29,6 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: doc/source/_assets/paper.pdf
+          path: ${{ env.PAPER_PATH }}/paper.pdf
           retention-days: 5
           overwrite: true
