### 2025-09-10 18fd9da
docs: shorten paper
scrubbed the solution section.

background on creating solutions could be added to the documentation
