+++ doc/source/_assets/paper.md
@@ -53,13 +53,11 @@ The NASA GSFC PREdict Tidal Heights (`PERTH3`) software is a Fortran program dev
 
 ## Ocean and Load Tides
 
-Tides are frequently decomposed into harmonic constants, or constituents, associated with the relative positions of the sun, moon and Earth [@Padman:2018cv]. For ocean and load tides, `pyTMD.io` contains routines for reading major constituent values (amplitude and phase) from commonly available tide models, which typically fall within a few general formats: `OTIS-binary` [@Egbert:2002ge; @Padman:2008ec], `OTIS-compact`, `OTIS-netcdf`, `TMD3-netcdf` [@Greene:2024di], `GOT-ascii` [@Ray:1999vm], `GOT-netcdf`, `FES-ascii` [@LeProvost:1994ie] and `FES-netcdf` [@Lyard:2025tr]. Information for each of the supported tide models is stored within a JSON database. `pyTMD.io` reads or computes the amplitude and phase lag of tide model constituents, which are then interpolated to the output spatial coordinates. 
+With the harmonic method, tides are decomposed into harmonic constants, or constituents, associated with the relative positions of the sun, moon and Earth [@Padman:2018cv]. These constituents are typically classified into different "species" based on their approximate period: short-period, semi-diurnal, diurnal, and long-period. `pyTMD.io` contains routines for reading major constituent values (amplitude and phase) from commonly available tide models, which typically fall within a few general formats: `OTIS-binary` [@Egbert:2002ge; @Padman:2008ec], `OTIS-compact`, `OTIS-netcdf`, `TMD3-netcdf` [@Greene:2024di], `GOT-ascii` [@Ray:1999vm], `GOT-netcdf`, `FES-ascii` [@LeProvost:1994ie] and `FES-netcdf` [@Lyard:2025tr]. Information for each of the supported tide models is stored within a JSON database. For tidal predictions, `pyTMD.io` interpolates the amplitude and phase lag of tide model constituents to sets of spatial coordinates. 
 
-`pyTMD` uses the astronomical argument formalism outlined in @Doodson:1921kt to predict ocean and load tides. After the computation of the amplitude and phase lag for the spatial coordinates, `pyTMD` computes the temporal elements using by 1) calculating the astronomical angles ($S$, $H$, $P$, $N$, $P_s$) [@Meeus:1991vh; @Simon:1994vo], 2) combining these angles with the "Doodson numbers" in a Fourier series to compute the equilibrium tide phase ($G$) of each constituent, and 3) computing the 18.6-year nodal amplitude and phase corrections ($f$ and $u$) of each constituent [@Doodson:1921kt; @Dietrich:1980ua; @Pugh:2014di]. The spatial and temporal components are then combined, and the output tidal time series is calculated through a summation over all constituents [@Egbert:2002ge]. The contributions of "minor" constituents can then be estimated using inference methods [@Schureman:1958ty; @Ray:2017jx].
[…]
