### 2025-09-12 85c498d
docs: add spherical harmonic plots
docs: include links to spherical harmonics

docs: add second row for degree-2

docs: add projection text
