+++ doc/source/_assets/pytmd-refs.bib
@@ -859,6 +859,19 @@ doi = {10.1175/jtech-d-16-0142.1},
 pages = {437--446},
 publisher = {American Meteorological Society},
 }
+@article{Ray:2020gn,
+author = {Ray, R D},
+title = {{First global observations of third-degree ocean tides}},
+journal = {Science Advances},
+year = {2020},
+month = nov,
+volume = {6},
+number = {48},
+issn = {2375-2548},
+url = {https://doi.org/10.1126/sciadv.abd4744},
+doi = {10.1126/sciadv.abd4744},
+publisher = {American Association for the Advancement of Science (AAAS)},
+}
 @article{Ray:2022hx,
 author = {Ray, Richard D},
 title = {{Technical note: On seasonal variability of the M$_2$ tide}},
