### 2025-09-13 5488c7f
docs: add table of max tide potential
chore: update pixi lock
