### 2025-09-13 23558cd
Update Spherical-Harmonics.rst
