### 2025-09-23 3316c6c
refactor: generalize the calculation of body tides for degrees 3+
docs: added IERS Conventions references for Love number calculations
