### 2025-09-23 e70f5a6
shorten paper. still long
