### 2025-09-23 10c671d
remove developers to shorten paper
