+++ doc/source/_assets/paper.md
@@ -42,7 +42,7 @@ bibliography: pytmd-refs.bib
 
 # Statement of need
 
-There are several ocean tide prediction software options available. The OSU Tidal Inversion Software (`OTIS`) and OSU Tidal Prediction Software (`OTPS`) are Fortran programs developed by Lana Erofeeva and Gary Egbert for the [TPXO](https://www.tpxo.net/home) family of models. The Tidal Model Driver ([`TMD`](https://github.com/EarthAndSpaceResearch/TMD_Matlab_Toolbox_v2.5)) is a MATLAB Toolbox developed by Laurie Padman, Lana Erofeeva and Susan Howard for the same family of models. An updated version of the MATLAB Toolbox ([`TMD3`](https://github.com/chadagreene/tide-model-driver)) was developed by Chad Greene. [`pyFES`](https://cnes.github.io/aviso-fes/) is a Python library produced by LEGOS, NOVELTIS and CLS Ocean and Climate Division, and funded by CNES [@Lyard:2025tr] for the Finite Element Solution (FES) family of models. The NASA GSFC PREdict Tidal Heights (`PERTH3`) software is a Fortran program developed by Richard Ray and Remko Scharroo for the Goddard Ocean Tide (GOT) family of models. An updated and more versatile version of the NASA GSFC Fortran software ([`PERTH5`](https://codeberg.org/rray/perth5)) was developed by Richard Ray. `PERTH5` is a more generalized program, and can read from multiple different tide model formats. These software options are typically created by or for the model providers, and, with the exception of `PERTH5`, singly support their specific model formats.  
+There are several ocean tide prediction software options available. The OSU Tidal Inversion Software (`OTIS`) and OSU Tidal Prediction Software (`OTPS`) are Fortran programs developed for the [TPXO](https://www.tpxo.net/home) family of models [@Egbert:2002ge]. The Tidal Model Driver ([`TMD`](https://github.com/EarthAndSpaceResearch/TMD_Matlab_Toolbox_v2.5)) is a MATLAB Toolbox developed for the same family of models. An updated version of the MATLAB Toolbox ([`TMD3`](https://github.com/chadagreene/tide-model-driver)) was developed to use a custom consolidated netCDF4 format [@Greene:2024di]. [`pyFES`](https://cnes.github.io/aviso-fes/) is a Python library produced for the Finite Element Solution (FES) family of models [@Lyard:2025tr] funded by CNES. The NASA GSFC PREdict Tidal Heights (`PERTH3`) software is a Fortran program developed for the Goddard Ocean Tide (GOT) family of models. An updated and more versatile version of the NASA GSFC Fortran software ([`PERTH5`](https://codeberg.org/rray/perth5)) can read from multiple different tide model formats. These software options are typically created by or for the model providers, and, with the exception of `PERTH5`, singly support their specific model formats.  
 
 `pyTMD` is a generalized tide program that allows users to calculate both tide deflections and currents from a broad suite of models. Over 50 different models are presently supported within `pyTMD`, and additional model schemas can be defined with a JSON file. 
 
