### 2025-09-23 756b7de
italicize provided by the tide model
reorder pixi install in instructions
