+++ doc/source/getting_started/Install.rst
@@ -34,6 +34,9 @@ The simplest installation for most users will likely be using ``conda`` or ``mam
 
     conda update pytmd
 
+Development Install
+###################
+
 To use the development repository, please fork ``pyTMD`` into your own account and then clone onto your system:
 
 .. code-block:: bash
@@ -46,12 +49,6 @@ To use the development repository, please fork ``pyTMD`` into your own account a
 
     python3 -m pip install --user .
 
-or alternatively ``pixi`` for a `streamlined environment <https://pixi.sh/>`_:
-
-.. code-block:: bash
-
-    pixi install
-
 To include all optional dependencies:
 
 .. code-block:: bash
@@ -63,3 +60,22 @@ The development version of ``pyTMD`` can also be installed directly from GitHub 
 .. code-block:: bash
 
     python3 -m pip install --user git+https://github.com/pyTMD/pyTMD.git
+
+Package Management with ``pixi``
+################################
+
+Alternatively ``pixi`` can be used to create a `streamlined environment <https://pixi.sh/>`_ after cloning the repository:
+
+.. code-block:: bash
+
+    pixi install
+
+``pixi`` maintains isolated environments for each project, allowing for different versions of
+``pyTMD`` and its dependencies to be used without conflict. The ``pixi.lock`` file within the
+repository defines the required packages and versions for the environment.
+
+``pixi`` can also create shells for running programs within the environment:
+
+.. code-block:: bash
+
+    pixi shell
