### 2025-09-23 daa4cd5
refactor: generalize the calculation of body tides for degrees 3+ (#467)
docs: added IERS Conventions references for Love number calculations
