### 2025-09-24 d814146
chore: add more `ruff` lint parameters
