### 2025-09-29 d56b850
refactor: functional access to downloading scripts for #470
docs: expand on the directory structure
docs: update links (now functions) for downloading from ArcticData, GOT or AVISO for #469
