+++ pyTMD/datasets/__init__.py
@@ -1,0 +1,8 @@
+"""
+Input/output functions for reading and writing tidal data
+"""
+from .fetch_arcticdata import fetch_arcticdata
+from .fetch_aviso_fes import fetch_aviso_fes
+from .fetch_gsfc_got import fetch_gsfc_got
+from .reduce_otis import reduce_otis
+from .verify_box_tpxo import verify_box_tpxo
