+++ doc/source/api_reference/scripts/reduce_OTIS_files.rst
@@ -1,19 +1,0 @@
-====================
-reduce_OTIS_files.py
-====================
-
-- Reduces OTIS-format tidal files provided by Oregon State University and ESR to a regional subset
-
-`Source code`__
-
-.. __: https://github.com/pyTMD/pyTMD/blob/main/pyTMD/scripts/reduce_OTIS_files.py
-
-Calling Sequence
-################
-
-.. argparse::
-    :module: pyTMD.scripts.reduce_OTIS_files
-    :func: arguments
-    :prog: reduce_OTIS_files.py
-    :nodescription:
-    :nodefault:

[…]
