### 2025-09-29 836a701
docs: update function call for fetching GOT model
