+++ pyTMD/datasets/fetch_arcticdata.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 fetch_arcticdata.py
-Written by Tyler Sutterley (07/2025)
+Written by Tyler Sutterley (09/2025)
 Download Arctic Ocean Tide Models from the NSF ArcticData archive
 
 AODTM-5: https://arcticdata.io/catalog/view/doi:10.18739/A2901ZG3N
@@ -33,6 +33,8 @@ PROGRAM DEPENDENCIES:
     utilities.py: download and management utilities for syncing files
 
 UPDATE HISTORY:
+    Updated 09/2025: renamed module and function to fetch_arcticdata
+        made a callable function and added function docstrings
     Updated 07/2025: add a default directory for tide models
     Updated 05/2023: added option to change connection timeout
     Updated 04/2023: using pathlib to define and expand paths
