### 2025-09-29 db30acf
docs: add admonitions noting models will need to be downloaded
docs: admonition in xarray recipe
