+++ pyTMD/datasets/reduce_otis.py
@@ -135,32 +135,32 @@ def reduce_otis(MODEL,
     crs1 = pyTMD.crs().from_input(PROJECTION)
     crs2 = pyproj.CRS.from_epsg(4326)
     transformer = pyproj.Transformer.from_crs(crs1, crs2, always_xy=True)
-    xbox = np.array([BOUNDS[0],BOUNDS[1],BOUNDS[1],BOUNDS[0],BOUNDS[0]])
-    ybox = np.array([BOUNDS[2],BOUNDS[2],BOUNDS[3],BOUNDS[3],BOUNDS[2]])
-    lon,lat = transformer.transform(xbox,ybox)
+    xbox = np.array([BOUNDS[0], BOUNDS[1], BOUNDS[1], BOUNDS[0], BOUNDS[0]])
+    ybox = np.array([BOUNDS[2], BOUNDS[2], BOUNDS[3], BOUNDS[3], BOUNDS[2]])
+    lon, lat = transformer.transform(xbox, ybox)
 
     # convert bounds from latitude/longitude to model coordinates
-    x,y = pyTMD.crs().convert(lon,lat,model.projection,'F')
+    x, y = pyTMD.crs().convert(lon, lat, model.projection, 'F')
 
     # find indices to reduce to xmin,xmax,ymin,ymax
-    gridx,gridy = np.meshgrid(xi,yi)
-    indy,indx = np.nonzero((gridx >= x.min()) & (gridx <= x.max()) &
+    gridx, gridy = np.meshgrid(xi, yi)
+    indy, indx = np.nonzero((gridx >= x.min()) & (gridx <= x.max()) &
         (gridy >= y.min()) & (gridy <= y.max()))
     nx = np.count_nonzero((xi >= x.min()) & (xi <= x.max()))
     ny = np.count_nonzero((yi >= y.min()) & (yi <= y.max()))
     # calculate new grid limits and convert back to grid-cell edges
-    dx = np.abs(xi[1]-xi[0])
-    dy = np.abs(yi[1]-yi[0])
-    xlim = np.array([xi[indx[0]]-dx/2.0,xi[indx[-1]]+dx/2.0],dtype='>f4')
-    ylim = np.array([yi[indy[0]]-dy/2.0,yi[indy[-1]]+dy/2.0],dtype='>f4')
+    dx = np.abs(xi[1] - xi[0])
+    dy = np.abs(yi[1] - yi[0])
+    xlim = np.array([xi[indx[0]] - dx/2.0, xi[indx[-1]] + dx/2.0], dtype='>f4')
+    ylim = np.array([yi[indy[0]] - dy/2.0, yi[indy[-1]] + dy/2.0], dtype='>f4')
     # reduce grid and mask to new bounds
-    hz1 = np.zeros((ny,nx),dtype='>f4')
-    mz1 = np.zeros((ny,nx),dtype='>i4')
+    hz1 = np.zeros((ny,nx), dtype='>f4')
+    mz1 = np.zeros((ny,nx), dtype='>i4')
     hz1[:,:] = hz[indy,indx].reshape(ny,nx)
     mz1[:,:] = mz[indy,indx].reshape(ny,nx)
     # output reduced grid to file
     new_grid_file = create_unique_filename(model.grid_file)
-    pyTMD.io.OTIS.output_otis_grid(new_grid_file,xlim,ylim,hz1,mz1,iob,dt)
+    pyTMD.io.OTIS.output_otis_grid(new_grid_file, xlim, ylim, hz1, mz1, iob, dt)
     # change the permissions level to MODE
     new_grid_file.chmod(mode=MODE)
 
@@ -178,10 +178,10 @@ def reduce_otis(MODEL,
         for i,c in enumerate(constituents):
             # read constituent from elevation file
[…]
