### 2025-10-01 68f7586
test: add developer download routines for #471
fix: make pixi tasks associated with features for #472
docs: update data access for #470
docs: add notes about AVISO+ registration
