### 2025-10-01 ce82c15
docs: fix ref to model directories
idk it is weird
