+++ doc/source/user_guide/Examples.rst
@@ -7,7 +7,7 @@ Examples
 .. important::
   Running the example notebooks requires a `full installation <../getting_started/Install.html>`_ to include all of the optional dependencies.
 
-  Some notebooks require first downloading a tide model [see details on `getting started <../getting_started/Getting-Started.html#tide-model-formats>`_].
+  Some notebooks require first downloading a tide model [see details on `data access <../getting_started/Getting-Started.html#data-access>`_].
 
 .. toctree::
    :maxdepth: 1
