### 2025-10-02 5c545ae
fix: move xarray to `all` dependencies
