### 2025-10-02 b9a0fb3
docs: fix pytest parallel for #473
