### 2025-10-02 ae0be7c
ci: refactor to use pixi
ci: fetch data from open repository
ci: no longer reliant on AWS s3
refactor: remove environment.yml and add pixi task to export
chore: clean up pyproject.toml
