### 2025-10-03 ac87362
docs: change build from micromamba to pixi
docs: add note for fetching test data for #471
