+++ doc/environment.yml
@@ -1,34 +1,0 @@
-name: pytmd-docs
-channels:
-  - conda-forge
-dependencies:
-  - cartopy
-  - docutils
-  - fontconfig
-  - freetype
-  - graphviz
-  - ipyleaflet
-  - ipympl
-  - ipywidgets
-  - jplephem
-  - lxml
-  - matplotlib
-  - myst-nb
-  - netCDF4
-  - notebook
-  - numpy
-  - numpydoc
-  - pandas
-  - pip
-  - python>=3.6
-  - python-dateutil
-  - scipy>=1.16.0
-  - sphinx
-  - sphinx-argparse>=0.4
-  - sphinxcontrib-bibtex
-  - sphinx-design
-  - sphinx_rtd_theme
-  - timescale>=0.0.8
-  - xarray
-  - pip:
-    - ..
