### 2025-10-03 2326294
ci: try adding macos tahoe `macos-26`
