+++ .github/workflows/python-request.yml
@@ -20,7 +20,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, windows-latest]
+        os: [ubuntu-latest, macos-26, windows-latest]
     env:
       OS: ${{ matrix.os }}
     defaults:
