### 2025-10-03 ba17f24
ci: drop macos-latest build
:(
