### 2025-10-03 2aabfdb
docs: add a draft of a JOSS paper (#415)
* ci: add JOSS paper build
* docs: add non-antarctic pyTMD uses (#416)
* docs: add cite to ESA cci+ velocities
* docs: add doi for Kantha and Clayson
* mention PERTH is a fortran software
* typo (#466)
* docs: add blurbs about the harmonic method
* docs: add tide potential catalog table
* ci: add environmental variable for paper path
* docs: add spherical harmonic plots
* docs: add table of max tide potential
* chore: add more `ruff` lint parameters
* refactor: functional access to downloading scripts for #470
* docs: expand on the directory structure
*
[…]
