+++ pyTMD/datasets/fetch_gsfc_got.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 u"""
-gsfc_got_tides.py
-Written by Tyler Sutterley (0/2025)
+fetch_gsfc_got.py
+Written by Tyler Sutterley (09/2025)
 Download Goddard Ocean Tide (GOT) models
 
 CALLING SEQUENCE:
-    python gsfc_got_tides.py --tide=GOT5.6
+    python fetch_gsfc_got.py --tide=GOT5.6
 
 COMMAND LINE OPTIONS:
     --help: list the command line options
@@ -32,6 +32,8 @@ PROGRAM DEPENDENCIES:
     utilities.py: download and management utilities for syncing files
 
 UPDATE HISTORY:
+    Updated 09/2025: renamed module and function to fetch_gsfc_got
+        made a callable function and added function docstrings
     Updated 07/2025: add a default directory for tide models
     Updated 01/2025: scrubbed use of pathlib.os to just use os directly
     Updated 09/2024: added Ray and Erofeeva (2014) long-period tide model
@@ -51,18 +53,41 @@ import argparse
 import posixpath
 import pyTMD.utilities
 
+# default data directory for tide models
+_default_directory = pyTMD.utilities.get_data_path('data')
+
 # PURPOSE: Download Arctic Ocean Tide Models from the NSF ArcticData archive
-def gsfc_got_tides(MODEL: str,
-    DIRECTORY: str | pathlib.Path | None = None,
-    FORMAT: str = 'netcdf',
-    GZIP: bool = False,
-    TIMEOUT: int | None = None,
-    MODE: oct = 0o775):
+def fetch_gsfc_got(model: str,
+        directory: str | pathlib.Path | None = _default_directory,
+        format: str = 'netcdf',
+        compressed: bool = False,
+        timeout: int | None = None,
+        mode: oct = 0o775
+    ):
+    """
+    Download Goddard Ocean Tide (GOT) models from NASA Goddard Space Flight
+    Center (GSFC)
+
+    Parameters
+    ----------
+    model: str
+        GOT tide model to download
+    directory: str or pathlib.Path
+        Working data directory
+    format: str, default 'netcdf'
+        GOT tide model format to download
+    compressed: bool, default False
+        Compress output ascii and netCDF4 tide files
+    timeout: int, default None
+        Timeout in seconds for blocking operations
+    mode: oct, default 0o775
+        Local permissions mode of the files downloaded
+    """
 
     # create logger for verbosity level
-    logger = pyTMD.utilities.build_logger(__name__,level=logging.INFO)
+    logger = pyTMD.utilities.build_logger(__name__, level=logging.INFO)
     # if compressing the output files
-    opener = gzip.open if GZIP else open
+    opener = gzip.open if compressed else open
 
     # url for each tide model tarfile
     PATH = {}
@@ -82,33 +107,37 @@ def gsfc_got_tides(MODEL: str,
     TAR['RE14'] = 'r'
 
     # recursively create directories if non-existent
-    DIRECTORY = pathlib.Path(DIRECTORY).expanduser().absolute()
-    DIRECTORY.mkdir(MODE, parents=True, exist_ok=True)
+    directory = pathlib.Path(directory).expanduser().absolute()
+    directory.mkdir(mode=mode, parents=True, exist_ok=True)
 
     # build host url for model
-    URL = ['https://earth.gsfc.nasa.gov','sites','default','files',*PATH[MODEL]]
+    url = ['https://earth.gsfc.nasa.gov','sites','default','files',*PATH[model]]
     # download tarfile from host
-    logger.info(f'{posixpath.join(*URL)} -->\n')
[…]
