### 2025-10-03 #8566 @tsutterley
Hi all,
Issues are closed and PR from the paper branch is merged into main! Thanks for all the comments!! :)

The two unrendered notebooks are expected. The "compact" notebook uses a pretty large dataset (~2GB) that readthedocs might not like. Similar with the tidal currents, but I might find a good solution in the future.
