### 2025-10-06 #8566 @rcaneill
Hi all,
I am now done with the review. As a non native English speaker, I cannot be sure that the English is 100% correct, as far as I have seen it is all good.
Hi hope that @tsutterley you’re happy with the new version of the paper, the doc, the tests, etc. Personally, I think the overall quality has improved! :)
