### 2025-10-06 a005e43
docs: add acknowledgment to Romain and Gui (#474)
