### 2025-10-07 175c84d
test: fix pixi task for testing (#475)
