### 2025-10-08 556da5b
feat: can solve with inferred minor constituents (post-fit)
