### 2025-10-08 a2f43a6
fix: docstring of solve.constants
