### 2025-10-08 90f4be4
feat: can solve with inferred minor constituents (post-fit) (#476)
