+++ pyTMD/solve/constants.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 constants.py
-Written by Tyler Sutterley (08/2025)
+Written by Tyler Sutterley (10/2025)
 Routines for estimating the harmonic constants for ocean tides
 
 REFERENCES:
@@ -25,6 +25,7 @@ PROGRAM DEPENDENCIES:
     astro.py: computes the basic astronomical mean longitudes
 
 UPDATE HISTORY:
+    Updated 10/2025: added option to infer minor constituents (post-fit)
     Updated 08/2025: use numpy degree to radian conversions
     Updated 06/2025: verify that height values are all finite
     Updated 05/2025: added option to include higher order polynomials
@@ -40,6 +41,7 @@ import numpy as np
 import scipy.linalg
 import scipy.optimize
 import pyTMD.arguments
+import pyTMD.predict
 
 __all__ = [
     'constants'
@@ -55,11 +57,14 @@ def constants(t: float | np.ndarray,
         corrections: str = 'OTIS',
         solver: str = 'lstsq',
         order: int = 0,
+        infer_minor: bool = False,
+        minor_constituents: list = [],
         bounds: tuple = (-np.inf, np.inf),
-        max_iter: int | None = None
+        max_iter: int | None = None,
+        infer_iter: int = 1
     ):
     """
-    Estimate the harmonic constants for an elevation time series
+    Estimate the harmonic constants for a time series
     :cite:p:`Egbert:2002ge,Foreman:2009bg,Ray:1999vm`
 
     Parameters
@@ -67,7 +72,7 @@ def constants(t: float | np.ndarray,
     t: float or np.ndarray
         days relative to 1992-01-01T00:00:00
     ht: np.ndarray
-        elevation time series (meters)
+        input time series (elevation or currents)
     constituents: str, list or np.ndarray
         tidal constituent ID(s)
     deltat: float or np.ndarray, default 0.0
@@ -84,15 +89,21 @@ def constants(t: float | np.ndarray,
         - ``'bvls'``: bounded-variable least-squares
     order: int, default 0
         degree of the polynomial to add to fit
+    infer_minor: bool, default False
+        infer minor tidal constituents
+    minor_constituents: list or None, default None
+        Specify constituents to infer
     bounds: tuple, default (None, None)
         Lower and upper bounds on parameters for ``'bvls'``
     max_iter: int or None, default None
         Maximum number of iterations for ``'bvls'``
+    infer_iter: int, default 1
+        Maximum number of iterations for inferring minor constituents
 
     Returns
     -------
     amp: np.ndarray
-        amplitude of each harmonic constant (meters)
+        amplitude of each harmonic constant (units of input data)
     phase: np.ndarray
         phase of each harmonic constant (degrees)
     """
@@ -152,6 +163,33 @@ def constants(t: float | np.ndarray,
         p = scipy.optimize.lsq_linear(M, ht,
             method=solver, bounds=bounds,
             max_iter=max_iter).x
+
+    # infer minor using a post-fit technique
+    infer_iter = np.maximum(1, infer_iter) if infer_minor else 0  
+    # iterate solutions to improve minor constituent estimates
+    for i in range(infer_iter):
+        # indices for the sine and cosine terms
+        # skip over the polynomial terms
+        isin = 2*np.arange(nc) + order + 2
+        icos = 2*np.arange(nc) + order + 1
+        # complex model amplitudes for major constituents
+        hmajor = p[icos] + 1j*p[isin]
+        # inferred minor constituent time series
+        hminor = pyTMD.predict.infer_minor(t, hmajor, constituents,
+            deltat=deltat, corrections=corrections,
[…]
