### 2025-10-10 528af17
docs: add `:func:` and `:py:class:` pointers
refactor: remove some deprecated functions
