### 2025-10-13 2df865d
docs: add `:func:` and `:py:class:` pointers  (#477)
refactor: remove some deprecated functions
