+++ pyTMD/compute_tide_corrections.py
@@ -1,174 +1,0 @@
-#!/usr/bin/env python
-u"""
-compute_tide_corrections.py
-Written by Tyler Sutterley (01/2024)
-Calculates tidal elevations for correcting elevation or imagery data
-
-Ocean and Load Tides
-Uses OTIS format tidal solutions provided by Oregon State University and ESR
-    http://volkov.oce.orst.edu/tides/region.html
-    https://www.esr.org/research/polar-tide-models/list-of-polar-tide-models/
-    ftp://ftp.esr.org/pub/datasets/tmd/
-Global Tide Model (GOT) solutions provided by Richard Ray at GSFC
-or Finite Element Solution (FES) models provided by AVISO
-
-Long-Period Equilibrium Tides (LPET)
-Calculates long-period equilibrium tidal elevations for correcting
-elevation or imagery data from the summation of fifteen spectral lines
-    https://doi.org/10.1111/j.1365-246X.1973.tb03420.x
-
-Load Pole Tides (LPT)
-Calculates radial load pole tide displacements following IERS Convention
-(2010) guidelines for correcting elevation or imagery data
-    https://iers-conventions.obspm.fr/chapter7.php
-
-Ocean Pole Tides (OPT)
-Calculates radial ocean pole load tide displacements following IERS Convention
-(2010) guidelines for correcting elevation or imagery data
-    https://iers-conventions.obspm.fr/chapter7.php
-
-Ocean Pole Tides (SET)
-Calculates radial Solid Earth tide displacements following IERS Convention
-(2010) guidelines for correcting elevation or imagery data
-    https://iers-conventions.obspm.fr/chapter7.php
-
-PYTHON DEPENDENCIES:
-    numpy: Scientific Computing Tools For Python
-        https://numpy.org
-        https://numpy.org/doc/stable/user/numpy-for-matlab-users.html
-    scipy: Scientific Tools for Python
-        https://docs.scipy.org/doc/
-    netCDF4: Python interface to the netCDF C library
-        https://unidata.github.io/netcdf4-python/netCDF4/index.html
-    pyproj: Python interface to PROJ library
-        https://pypi.org/project/pyproj/
-
-PROGRAM DEPENDENCIES:
-    time.py: utilities for calculating time operations
-    spatial: utilities for reading, writing and operating on spatial data
-    utilities.py: download and management utilities for syncing files
-    arguments.py: load the nodal corrections for tidal constituents
-    astro.py: computes the basic astronomical mean longitudes
-    crs.py: Coordinate Reference System (CRS) routines
-    predict.py: predict tide values using harmonic constants
-    io/model.py: retrieves tide model parameters for named tide models
-    io/OTIS.py: extract tidal harmonic constants from OTIS tide models
-    io/ATLAS.py: extract tidal harmonic constants from netcdf models
-    io/GOT.py: extract tidal harmonic constants from GSFC GOT models
-    io/FES.py: extract tidal harmonic constants from FES tide models
-    interpolate.py: interpolation routines for spatial data
-
-UPDATE HISTORY:
-    Updated 01/2024: made the inference of minor constituents an option
-        refactored lunisolar ephemerides functions
-        deprecated in favor of refactored compute.py module
-    Updated 12/2023: use new crs class for coordinate reprojection
-    Updated 08/2023: changed ESR netCDF4 format to TMD3 format
-    Updated 05/2023: use timescale class for time conversion operations
-        use defaults from eop module for pole tide and EOP files
-        add option for using higher resolution ephemerides from JPL
-    Updated 04/2023: added function for radial solid earth tides
-        using pathlib to define and expand paths
-    Updated 03/2023: add basic variable typing to function inputs
[…]
