### 2025-10-14 e71fe14
feat: added option to use memory mapping for reading large files
