+++ pyTMD/io/OTIS.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 OTIS.py
-Written by Tyler Sutterley (08/2025)
+Written by Tyler Sutterley (10/2025)
 
 Reads files for a tidal model and makes initial calculations to run tide program
 Includes functions to extract tidal harmonic constants from OTIS tide models for
@@ -60,6 +60,8 @@ PROGRAM DEPENDENCIES:
     interpolate.py: interpolation routines for spatial data
 
 UPDATE HISTORY:
+    Updated 10/2025: refactored binary read programs
+        added option to use memory mapping for reading large files
     Updated 08/2025: use numpy degree to radian conversions
         added option to gap fill when reading constituent grids
     Updated 05/2025: added option to select constituents to read from model
@@ -215,6 +217,8 @@ def extract_constants(
             - ``'ATLAS'``: reading a global solution with localized solutions
             - ``'OTIS'``: combined global or local solution
             - ``'TMD3'``: combined global or local netCDF4 solution
+    use_mmap: bool, default False
+        Use memory mapping to read binary OTIS or ATLAS files
     constituents: list or None, default None
         Specify constituents to read from model
     crop: bool, default False
@@ -252,6 +256,7 @@ def extract_constants(
     # set default keyword arguments
     kwargs.setdefault('type', 'z')
     kwargs.setdefault('grid', 'OTIS')
+    kwargs.setdefault('use_mmap', False)
     kwargs.setdefault('constituents', None)
     kwargs.setdefault('crop', False)
     kwargs.setdefault('bounds', None)
@@ -278,15 +283,19 @@ def extract_constants(
     # read the OTIS-format tide grid file
     if (kwargs['grid'] == 'ATLAS'):
         # if reading a global solution with localized solutions
-        x0,y0,hz0,mz0,iob,dt,pmask,local = read_atlas_grid(grid_file)
-        xi,yi,hz = combine_atlas_model(x0,y0,hz0,pmask,local,variable='depth')
-        mz = create_atlas_mask(x0,y0,mz0,local,variable='depth')
+        x0,y0,hz0,mz0,iob,dt,pmask,local = read_atlas_grid(
+            grid_file, use_mmap=kwargs['use_mmap'])
+        xi,yi,hz = combine_atlas_model(x0,y0,hz0,pmask,local,
+            variable='depth')
+        mz = create_atlas_mask(x0,y0,mz0,local,
+            variable='depth')
     elif (kwargs['grid'] == 'TMD3'):
         # if reading a single TMD3 netCDF4 solution
         xi,yi,hz,mz,sf = read_netcdf_grid(grid_file)
     else:
         # if reading a single OTIS solution
-        xi,yi,hz,mz,iob,dt = read_otis_grid(grid_file)
+        xi,yi,hz,mz,iob,dt = read_otis_grid(grid_file,
+            use_mmap=kwargs['use_mmap'])
     # invert tide mask to be True for invalid points
     mz = np.logical_not(mz).astype(mz.dtype)
 
@@ -407,7 +416,8 @@ def extract_constants(
         cons = [read_constituents(m)[0].pop() for m in model_file]
         nc = len(cons)
     else:
-        cons,nc = read_constituents(model_file, grid=kwargs['grid'])
+        cons,nc = read_constituents(model_file,
+            grid=kwargs['grid'], use_mmap=kwargs['use_mmap'])
     # reduce number of constituents
     if kwargs['constituents'] is not None:
         # verify that constituents is a list
@@ -436,7 +446,8 @@ def extract_constants(
         if (kwargs['type'] == 'z'):
             # read z constituent from elevation file
             if (kwargs['grid'] == 'ATLAS'):
[…]
