### 2025-10-14 58feb48
feat: added option to use memory mapping for reading large files (#478)
