### 2025-10-14 06e270d
fix: assign order to reshape if not using mmap
