### 2025-10-16 42f2ffb
feat:  create fetch class to group fetching functions
