### 2025-10-24 d882dc6
fix: add `platformdirs` to dependencies
