+++ README.md
@@ -72,6 +72,7 @@ This will automatically create the environment, install all dependencies, and la
 - [lxml: processing XML and HTML in Python](https://pypi.python.org/pypi/lxml)
 - [netCDF4: Python interface to the netCDF C library](https://unidata.github.io/netcdf4-python/)
 - [numpy: Scientific Computing Tools For Python](https://www.numpy.org)
+- [platformdirs: Python module for determining platform-specific directories](https://pypi.org/project/platformdirs/)
 - [pyproj: Python interface to PROJ library](https://pypi.org/project/pyproj/)
 - [scipy: Scientific Tools for Python](https://www.scipy.org/)
 - [timescale: Python tools for time and astronomical calculations](https://pypi.org/project/timescale/)
