### 2025-10-24 933a636
refactor: change default directory for tide models to cache (#481)
* feat: add `platformdirs` to dependencies
