+++ test/conftest.py
@@ -1,9 +1,9 @@
 import pytest
 import pathlib
-from pyTMD.utilities import get_data_path
+from pyTMD.utilities import get_cache_path
 
 # default working data directory for tide models
-_default_directory = get_data_path('data')
+_default_directory = get_cache_path()
 
 def pytest_addoption(parser):
     parser.addoption("--directory", action="store", help="Directory for test data", default=_default_directory, type=pathlib.Path)
