+++ doc/source/api_reference/utilities.rst
@@ -14,6 +14,8 @@ General Methods
 
 .. autofunction:: pyTMD.utilities.get_data_path
 
+.. autofunction:: pyTMD.utilities.get_cache_path
+
 .. autofunction:: pyTMD.utilities.import_dependency
 
 .. autoclass:: pyTMD.utilities.reify
