+++ doc/source/project/Citations.rst
@@ -25,6 +25,7 @@ This software is also dependent on other commonly used Python packages:
 - `lxml: processing XML and HTML in Python <https://pypi.python.org/pypi/lxml>`_
 - `netCDF4: Python interface to the netCDF C library <https://unidata.github.io/netcdf4-python/>`_
 - `numpy: Scientific Computing Tools For Python <https://www.numpy.org>`_
+- `platformdirs: Python module for determining platform-specific directories <https://pypi.org/project/platformdirs/>`_
 - `pyproj: Python interface to PROJ library <https://pypi.org/project/pyproj/>`_
 - `scipy: Scientific Tools for Python <https://www.scipy.org/>`_
 - `timescale: Python tools for time and astronomical calculations <https://pypi.org/project/timescale/>`_
