### 2025-10-27 908c395
fix: simplify `pixi` build with `matplotlib-base`
