### 2025-10-27 9273ae8
docs: elaborate on Julian centuries
