### 2025-10-29 4a3ce47
docs: add Taher to contributors
