+++ doc/source/project/Contributors.rst
@@ -5,6 +5,7 @@ Contributors
 .. include:: ../../../CONTRIBUTORS.md
     :parser: commonmark
 
-We also wish to acknowledge the contributions and comments from our two JOSS reviewers Romain Caneill and Guilherme Castel\ |atilde|\ o.
+We also wish to acknowledge the contributions and comments from our JOSS editor Taher Chegini,
+and reviewers Romain Caneill and Guilherme Castel\ |atilde|\ o.
 
 .. |atilde|    unicode:: U+00E3 .. LATIN SMALL LETTER A WITH TILDE
