### 2025-10-29 70f25ef
fix: skip 4 bytes if no boundary conditions
