+++ pyTMD/io/OTIS.py
@@ -1142,6 +1142,7 @@ def read_otis_grid(
     # read iob from file
     if (nob == 0):
         iob = []
+        offset += 4
     else:
         iob = read_raw_binary(input_file,
             dtype='>i4',
