### 2025-10-31 582defb
Update GOT.py
