### 2025-11-02 6db967b
fix: simplify `pixi` build with `matplotlib-base` (#482)
* docs: elaborate on Julian centuries
* fix: skip 4 bytes if no boundary conditions
