### 2025-11-03 3c7e5f4
fix: include filtering mask out from possible parsed names
