### 2025-11-03 6ba526b
Update model.py
