+++ pyTMD/io/model.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 model.py
-Written by Tyler Sutterley (08/2025)
+Written by Tyler Sutterley (11/2025)
 Retrieves tide model parameters for named tide models and
     from model definition files
 
@@ -11,6 +11,7 @@ PYTHON DEPENDENCIES:
         https://numpy.org/doc/stable/user/numpy-for-matlab-users.html
 
 UPDATE HISTORY:
+    Updated 11/2025: use default cache directory if directory is None
     Updated 08/2025: use numpy degree to radian conversions
         update node equilibrium tide estimation
         add functions for converting a model to an xarray DataTree object
@@ -77,7 +78,7 @@ import copy
 import json
 import pathlib
 import numpy as np
-from pyTMD.utilities import import_dependency, get_data_path
+from pyTMD.utilities import import_dependency, get_data_path, get_cache_path
 from collections.abc import Iterable
 from dataclasses import dataclass
 
@@ -89,6 +90,9 @@ __all__ = [
     'load_database',
     'model'
 ]
+
+# default working data directory for tide models
+_default_directory = get_cache_path()
 
 @dataclass
 class DataBase:
@@ -144,7 +148,7 @@ def load_database(extra_databases: list = []):
         # otherwise load parameters from JSON file path
         else:
             # verify that extra database file exists
-            db = pathlib.Path(db).expanduser().absolute()
+            db = pathlib.Path(db).expanduser()
             if not db.exists():
                 raise FileNotFoundError(db)
             # extract JSON data
@@ -268,7 +272,7 @@ class model:
         """
         # set working data directory if unset
         if self.directory is None:
-            self.directory = pathlib.Path().absolute()
+            self.directory = pathlib.Path(_default_directory)
         # select between known tide models
         parameters = load_database(extra_databases=self.extra_databases)
         # try to extract parameters for model
@@ -299,7 +303,7 @@ class model:
         """
         # set working data directory if unset
         if self.directory is None:
-            self.directory = pathlib.Path().absolute()
+            self.directory = pathlib.Path(_default_directory)
         # select between tide models
         parameters = load_database(extra_databases=self.extra_databases)
         # try to extract parameters for model
@@ -639,7 +643,7 @@ class model:
         """
         # set working data directory if unset
         if self.directory is None:
-            self.directory = pathlib.Path().absolute()
+            self.directory = pathlib.Path(_default_directory)
         # complete model file paths
         if isinstance(model_file, list):
             output_file = [self.pathfinder(f) for f in model_file]
