### 2025-11-04 4e5ef7c
feat: added crs property for model coordinate reference system
