### 2025-11-04 c35563b
feat: made some additions to the `xarray` accessor
feat: added `is_global` property for models covering a global domain
feat: added `pad` function to pad global datasets along boundaries
feat: added `inpaint` function to fill missing data in datasets
