+++ pyTMD/io/dataset.py
@@ -5,30 +5,39 @@ Written by Tyler Sutterley (11/2025)
 An xarray.Dataset extension for tidal model data
 
 PYTHON DEPENDENCIES:
+    numpy: Scientific Computing Tools For Python
+        https://numpy.org
+        https://numpy.org/doc/stable/user/numpy-for-matlab-users.html
     pyproj: Python interface to PROJ library
         https://pypi.org/project/pyproj/
         https://pyproj4.github.io/pyproj/
+    scipy: Scientific Tools for Python
+        https://docs.scipy.org/doc/
     xarray: N-D labeled arrays and datasets in Python
         https://docs.xarray.dev/en/stable/
 
 UPDATE HISTORY:
     Updated 11/2025: get crs directly using pyTMD.CRS.from_user_input
         set variable name to constituent for to_dataarray method
+        added is_global property for models covering a global domain
+        added pad function to pad global datasets along boundaries
+        added inpaint function to fill missing data in datasets
     Updated 09/2025: added argument to limit the list of constituents
         when converting to an xarray DataArray
     Written 08/2025
 """
+import numpy as np
 from pyTMD.utilities import import_dependency
 # attempt imports
 xr = import_dependency('xarray')
 pyproj = import_dependency('pyproj')
 
 __all__ = [
-    'dataset',
+    'Dataset',
 ]
 
 @xr.register_dataset_accessor('tmd')
-class dataset:
+class Dataset:
     """Accessor for extending an ``xarray.Dataset`` for tidal model data
     """
     def __init__(self, ds):
@@ -44,6 +53,97 @@ class dataset:
         da = self._ds[kwargs['constituents']].to_dataarray(dim='constituent').T
         da = da.assign_coords(constituent=kwargs['constituents'])
         return da
+    
+    def inpaint(self, **kwargs):
+        """
+        Inpaint over missing data in ``Dataset``
+
+        Parameters
+        ----------
+        kwargs: keyword arguments
+            keyword arguments for ``pyTMD.interpolate.inpaint``
+
+        Returns
+        -------
+        ds: xarray.Dataset
+            interpolated xarray Dataset
+        """
+        # import inpaint function
+        from pyTMD.interpolate import inpaint
+        # create copy of dataset
+        ds = self._ds.copy()
+        # inpaint each variable in the dataset
+        for v in ds.data_vars.keys():
+            ds[v].values = inpaint(self._ds.x, self._ds.y,
+                self._ds[v].values, **kwargs)
+        # return the dataset
+        return ds
+
+    def interp(self, 
+            x: np.ndarray,
+            y: np.ndarray, 
+            method='linear',              
+            **kwargs
+        ):
+        """
+        Interpolate ``Dataset`` to input coordinates
+        
+        Parameters
+        ----------
+        x: np.ndarray
+            input x-coordinates
+        y: np.ndarray
+            input y-coordinates
+        method: str, default 'linear'
+            Interpolation method
+        **kwargs: dict
+            Additional keyword arguments for reading the dataset
+            
+        Returns
+        -------
+        ds: xarray.Dataset
+            interpolated tidal constants
+        """
+        # pad global grids along x-dimension (if necessary)
+        if self.is_global:
+            self._ds = self.pad(n=1)
+        # check longitudinal convention for geographic models
+        if self.crs.is_geographic:
+            # grid spacing in x-direction
+            dx = self._ds.x[1] - self._ds.x[0]
+            # adjust input longitudes to be consistent with model
+            if (np.min(x) < 0.0) & (self._ds.x.max() > (180.0 + dx)):
+                # input points convention (-180:180)
[…]
