+++ pyTMD/io/__init__.py
@@ -18,6 +18,6 @@ from .model import (
 
 # Attempt to import the dataset module for xarray support
 try:
-    from .dataset import dataset
+    from .dataset import Dataset
 except (ImportError, AttributeError):
-    dataset = None
+    Dataset = None
