### 2025-11-05 c3f8221
feat: add projection info for all models
