### 2025-11-06 70d1bf5
feat: add dataarray accessor for amp+phase
