### 2025-11-06 53cf115
feat: added crs property for model coordinate reference system (#483)
- fix: include filtering mask out from possible parsed names
- feat: added `crs` property for model coordinate reference system
- feat: added `is_global` property for models covering a global domain
- feat: added `pad` function to pad global datasets along boundaries
- feat: added `inpaint` function to fill missing data in datasets
- feat: add projection info for all models to `database.json`
- feat: add dataarray accessor for amp+phase
