+++ pyproject.toml
@@ -29,7 +29,7 @@ dependencies = [
     "netCDF4",
     "numpy",
     "platformdirs",
-    "pyproj",
+    "pyproj>=2.5.0",
     "python-dateutil",
     "scipy>=1.10.1",
     "timescale>=0.0.8",
@@ -181,7 +181,7 @@ numpy = "*"
 netcdf4 = "*"
 lxml = "*"
 platformdirs = "*"
-pyproj = "*"
+pyproj = ">=2.5.0"
 python-dateutil = "*"
 scipy = ">=1.16.0"
 timescale = ">=0.0.8"

[…]
