### 2025-11-06 106c730
feat: experimental xarray io
