+++ pyTMD/xio/ATLAS.py
@@ -1,0 +1,257 @@
+#!/usr/bin/env python
+u"""
+ATLAS.py
+Written by Tyler Sutterley (11/2025)
+
+Reads netCDF4 ATLAS tidal solutions provided by Oregon State University
+
+PYTHON DEPENDENCIES:
+    xarray: N-D labeled arrays and datasets in Python
+        https://docs.xarray.dev/en/stable/
+
+UPDATE HISTORY:
+    Updated 11/2025: near-complete rewrite of program to use xarray
+    Updated 08/2025: use numpy degree to radian conversions
+        added option to gap fill when reading constituent grids
+    Updated 11/2024: expose buffer distance for cropping tide model data
+    Updated 10/2024: fix error when using default bounds in extract_constants
+    Updated 07/2024: added crop and bounds keywords for trimming model data
+    Updated 02/2024: changed variable for setting global grid flag to is_global
+    Updated 10/2023: add generic wrapper function for reading constituents
+    Updated 04/2023: using pathlib to define and expand tide model paths
+    Updated 03/2023: add basic variable typing to function inputs
+    Updated 12/2022: refactor tide read programs under io
+        new functions to read and interpolate from constituents class
+        new functions to output ATLAS formatted netCDF4 files
+        refactored interpolation routines into new module
+    Updated 11/2022: place some imports within try/except statements
+        use f-strings for formatting verbose or ascii output
+    Updated 07/2022: fix setting of masked array data to NaN
+    Updated 05/2022: reformat arguments to extract_netcdf_constants definition
+        changed keyword arguments to camel case
+    Updated 04/2022: updated docstrings to numpy documentation format
+    Updated 12/2021: adjust longitude convention based on model longitude
+    Updated 09/2021: fix cases where there is no mask on constituent files
+    Updated 07/2021: added check that tide model files are accessible
+    Updated 06/2021: add warning for tide models being entered as string
+    Updated 05/2021: added option for extrapolation cutoff in kilometers
+    Updated 03/2021: add extrapolation check where there are no invalid points
+        prevent ComplexWarning for fill values when calculating amplitudes
+        simplified inputs to be similar to binary OTIS read program
+    Updated 02/2021: set invalid values to nan in extrapolation
+        replaced numpy bool to prevent deprecation warning
+    Updated 12/2020: added valid data extrapolation with nearest_extrap
+        replace tostring with tobytes to fix DeprecationWarning
+    Updated 11/2020: create function to read bathymetry and spatial coordinates
+    Updated 09/2020: set bounds error to false for regular grid interpolations
+        adjust dimensions of input coordinates to be iterable
+        reduce number of interpolations by copying bathymetry mask to variables
+    Updated 08/2020: replaced griddata with scipy regular grid interpolators
+    Updated 07/2020: added function docstrings. separate bilinear interpolation
+        changed TYPE variable to keyword argument. update griddata interpolation
+    Updated 06/2020: use argmin and argmax in bilinear interpolation
+    Written 09/2019
+"""
+from __future__ import annotations
+
+import gzip
+import pathlib
+import xarray as xr
+
+__all__ = [
+    'open_dataset',
+    'open_mfdataset',
+    'open_atlas_grid',
+    'open_atlas_dataset',
+    'ATLASDataTree',
+]
+
[…]
