+++ pyTMD/xio/OTIS.py
@@ -1,0 +1,2037 @@
+#!/usr/bin/env python
+u"""
+OTIS.py
+Written by Tyler Sutterley (11/2025)
+
+Reads OTIS format tidal solutions provided by Oregon State University and ESR
+    http://volkov.oce.orst.edu/tides/region.html
+    https://www.esr.org/research/polar-tide-models/list-of-polar-tide-models/
+    ftp://ftp.esr.org/pub/datasets/tmd/
+
+PYTHON DEPENDENCIES:
+    numpy: Scientific Computing Tools For Python
+        https://numpy.org
+        https://numpy.org/doc/stable/user/numpy-for-matlab-users.html
+    pyproj: Python interface to PROJ library
+        https://pypi.org/project/pyproj/
+        https://pyproj4.github.io/pyproj/
+    xarray: N-D labeled arrays and datasets in Python
+        https://docs.xarray.dev/en/stable/
+
+UPDATE HISTORY:
+    Updated 11/2025: near-complete rewrite of program to use xarray
+    Updated 10/2025: refactored binary read programs
+        added option to use memory mapping for reading large files
+    Updated 08/2025: use numpy degree to radian conversions
+        added option to gap fill when reading constituent grids
+    Updated 05/2025: added option to select constituents to read from model
+    Updated 12/2024: released version of TMD3 has different variable names
+    Updated 11/2024: expose buffer distance for cropping tide model data
+    Updated 10/2024: save latitude and longitude to output constituent object
+        fix error when using default bounds in extract_constants
+    Updated 09/2024: using new JSON dictionary format for model projections
+    Updated 08/2024: revert change and assume crop bounds are projected
+    Updated 07/2024: added crop and bounds keywords for trimming model data
+        convert the crs of bounds when cropping model data
+    Updated 06/2024: change int32 to int to prevent overflows with numpy 2.0
+    Updated 02/2024: don't overwrite hu and hv in _interpolate_zeta
+        changed variable for setting global grid flag to is_global
+    Updated 01/2024: construct currents masks differently if not global
+        renamed currents masks and bathymetry interpolation functions
+    Updated 12/2023: use new crs class for coordinate reprojection
+    Updated 10/2023: fix transport variable entry for TMD3 models
+    Updated 09/2023: prevent overwriting ATLAS compact x and y coordinates
+    Updated 08/2023: changed ESR netCDF4 format to TMD3 format
+    Updated 04/2023: using pathlib to define and expand tide model paths
+    Updated 03/2023: add basic variable typing to function inputs
+        new function name for converting coordinate reference systems
+    Updated 12/2022: refactor tide read programs under io
+        new functions to read and interpolate from constituents class
+        refactored interpolation routines into new module
+    Updated 11/2022: place some imports within try/except statements
+        fix variable reads for ATLAS compact data formats
+        use f-strings for formatting verbose or ascii output
+    Updated 10/2022: invert current tide masks to be True for invalid points
+    Updated 06/2022: unit updates in the ESR netCDF4 format
+    Updated 05/2022: add functions for using ESR netCDF4 format models
+        changed keyword arguments to camel case
+    Updated 04/2022: updated docstrings to numpy documentation format
+        use longcomplex data format to be windows compliant
+    Updated 03/2022: invert tide mask to be True for invalid points
[…]
