### 2025-11-06 123a8ae
feat: experimental xarray io
