+++ pyTMD/xio/FES.py
@@ -1,0 +1,401 @@
+#!/usr/bin/env python
+u"""
+FES.py
+Written by Tyler Sutterley (11/2025)
+
+Reads ascii and netCDF4 files for FES tidal solutions provided by AVISO
+    https://www.aviso.altimetry.fr/data/products/auxiliary-products/
+        global-tide-fes.html
+
+PYTHON DEPENDENCIES:
+    numpy: Scientific Computing Tools For Python
+        https://numpy.org
+        https://numpy.org/doc/stable/user/numpy-for-matlab-users.html
+    xarray: N-D labeled arrays and datasets in Python
+        https://docs.xarray.dev/en/stable/
+
+UPDATE HISTORY:
+    Updated 11/2025: near-complete rewrite of program to use xarray
+    Updated 10/2025: simplify ascii read function to use masked_equal
+    Updated 08/2025: use numpy degree to radian conversions
+        added option to gap fill when reading constituent grids
+    Updated 11/2024: expose buffer distance for cropping tide model data
+    Updated 10/2024: fix error when using default bounds in extract_constants
+    Updated 07/2024: added new FES2022 to available known model versions
+        FES2022 have masked longitudes, only extract longitude data
+        FES2022 extrapolated data have zeroed out inland water bodies
+        added crop and bounds keywords for trimming model data
+    Updated 01/2024: attempt to extract constituent IDs from filenames
+    Updated 06/2023: extract ocean tide model variables for FES2012
+    Updated 04/2023: added global HAMTIDE11 model
+        using pathlib to define and expand tide model paths
+    Updated 03/2023: add basic variable typing to function inputs
+    Updated 12/2022: refactor tide read programs under io
+        new functions to read and interpolate from constituents class
+        new functions to output FES formatted netCDF4 files
+        refactored interpolation routines into new module
+    Updated 11/2022: place some imports within try/except statements
+        use f-strings for formatting verbose or ascii output
+    Updated 05/2022: reformat arguments to extract_FES_constants definition
+        changed keyword arguments to camel case
+    Updated 04/2022: updated docstrings to numpy documentation format
+        include utf-8 encoding in reads to be windows compliant
+        fix netCDF4 masks for nan values
+    Updated 01/2022: added global Empirical Ocean Tide model (EOT20)
+    Updated 12/2021: adjust longitude convention based on model longitude
+    Updated 07/2021: added check that tide model files are accessible
+    Updated 06/2021: add warning for tide models being entered as string
+    Updated 05/2021: added option for extrapolation cutoff in kilometers
+    Updated 03/2021: add extrapolation check where there are no invalid points
+        prevent ComplexWarning for fill values when calculating amplitudes
+        simplified inputs to be similar to binary OTIS read program
+        replaced numpy bool/int to prevent deprecation warnings
+        use uuid for reading from gzipped netCDF4 files
+    Updated 02/2021: set invalid values to nan in extrapolation
+        replaced numpy bool to prevent deprecation warning
+    Updated 12/2020: added nearest-neighbor data extrapolation
+    Updated 09/2020: set bounds error to false for regular grid interpolations
+        adjust dimensions of input coordinates to be iterable
+    Updated 08/2020: replaced griddata with scipy regular grid interpolators
+    Written 07/2020
+"""
[…]
