+++ pyTMD/xio/Model.py
@@ -1,0 +1,157 @@
+import pyproj
+import pyTMD.io
+import numpy as np
+import xarray as xr
+
+class Model(pyTMD.io.model):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def open_dataset(self, **kwargs):
+        # import tide model functions
+        # set default keyword arguments
+        kwargs.setdefault('type', self.type)
+        kwargs.setdefault('append_node', False)
+        kwargs.setdefault('scale', self.scale)
+        kwargs.setdefault('compressed', self.compressed)
+        kwargs.setdefault('constituents', None)
+        # reduce constituents if specified
+        self.reduce_constituents(kwargs['constituents'])
+        if self.format in ('OTIS', 'ATLAS-compact', 'TMD3'):
+            # extract model file in case of currents
+            if isinstance(self.model_file, dict):
+                model_file = self.model_file['u']
+            else:
+                model_file = self.model_file
+            # open OTIS/TMD3/ATLAS-compact files as xarray Dataset
+            ds = pyTMD.xio.OTIS.open_dataset(model_file, self.grid_file,
+                format=self.file_format, crs=self.crs, 
+                **kwargs)
+        elif self.format in ('ATLAS-netcdf',):
+            # extract model file in case of currents
+            if isinstance(self.model_file, dict):
+                TYPE = kwargs['type'].lower()
+                model_file = self.model_file[TYPE]
+            else:
+                model_file = self.model_file
+            # open ATLAS netCDF4 files as xarray Dataset
+            ds = pyTMD.xio.ATLAS.open_dataset(model_file, self.grid_file,
+                format=self.file_format, **kwargs)
+        elif self.format in ('GOT-ascii', 'GOT-netcdf'):
+            # open GOT ASCII/netCDF4 files as xarray Dataset
+            ds = pyTMD.xio.GOT.open_mfdataset(self.model_file,
+                format=self.file_format, **kwargs)
+        elif self.format in ('FES-ascii', 'FES-netcdf'):
+            # extract model file in case of currents
+            if isinstance(self.model_file, dict):
+                TYPE = kwargs['type'].lower()
+                model_file = self.model_file[TYPE]
+            else:
+                model_file = self.model_file
+            # open FES ASCII/netCDF4 files as xarray Dataset
+            ds = pyTMD.xio.FES.open_mfdataset(model_file,
+                version=self.version, **kwargs)
+        # add coordinate reference system to Dataset
+        ds.attrs['crs'] = self.crs.to_dict()
+        # return xarray dataset
+        return ds
+    
+    def interp(self, 
+            x: np.ndarray,
+            y: np.ndarray, 
+            crs: str | int | dict = 4326, 
+            method='linear',              
+            **kwargs
+        ):
+        """
+        Interpolate tidal constants to input coordinates
+        
+        Parameters
+        ----------
+        x: np.ndarray
+            input x-coordinates
+        y: np.ndarray
+            input y-coordinates
+        crs: str, int, or dict, default 4326
+            Coordinate reference system of input coordinates
+        method: str, default 'linear'
+            Interpolation method
+        **kwargs: dict
+            Additional keyword arguments for reading the dataset
+            
+        Returns
+        -------
+        ds: xarray.Dataset
+            interpolated tidal constants
+        """
+        # open dataset (if not already cached)
+        if not hasattr(self, '_ds'):
+            self._ds = self.open_dataset(**kwargs)
+        # pad global grids along x-dimension (if necessary)
+        if self._ds.tmd.is_global:
+            self._ds = self._ds.tmd.pad(n=1)
[…]
