+++ pyTMD/__init__.py
@@ -27,6 +27,12 @@ from pyTMD import io
 from pyTMD import solve
 from pyTMD.crs import crs
 
+# Attempt to import the experimental xarray support
+try:
+    from pyTMD import xio
+except (ImportError, AttributeError):
+    pass
+
 # Deprecated functions
 def compute_tide_corrections(*args, **kwargs):
     """Wrapper function to compute tide corrections
