### 2025-11-07 639eae3
feat: add output ATLAS-netcdf functions
