### 2025-11-07 12652b2
feat: add output ATLAS-netcdf functions
