### 2025-11-07 32e7bf4
refactor: make ATLAS-compact accessor `compact`
