### 2025-11-07 45239ed
refactor: make ATLAS-compact accessor `compact`
