### 2025-11-07 f1e95eb
refactor: restructure database and providers
