+++ pyTMD/io/model.py
@@ -18,6 +18,7 @@ PYTHON DEPENDENCIES:
 UPDATE HISTORY:
     Updated 11/2025: use default cache directory if directory is None
         added crs property for model coordinate reference system
+        refactor to use new simpler (flattened) database format
     Updated 08/2025: use numpy degree to radian conversions
         update node equilibrium tide estimation
         add functions for converting a model to an xarray DataTree object
@@ -104,8 +105,11 @@ _default_directory = get_cache_path()
 @dataclass
 class DataBase:
     """Class for pyTMD model database"""
-    current: dict
-    elevation: dict
+    __dict__: dict
+
+    def update(self, d: dict):
+        """Update the keys of the model database"""
+        self.__dict__.update(d)
 
     def keys(self):
         """Returns the keys of the model database"""
@@ -119,9 +123,13 @@ class DataBase:
         """Returns the items of the model database"""
         return self.__dict__.items()
 
+    def __str__(self):
+        """String representation of the ``DataBase`` object
+        """
+        return str(self.__dict__)
+
     def __getitem__(self, key):
         return getattr(self, key)
-
 
 # PURPOSE: load the JSON database of model files
 def load_database(extra_databases: list = []):
@@ -161,11 +169,9 @@ def load_database(extra_databases: list = []):
             # extract JSON data
             with db.open(mode='r', encoding='utf-8') as fid:
                 extra_database = json.load(fid)
-        # Add additional models to database, accounting
-        # for top level elevation and current dict keys
-        for key, val in extra_database.items():
-            parameters[key].update(val)
-    return DataBase(**parameters)
+        # Add additional models to database
+        parameters.update(extra_database)
+    return DataBase(parameters)
 
 class model:
     """Retrieves tide model parameters for named models or
@@ -174,22 +180,9 @@ class model:
 
     Attributes
     ----------
-    atl03: str
-        HDF5 dataset string for output ATL03 tide heights
-    atl06: str
-        HDF5 dataset string for output ATL06 tide heights
-    atl07: str
-        HDF5 dataset string for output ATL07 tide heights
-    atl10: str
-        HDF5 dataset string for output ATL10 tide heights
-    atl11: str
-        HDF5 dataset string for output ATL11 tide heights
-    atl12: str
-        HDF5 dataset string for output ATL12 tide heights
+
     compressed: bool
         Model files are gzip compressed
-    constituents: list or None
-        Model constituents for ``FES`` models
     description: str
         HDF5 ``description`` attribute string for output tide heights
     directory: str, pathlib.Path or None, default None
@@ -203,38 +196,28 @@ class model:
     format: str
         Model format
 
-            - ``OTIS``
             - ``ATLAS-compact``
-            - ``TMD3``
             - ``ATLAS-netcdf``
             - ``GOT-ascii``
             - ``GOT-netcdf``
             - ``FES-ascii``
             - ``FES-netcdf``
-    gla12: str
-        HDF5 dataset string for output GLA12 tide heights
+            - ``OTIS``
+            - ``TMD3``
     grid_file: pathlib.Path
-        Model grid file for ``OTIS``, ``ATLAS-compact``, ``ATLAS-netcdf``, and ``TMD3`` models
+        Model grid file for ``OTIS``, ``ATLAS-compact``, and ``ATLAS-netcdf`` models
     gzip: bool
         Suffix if model is compressed
-    long_name: str
-        HDF5 ``long_name`` attribute string for output tide heights
     minor: list or None
         Minor constituents for inference
     model_file: pathlib.Path or list
         Model constituent file or list of files
     name: str
         Model name
-    projection: str
-        Model projection for ``OTIS``, ``ATLAS-compact`` and ``TMD3`` models
[…]
