### 2025-11-07 fbfe77a
refactor: restructure database and providers
