+++ pyTMD/__init__.py
@@ -44,4 +44,4 @@ def compute_tide_corrections(*args, **kwargs):
 __version__ = pyTMD.version.version
 # read model database
 models = io.load_database()
-__models__ = {key:list(val.keys()) for key, val in models.items()}
+__models__ = sorted(models.keys())
