### 2025-11-07 46855c2
Update model.py
