+++ pyTMD/io/model.py
@@ -807,18 +807,10 @@ class model:
         if isinstance(self[type].model_file, list):
             # multiple file elevation case
             # filter model files to constituents
-            self[type].model_file = [self[type].model_file[self.constituents.index(c)]
+            self[type].model_file = [
+                self[type].model_file[self.constituents.index(c)]
                 for c in constituents if (c in self.constituents)
             ]
-        elif isinstance(self[type].model_file, dict) and \
-            isinstance(self[type].model_file['u'], list):
-            # multiple file currents case
-            for key, val in self[type].model_file.items():
-                # reduce list of model files to constituents
-                # filter model files to constituents
-                self[type].model_file[key] = [val[self.constituents.index(c)]
-                    for c in constituents if (c in self.constituents)
-                ]
         # update list of constituents
         self.parse_constituents(type=type)
         # return self
@@ -855,52 +847,46 @@ class model:
         # import tide model functions
         from pyTMD.io import OTIS, ATLAS, GOT, FES
         # set default keyword arguments
-        kwargs.setdefault('type', self.type)
+        kwargs.setdefault('type', 'z')
         kwargs.setdefault('append_node', False)
         kwargs.setdefault('scale', self.scale)
         kwargs.setdefault('constituents', None)
         # reduce constituents if specified
         self.reduce_constituents(kwargs['constituents'])
+        # model type
+        mtype = kwargs['type'].lower()
+        # extract model file
+        model_file = self[mtype].model_file
         # read tidal constants and interpolate to grid points
         if self.format in ('OTIS', 'ATLAS-compact', 'TMD3'):
-            # extract model file in case of currents
-            if isinstance(self.model_file, dict):
-                model_file = self.model_file['u']
+            # model grid file
+            if hasattr(self[mtype], 'grid_file'):
+                grid_file = self[mtype].grid_file
             else:
-                model_file = self.model_file
+                grid_file = model_file
             # extract tidal constants for model type
             amp,ph,D,c = OTIS.extract_constants(lon, lat,
-                self.grid_file, model_file, self.projection,
+                grid_file, model_file, self.projection,
                 grid=self.file_format, **kwargs)
         elif self.format in ('ATLAS-netcdf',):
-            # extract model file in case of currents
-            if isinstance(self.model_file, dict):
-                TYPE = kwargs['type'].lower()
-                model_file = self.model_file[TYPE]
-            else:
-                model_file = self.model_file
+            # model grid file
+            grid_file = self[mtype].grid_file
             # extract tidal constants for model type
             amp,ph,D,c = ATLAS.extract_constants(lon, lat,
-                self.grid_file, model_file,
+                grid_file, model_file,
                 compressed=self.compressed, **kwargs)
         elif self.format in ('GOT-ascii', 'GOT-netcdf'):
             # extract tidal constants for model type
             amp,ph,c = GOT.extract_constants(lon, lat,
-                self.model_file, grid=self.file_format,
+                model_file, grid=self.file_format,
                 compressed=self.compressed, **kwargs)
         elif self.format in ('FES-ascii', 'FES-netcdf'):
-            # extract model file in case of currents
-            if isinstance(self.model_file, dict):
-                TYPE = kwargs['type'].lower()
-                model_file = self.model_file[TYPE]
-            else:
-                model_file = self.model_file
             # extract tidal constants for model type
             amp,ph = FES.extract_constants(lon, lat,
                 model_file, version=self.version,
                 compressed=self.compressed, **kwargs)
             # available model constituents
-            c = self.constituents
[…]
