### 2025-11-07 19ec9fb
Update model.py
