### 2025-11-08 c8e8dda
feat: unit-aware conversions
