+++ pyproject.toml
@@ -61,7 +61,7 @@ Issues = "https://github.com/pyTMD/pyTMD/issues"
 [project.optional-dependencies]
 doc = ["docutils", "graphviz", "myst-nb", "numpydoc", "sphinx", "sphinx-argparse>=0.4", "sphinxcontrib-bibtex", "sphinx-design", "sphinx_rtd_theme"]
 all = ["cartopy", "ipyleaflet", "ipywidgets", "jplephem", "jupyterlab", "matplotlib", "notebook", "pandas", "xarray"]
-aws = ["dask", "geopandas", "h5netcdf", "obstore", "pyarrow", "s3fs", "zarr>=3"]
+aws = ["dask", "geopandas", "h5netcdf", "obstore", "pint", "pyarrow", "s3fs", "zarr>=3"]
 dev = ["flake8", "gh", "oct2py", "pytest>=4.6", "pytest-cov", "pytest-xdist"]
 
 [project.scripts]
@@ -202,6 +202,7 @@ dask = "*"
 geopandas = "*"
 h5netcdf = "*"
 obstore = "*"
+pint = "*"
 pyarrow = "*"
 s3fs = "*"
 zarr = ">=3"

[…]
