### 2025-11-08 4334fd9
feat: unit-aware conversions
