+++ pyTMD/io/model.py
@@ -128,6 +128,9 @@ class DataBase:
         """
         return str(self.__dict__)
 
+    def get(self, key, default=None):
+        return getattr(self, key, default) or default
+
     def __getitem__(self, key):
         return getattr(self, key)
 
