### 2025-11-08 2e43833
feat: add extrapolation option to dataset interp
