### 2025-11-08 ba45ae0
ci: don't run actions for WIP PRs
