+++ .github/workflows/python-request.yml
@@ -10,6 +10,7 @@ on:
       - scripts/**
       - test/**
       - .github/workflows/python-request.yml
+    types: [opened, synchronize, reopened, ready_for_review]
   schedule:
     # * is a special character in YAML so you have to quote this string
     - cron:  '0 0 1 * *'
@@ -17,6 +18,7 @@ on:
 jobs:
   build:
 
+    if: github.event.pull_request.draft == false
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
