### 2025-11-10 eb3884b
feat: add OTIS variable attributes
