+++ pyTMD/xio/OTIS.py
@@ -120,6 +120,71 @@ __all__ = [
     'ATLASDataset',
 ]
 
+# variable attributes
+_attributes = dict(
+    z = dict(
+        bathymetry = dict(
+            standard_name="ocean_depth",
+            long_name="ocean depth",
+            units="m",
+        ),
+        mask = dict(
+            standard_name="sea_binary_mask",
+            long_name="land-sea mask",
+            units="1",
+        ),
+        elevation = dict(
+            standard_name="tide_height",
+            long_name="tidal_elevation",
+            units="m",
+        )
+    ),
+    u = dict(
+        bathymetry=dict(
+            standard_name="ocean_depth",
+            long_name="ocean depth at u-locations",
+            units="m",
+        ),
+        mask = dict(
+            standard_name="sea_binary_mask",
+            long_name="land-sea mask at u-locations",
+            units="1",
+        ),
+        current = dict(
+            standard_name="eastward_tidal_current",
+            long_name="zonal tidal currents on u-grid",
+            units="m/s",
+        ),
+        transport = dict(
+            standard_name="eastward_tidal_transport",
+            long_name="zonal tidal transports on u-grid",
+            units="m**2/s",
+        )
+    ),
+    v = dict(
+        bathymetry = dict(
+            standard_name="ocean_depth",
+            long_name="ocean depth at v-locations",
+            units="m",
+        ),
+        mask = dict(
+            standard_name="sea_binary_mask",
+            long_name="land-sea mask at v-locations",
+            units="1",
+        ),
+        current = dict(
+            standard_name="northward_tidal_current",
+            long_name="meridional tidal currents on v-grid",
+            units="m/s",
+        ),
+        transport = dict(
+            standard_name="northward_tidal_transport",
+            long_name="meridional tidal transports on v-grid",
+            units="m**2/s",
+        )
+    )
+)
+
 # PURPOSE: read tide model files
 def open_dataset(
         model_file: str | list | pathlib.Path,
@@ -150,6 +215,9 @@ def open_dataset(
     ds: xarray.Dataset
         tide model data
     """
+    # set default keyword arguments
+    kwargs.setdefault('type', 'z')
+    mtype = kwargs.get('type').lower()
     # open file(s) as xarray dataset
     if (format == 'OTIS'):
         # OTIS (single or multi-file)
@@ -168,6 +236,13 @@ def open_dataset(
         ds = open_tmd3_dataset(model_file, **kwargs)
     # add attributes
     ds.attrs['format'] = format
+    # convert transports to currents if necessary
+    if kwargs['type'] in ('u','v'):
+        # convert transports to currents
+        ds[ds.tmd.constituents] /= ds['bathymetry']
+        # update attributes
+        ds[ds.tmd.constituents].attrs.update(_attributes[mtype]['current'])
+        ds.attrs['units'] = _attributes[mtype]['current']['units']
     # return xarray dataset
     return ds
 
@@ -188,10 +263,10 @@ def open_mfdataset(
         Tidal variable to read
 
             - ``'z'``: heights
-            - ``'u'``: horizontal transport velocities
-            - ``'U'``: horizontal depth-averaged transport
-            - ``'v'``: vertical transport velocities
-            - ``'V'``: vertical depth-averaged transport
+            - ``'u'``: zonal currents
+            - ``'U'``: zonal depth-averaged transport
+            - ``'v'``: meridional currents
+            - ``'V'``: meridional depth-averaged transport
     **kwargs: dict
[…]
