### 2025-11-10 806db63
feat: add OTIS variable attributes
