### 2025-11-10 be8af12
feat: add ATLAS currents conversion
