### 2025-11-10 e48f01f
feat: add ATLAS currents conversion
