### 2025-11-10 a130661
refactor: do units conversions with dataarray accessor
