### 2025-11-10 645db8d
refactor: do units conversions with dataarray accessor
