### 2025-11-10 0e82895
assert model type is z, u, or v
