### 2025-11-10 119dc67
assert model type is z, u, or v
