### 2025-11-10 41c4ac6
feat: allow chunking and dask computation
