### 2025-11-10 923618d
feat: allow chunking and dask computation
