+++ pyTMD/xio/OTIS.py
@@ -295,7 +295,6 @@ def open_otis_dataset(
         model_file: str | list | pathlib.Path,
         grid_file: str | pathlib.Path,
         type: str | None = 'z',
-        use_mmap : bool = False,
         **kwargs
     ):
     """
@@ -328,22 +327,22 @@ def open_otis_dataset(
     # default coordinate reference system
     crs = kwargs.get('crs', 4326)
     # open grid file
-    ds1 = open_otis_grid(grid_file, use_mmap=use_mmap)
+    ds1 = open_otis_grid(grid_file, **kwargs)
     # add attributes
     ds1.attrs['crs'] = pyproj.CRS.from_user_input(crs).to_dict()
     # open model file(s)
     if isinstance(model_file, list):
         # multi-file datasets
-        ds2 = open_mfdataset(model_file, type=type, use_mmap=use_mmap)
+        ds2 = open_mfdataset(model_file, type=type, **kwargs)
     elif (type == 'z'):
         # elevations
-        ds2 = open_otis_elevation(model_file, use_mmap=use_mmap)
+        ds2 = open_otis_elevation(model_file, **kwargs)
     elif type in ('u', 'U'):
         # transports are returned as (u,v)
-        ds2 = open_otis_transport(model_file, use_mmap=use_mmap)[0]
+        ds2 = open_otis_transport(model_file, **kwargs)[0]
     elif type in ('v', 'V'):
         # transports are returned as (u,v)
-        ds2 = open_otis_transport(model_file, use_mmap=use_mmap)[1]
+        ds2 = open_otis_transport(model_file, **kwargs)[1]
     # merge datasets
     ds = ds1.otis.merge(ds2, type=type)
     # add attributes
@@ -355,6 +354,7 @@ def open_atlas_dataset(
         model_file: str | pathlib.Path,
         grid_file: str | pathlib.Path,
         type: str | None = 'z',
+        chunks: int | dict | str | None = None,
         use_mmap : bool = False,
         **kwargs
     ):
@@ -377,6 +377,8 @@ def open_atlas_dataset(
             - ``'V'``: meridional depth-averaged transport
     crs: int, str or dict, default 4326
         coordinate reference system for the model data
+    chunks: int, dict, str, or None, default None
+        coerce output to specified chunks
     use_mmap: bool, default False
         use memory mapping to read data
 
@@ -389,22 +391,22 @@ def open_atlas_dataset(
     crs = kwargs.get('crs', 4326)
     # open grid file
     dsg, dtg = open_atlas_grid(grid_file, use_mmap=use_mmap)
-    ds1 = dsg.compact.combine_local(dtg)
+    ds1 = dsg.compact.combine_local(dtg, chunks=chunks)
     # add attributes
     ds1.attrs['crs'] = pyproj.CRS.from_user_input(crs).to_dict()
     # open model file(s)
     if (type == 'z'):
         # elevations are returned as (z, localz)
         dsh, dth = open_atlas_elevation(model_file, use_mmap=use_mmap)
-        ds2 = dsh.compact.combine_local(dth)
+        ds2 = dsh.compact.combine_local(dth, chunks=chunks)
     elif type in ('u', 'U'):
         # transports are returned as (u, v, localu, localv)
         dsu, dtu, dsv, dtv = open_atlas_transport(model_file, use_mmap=use_mmap)
-        ds2 = dsu.compact.combine_local(dtu)
+        ds2 = dsu.compact.combine_local(dtu, chunks=chunks)
     elif type in ('v', 'V'):
         # transports are returned as (u, v, localu, localv)
[…]
