+++ pyTMD/xio/__init__.py
@@ -5,4 +5,5 @@ from .ATLAS import *
 from .FES import *
 from .GOT import *
 from .OTIS import *
+from .IERS import *
 from .Model import Model
