### 2025-11-11 f6d12ff
feat: add xarray ocean pole tide opener
