+++ pyTMD/xio/IERS.py
@@ -1,0 +1,150 @@
+#!/usr/bin/env python
+u"""
+IERS.py
+Written by Tyler Sutterley (11/2025)
+
+Reads ocean pole load tide coefficients provided by IERS
+http://maia.usno.navy.mil/conventions/2010/2010_official/chapter7/tn36_c7.pdf
+http://maia.usno.navy.mil/conventions/2010/2010_update/chapter7/icc7.pdf
+
+PYTHON DEPENDENCIES:
+    numpy: Scientific Computing Tools For Python
+        https://numpy.org
+        https://numpy.org/doc/stable/user/numpy-for-matlab-users.html
+    xarray: N-D labeled arrays and datasets in Python
+        https://docs.xarray.dev/en/stable/
+
+REFERENCES:
+    S. Desai, "Observing the pole tide with satellite altimetry", Journal of
+        Geophysical Research: Oceans, 107(C11), 2002. doi: 10.1029/2001JC001224
+    S. Desai, J. Wahr and B. Beckley "Revisiting the pole tide for and from
+        satellite altimetry", Journal of Geodesy, 89(12), p1233-1243, 2015.
+        doi: 10.1007/s00190-015-0848-7
+
+UPDATE HISTORY:
+    Updated 11/2025: near-complete rewrite of program to use xarray
+    Updated 08/2024: convert outputs to be in -180:180 longitude convention
+        added function to interpolate ocean pole tide values to coordinates
+        renamed from ocean_pole_tide to IERS
+    Updated 06/2024: use np.clongdouble instead of np.longcomplex
+    Updated 05/2023: add default for ocean pole tide file
+    Updated 04/2023: using pathlib to define and expand paths
+    Updated 03/2023: add basic variable typing to function inputs
+    Updated 12/2022: refactor ocean pole tide read programs under io
+    Updated 04/2022: updated docstrings to numpy documentation format
+        use longcomplex data format to be windows compliant
+    Updated 07/2021: added check that ocean pole tide file is accessible
+    Updated 03/2021: replaced numpy bool/int to prevent deprecation warnings
+    Updated 08/2020: output north load and east load deformation components
+    Updated 07/2020: added function docstrings
+    Updated 12/2018: Compatibility updates for Python3
+    Written 09/2017
+"""
+from __future__ import annotations
+
+import re
+import gzip
+import pathlib
+import numpy as np
+import xarray as xr
+from pyTMD.utilities import get_data_path
+
+__all__ = [
+    "open_dataset",
+]
+
+# ocean pole tide file from Desai (2002) and IERS conventions
+_ocean_pole_tide_file = get_data_path(['data','opoleloadcoefcmcor.txt.gz'])
+
+# PURPOSE: read real and imaginary ocean pole tide coefficients
+def open_dataset(
+        input_file: str | pathlib.Path = _ocean_pole_tide_file,
+        chunks: int | dict | str | None = None,
+        **kwargs
+    ):
+    """
+    Open Ocean Pole Tide ASCII files from
+    :cite:p:`Desai:2002ev,Desai:2015jr`
+
+    Parameters
+    ----------
+    input_file: str or pathlib.Path
+        Ocean pole tide file
+    chunks: int | dict | str | None, default None
+        coerce output to specified chunks
+    compressed: bool, default False
+        Input file is gzip compressed
+
+    Returns
+    -------
+    ds: xarray.Dataset
+        Ocean pole tide data
+    """
+    # set default keyword arguments
+    kwargs.setdefault('compressed', True)
+    # tilde-expand input file
+    input_file = pathlib.Path(input_file).expanduser()
+    # read compressed ocean pole tide file
+    if kwargs['compressed']:
[…]
