### 2025-11-12 b0692f5
move open_dataset over to model.py
