+++ pyTMD/xio/Model.py
@@ -1,163 +1,0 @@
-#!/usr/bin/env python
-u"""
-Model.py
-Written by Tyler Sutterley (11/2025)
-"""
-import pyproj
-import pyTMD.io
-import numpy as np
-import xarray as xr
-
-# PURPOSE: experimental extension of pyTMD.io.model for xarray I/O
-class Model(pyTMD.io.model):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def open_dataset(self, **kwargs):
-        # import tide model functions
-        # set default keyword arguments
-        kwargs.setdefault('type', 'z')
-        kwargs.setdefault('convert_units', True)
-        kwargs.setdefault('append_node', False)
-        kwargs.setdefault('compressed', self.compressed)
-        kwargs.setdefault('constituents', None)
-        # model type
-        mtype = kwargs['type'].lower()
-        assert mtype in ('z', 'u', 'v'), "Invalid model type"
-        # extract model file
-        model_file = self[mtype].get('model_file')
-        # reduce constituents if specified
-        self.reduce_constituents(kwargs['constituents'])
-        if self.format in ('OTIS', 'ATLAS-compact', 'TMD3'):
-            # open OTIS/TMD3/ATLAS-compact files as xarray Dataset
-            ds = pyTMD.xio.OTIS.open_dataset(model_file, 
-                grid_file=self[mtype].get('grid_file'),
-                format=self.file_format,
-                crs=self.crs, **kwargs)
-        elif self.format in ('ATLAS-netcdf',):
-            # open ATLAS netCDF4 files as xarray Dataset
-            ds = pyTMD.xio.ATLAS.open_dataset(model_file, 
-                grid_file=self[mtype].get('grid_file'),
-                format=self.file_format, **kwargs)
-        elif self.format in ('GOT-ascii', 'GOT-netcdf'):
-            # open GOT ASCII/netCDF4 files as xarray Dataset
-            ds = pyTMD.xio.GOT.open_mfdataset(model_file,
-                format=self.file_format, **kwargs)
-        elif self.format in ('FES-ascii', 'FES-netcdf'):
-            # open FES ASCII/netCDF4 files as xarray Dataset
-            ds = pyTMD.xio.FES.open_mfdataset(model_file,
-                version=self.version, **kwargs)
-        # add coordinate reference system to Dataset
-        ds.attrs['crs'] = self.crs.to_dict()
-        # list of constituents
-        c = ds.tmd.constituents
-        # set units attribute if not already set
-        # (uses value defined in the model database)
-        ds[c].attrs['units'] = ds[c].attrs.get('units', self[mtype].units)
-        # convert to default units
-        if kwargs['convert_units']:
-            ds = ds.tmd.to_default_units()
-        # return xarray dataset
-        return ds
-    
-    def interp(self, 
-            x: np.ndarray,
-            y: np.ndarray, 
-            crs: str | int | dict = 4326, 
-            method='linear',           
-            extrapolate: bool = False,
-            cutoff: int | float = np.inf,   
-            **kwargs
-        ):
-        """
-        Interpolate tidal constants to input coordinates
-        
-        Parameters
-        ----------
-        x: np.ndarray
-            input x-coordinates
-        y: np.ndarray
-            input y-coordinates
-        crs: str, int, or dict, default 4326
-            Coordinate reference system of input coordinates
-        method: str, default 'linear'
-            Interpolation method
-        extrapolate: bool, default False
-            Flag to extrapolate values using nearest-neighbors
-        cutoff: int or float, default np.inf
-            Maximum distance for extrapolation
-        **kwargs: dict
[…]
