### 2025-11-12 50f6d3a
move open_dataset over to model.py
