+++ pyTMD/xio/__init__.py
@@ -6,4 +6,3 @@ from .FES import *
 from .GOT import *
 from .OTIS import *
 from .IERS import *
-from .Model import Model
