### 2025-11-12 f33f24b
add scale back for backwards compat
might remove when xarray io is finalized
