### 2025-11-13 5342d1e
feat: add prediction functions
refactor: start working on notebooks
