### 2025-11-13 e99a21f
feat: add prediction functions
refactor: start working on notebooks
