### 2025-11-14 4453be4
feat: add infer function to dataset accessor
