+++ pyTMD/predict.py
@@ -129,7 +129,11 @@ _mjd_tide = 48622.0
 # number of days between the Julian day epoch and the tide epoch
 _jd_tide = _jd_mjd + _mjd_tide
 
-def dataset(ds, t: float | np.ndarray, **kwargs):
+def dataset(
+        ds: xr.Dataset,
+        t: float | np.ndarray,
+        **kwargs
+    ):
     """
     Predict tides from ``Dataset`` at times
 
@@ -156,6 +160,7 @@ def dataset(ds, t: float | np.ndarray, **kwargs):
     # load the nodal corrections
     pu, pf, G = pyTMD.arguments.arguments(MJD,
         constituents, **kwargs)
+    # dataset of arguments
     arguments = xr.Dataset(
         data_vars=dict(
             u=(['time','constituent'], pu),
@@ -178,7 +183,7 @@ def dataset(ds, t: float | np.ndarray, **kwargs):
         if i == 0:
             darr = arg.f*ds[c].real*np.cos(th) - \
                 arg.f*ds[c].imag*np.sin(th)
-            darr.attrs['units'] = str(ds[c].units)
+            darr.attrs['units'] = ds[c].attrs.get('units', None)
         else:
             darr += arg.f*ds[c].real*np.cos(th) - \
                 arg.f*ds[c].imag*np.sin(th)
@@ -379,9 +384,8 @@ def time_series(t: float | np.ndarray,
 
 # PURPOSE: infer the minor corrections from the major constituents
 def infer_minor(
+        ds: xr.Dataset,
         t: float | np.ndarray,
-        zmajor: np.ndarray,
-        constituents: list | np.ndarray,
         **kwargs
     ):
     """
@@ -391,12 +395,10 @@ def infer_minor(
 
     Parameters
     ----------
+    ds: xarray.Dataset
+        Dataset containing major tidal harmonic constants
     t: float or np.ndarray
         days relative to 1992-01-01T00:00:00
-    zmajor: np.ndarray
-        Complex HC for given constituents/points
-    constituents: list
-        tidal constituent IDs
     deltat: float or np.ndarray, default 0.0
         time correction for converting to Ephemeris Time (days)
     corrections: str, default 'OTIS'
@@ -424,21 +426,20 @@ def infer_minor(
     dh = 0.0
     # infer short-period tides for minor constituents
     if kwargs['corrections'] in ('GOT',):
-        dh += _infer_semi_diurnal(t, zmajor, constituents, **kwargs)
-        dh += _infer_diurnal(t, zmajor, constituents, **kwargs)
+        dh += _infer_semi_diurnal(ds, t, **kwargs)
+        dh += _infer_diurnal(ds, t, **kwargs)
     else:
-        dh += _infer_short_period(t, zmajor, constituents, **kwargs)
+        dh += _infer_short_period(ds, t, **kwargs)
     # infer long-period tides for minor constituents
     if kwargs['infer_long_period']:
-        dh += _infer_long_period(t, zmajor, constituents, **kwargs)
+        dh += _infer_long_period(ds, t, **kwargs)
     # return the inferred values
     return dh
 
 # PURPOSE: infer short-period minor constituents
 def _infer_short_period(
+        ds: xr.Dataset,
         t: float | np.ndarray,
-        zmajor: np.ndarray,
-        constituents: list | np.ndarray,
         **kwargs
     ):
     """
@@ -448,12 +449,10 @@ def _infer_short_period(
 
     Parameters
[…]
