### 2025-11-14 b33ceab
feat: add infer function to dataset accessor
