+++ doc/source/conf.py
@@ -76,6 +76,7 @@ if on_rtd:
     nb_execution_mode = "auto"
     nb_execution_excludepatterns = [
         "Plot-Antarctic-Tidal-Currents.ipynb",
+        "Plot-Antarctic-Tide-Range.ipynb",
         "Plot-ATLAS-Compact.ipynb"
     ]
 else:
