### 2025-11-15 ca2782a
test: update model tests to use xarray interface
