+++ test/test_perth3_read.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 u"""
-test_perth3_read.py (10/2025)
+test_perth3_read.py (11/2025)
 Tests the read program to verify that constituents are being extracted
 Tests that interpolated results are comparable to NASA PERTH3 program
 
@@ -14,6 +14,7 @@ PYTHON DEPENDENCIES:
         https://pypi.org/project/timescale/
 
 UPDATE HISTORY:
+    Updated 11/2025: using new xarray interface for tidal model data
     Updated 10/2025: split directories between validation and model data
         fetch data from pyTMD developers test data repository
     Updated 09/2025: added check if running on GitHub Actions or locally
@@ -52,18 +53,16 @@ import timescale
 filename = inspect.getframeinfo(inspect.currentframe()).filename
 filepath = pathlib.Path(filename).absolute().parent
 
-# parameterize interpolation method
-@pytest.mark.parametrize("METHOD", ['spline','linear','bilinear'])
-@pytest.mark.parametrize("CROP", [False, True])
 # PURPOSE: Tests that interpolated results are comparable to PERTH3 program
-def test_verify_GOT47(directory, METHOD, CROP):
+def test_verify_GOT47(directory):
     # model parameters for GOT4.7
-    model = pyTMD.io.model(directory,compressed=True).elevation('GOT4.7')
+    model = pyTMD.io.model(directory).from_database('GOT4.7')
     # perth3 test program infers m4 tidal constituent
     # constituent files included in test
     constituents = ['q1','o1','p1','k1','n2','m2','s2','k2','s1']
-    # keep units consistent with test outputs
-    model.scale = 1.0
+    model.reduce_constituents(constituents)
+    # open dataset
+    ds = model.open_dataset(type='z', use_default_units=False)
 
     # read validation dataset
     with gzip.open(filepath.joinpath('perth_output_got4.7.gz'),'r') as fid:
@@ -89,144 +88,25 @@ def test_verify_GOT47(directory, METHOD, CROP):
     # interpolate delta times
     deltat = ts.tt_ut1
 
+    # convert to xarray DataArrays
+    X = xr.DataArray(lon, dims=('time'))
+    Y = xr.DataArray(lat, dims=('time'))
     # extract amplitude and phase from tide model
-    amp,ph,cons = model.extract_constants(lon, lat,
-        method=METHOD, crop=CROP, constituents=constituents)
-    assert all(c in constituents for c in cons)
-    # calculate complex phase in radians for Euler's
-    cph = -1j*ph*np.pi/180.0
-    # calculate constituent oscillations
-    hc = amp*np.exp(cph)
+    local = ds.tmd.interp(X, Y, extrapolate=True)
 
-    # allocate for out tides at point
-    tide = np.ma.zeros((npts))
-    tide.mask = np.zeros((npts),dtype=bool)
     # predict tidal elevations at time and infer minor corrections
-    tide.mask[:] = np.any(hc.mask, axis=1)
-    tide.data[:] = pyTMD.predict.drift(ts.tide, hc, cons,
-        deltat=deltat, corrections='perth3')
-    minor = pyTMD.predict.infer_minor(ts.tide, hc, cons,
-        deltat=deltat, corrections='perth3', minor=model.minor)
-    tide.data[:] += minor.data[:]
+    tide = local.tmd.predict(ts.tide, deltat=deltat,
+        corrections='perth3')
+    tide += local.tmd.infer(ts.tide, deltat=deltat,
+        corrections='perth3', minor=model.minor)
 
     # will verify differences between model outputs are within tolerance
     eps = 0.01
     # calculate differences between perth3 and python version
     difference = np.ma.zeros((npts))
[…]
