+++ test/test_atlas_read.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 u"""
-test_atlas_read.py (10/2025)
+test_atlas_read.py (11/2025)
 Tests that ATLAS compact and netCDF4 data can be downloaded from AWS S3 bucket
 Tests the read program to verify that constituents are being extracted
 
@@ -16,6 +16,7 @@ PYTHON DEPENDENCIES:
         https://pypi.org/project/timescale/
 
 UPDATE HISTORY:
+    Updated 11/2025: using new xarray interface for tidal model data
     Updated 10/2025: split directories between validation and model data
         fetch data from pyTMD developers test data repository
     Updated 09/2025: added check if running on GitHub Actions or locally
@@ -42,6 +43,7 @@ import pytest
 import inspect
 import pathlib
 import numpy as np
+import xarray as xr
 import pyTMD
 import timescale
 
@@ -49,17 +51,14 @@ import timescale
 filename = inspect.getframeinfo(inspect.currentframe()).filename
 filepath = pathlib.Path(filename).absolute().parent
 
-# parameterize interpolation method
-@pytest.mark.parametrize("METHOD", ['spline','nearest'])
-@pytest.mark.parametrize("EXTRAPOLATE", [False])
-@pytest.mark.parametrize("CROP", [False, True])
 # PURPOSE: Tests that interpolated results are comparable to OTPSnc program
-def test_read_TPXO9_v2(directory, METHOD, EXTRAPOLATE, CROP):
+def test_read_TPXO9_v2(directory):
     # model parameters for TPXO9-atlas-v2
-    m = pyTMD.io.model(directory,compressed=True).elevation('TPXO9-atlas-v2-nc')
-    constituents = ['m2','s2','k1','o1']
-    # convert units from millimeters to meters
-    m.scale = 1.0/1000.0
+    m = pyTMD.io.model(directory).from_database('TPXO9-atlas-v2-nc', type='z')
+    # reduce to constituents for test
+    m.reduce_constituents(['m2','s2','k1','o1'])
+    # open dataset
+    ds = m.open_dataset(type='z')
 
     # read validation dataset (m2, s2, k1, o1)
     names = ('Lat', 'Lon', 'm2_amp', 'm2_ph', 's2_amp', 's2_ph',
@@ -68,166 +67,37 @@ def test_read_TPXO9_v2(directory, METHOD, EXTRAPOLATE, CROP):
     val = np.loadtxt(filepath.joinpath('extract_HC_sample_out.gz'),
         skiprows=3,dtype=dict(names=names,formats=formats))
 
+    # convert to xarray DataArrays
+    X = xr.DataArray(val['Lon'], dims=('time'))
+    Y = xr.DataArray(val['Lat'], dims=('time'))
     # extract amplitude and phase from tide model
-    amp,ph,c = m.extract_constants(val['Lon'], val['Lat'],
-        constituents=constituents, method=METHOD,
-        extrapolate=EXTRAPOLATE, crop=CROP)
-    # convert phase from 0:360 to -180:180
-    ph[ph > 180] -= 360.0
+    local = ds.tmd.interp(X, Y)
 
     # will verify differences between model outputs are within tolerance
     amp_eps = 0.05
     ph_eps = 10.0
     # calculate differences between OTPSnc and python version
-    for i,cons in enumerate(c):
-        # verify constituents
-        assert (cons == constituents[i])
-        # calculate difference in amplitude and phase
-        amp_diff = amp[:,i] - val[f'{cons}_amp']
-        ph_diff = ph[:,i] - val[f'{cons}_ph']
+    for i,cons in enumerate(local.tmd.constituents):
+        # amplitude and phase
[…]

[…]
