### 2025-11-15 26e4a92
test: update model tests to use xarray interface
