+++ pyTMD/compute.py
@@ -160,6 +160,7 @@ import timescale.eop
 import timescale.time
 # attempt imports
 pyproj = pyTMD.utilities.import_dependency('pyproj')
+xr = pyTMD.utilities.import_dependency('xarray')
 
 __all__ = [
     "corrections",
@@ -230,16 +231,12 @@ def tide_elevations(
         x: np.ndarray, y: np.ndarray, delta_time: np.ndarray,
         DIRECTORY: str | pathlib.Path | None = _default_directory,
         MODEL: str | None = None,
-        GZIP: bool = False,
         DEFINITION_FILE: str | pathlib.Path | IOBase | None = None,
-        CROP: bool = False,
-        BOUNDS: list | np.ndarray | None = None,
-        BUFFER: int | float | None = None,
         EPSG: str | int = 4326,
         EPOCH: list | tuple = (2000, 1, 1, 0, 0, 0),
         TYPE: str | None = 'drift',
         TIME: str = 'UTC',
-        METHOD: str = 'spline',
+        METHOD: str = 'linear',
         EXTRAPOLATE: bool = False,
         CUTOFF: int | float = 10.0,
         CORRECTIONS: str | None = None,
@@ -248,7 +245,6 @@ def tide_elevations(
         MINOR_CONSTITUENTS: list | None = None,
         APPEND_NODE: bool = False,
         APPLY_FLEXURE: bool = False,
-        FILL_VALUE: float = np.nan,
         **kwargs
     ):
     """
@@ -267,16 +263,8 @@ def tide_elevations(
         working data directory for tide models
     MODEL: str or NoneType, default None
         Tide model to use in correction
-    GZIP: bool, default False
-        Tide model files are gzip compressed
     DEFINITION_FILE: str, pathlib.Path, io.IOBase or NoneType, default None
         Tide model definition file for use
-    CROP: bool, default False
-        Crop tide model data to (buffered) bounds
-    BOUNDS: list, np.ndarray or NoneType, default None
-        Boundaries for cropping tide model data
-    BUFFER: int, float or NoneType, default None
-        Buffer distance for cropping tide model data
     EPSG: int, default: 4326 (WGS84 Latitude and Longitude)
         Input coordinate system
     EPOCH: tuple, default (2000,1,1,0,0,0)
@@ -323,12 +311,10 @@ def tide_elevations(
         Apply ice flexure scaling factor to height values
 
         Only valid for models containing flexure fields
-    FILL_VALUE: float, default np.nan
-        Output invalid value
 
     Returns
     -------
-    tide: np.ndarray
+    tide: xarray.DataArray
         tidal elevation in meters
     """
 
@@ -346,50 +332,43 @@ def tide_elevations(
     if DEFINITION_FILE is not None:
         model = pyTMD.io.model(DIRECTORY).from_file(DEFINITION_FILE)
     else:
-        model = pyTMD.io.model(DIRECTORY, compressed=GZIP).elevation(MODEL)
+        model = pyTMD.io.model(DIRECTORY).from_database(MODEL)
+    # open dataset
+    ds = model.open_dataset(type='z')
+    # append_node=APPEND_NODE, apply_flexure=APPLY_FLEXURE
+    # subset to constituents
+    if CONSTITUENTS:
+        ds = ds.subset(CONSTITUENTS)
 
     # determine input data type based on variable dimensions
     if not TYPE:
         TYPE = pyTMD.spatial.data_type(x, y, delta_time)
     assert TYPE.lower() in ('grid', 'drift', 'time series')
-    # reform coordinate dimensions for input grids
-    # or verify coordinate dimension shapes
-    if (TYPE.lower() == 'grid') and (np.size(x) != np.size(y)):
[…]

[…]
