### 2025-11-16 708711b
use to_ellipose and from_ellipse
add back crop functionality
