### 2025-11-16 2090f81
use to_ellipose and from_ellipse
add back crop functionality
