+++ pyTMD/io/dataset.py
@@ -57,79 +57,16 @@ class DataTree:
         # initialize DataTree
         self._dtree = dtree
 
-    def ellipse(self, **kwargs):
+    def crop(self, *args, **kwargs):
         """
-        Expresses currents in terms of four ellipse parameters
-
-        - ``major``: amplitude of the semi-major axis
-        - ``minor``: amplitude of the semi-minor axis
-        - ``incl``: angle of inclination of the northern semi-major axis
-        - ``phase``: phase lag of the maximum current behind the maximum tidal potential  
+        Crop ``DataTree`` to input bounding box
         """
-        # set the direction of the transformation
-        kwargs.setdefault('direction', 'FORWARD')
-        # allocate for output datatree
-        dtree = xr.DataTree()
-        # perform transformation (forward or backward)
-        if (kwargs['direction'].lower() == 'forward'):
-            # get u and v components from datatree
-            dsu = (self._dtree.get('u',None) or self._dtree.get('U',None)).to_dataset()
-            dsv = (self._dtree.get('v',None) or self._dtree.get('V',None)).to_dataset()
-            # calculate ellipse parameters for each constituent
-            dmajor = xr.Dataset()
-            dminor = xr.Dataset()
-            dincl = xr.Dataset()
-            dphase = xr.Dataset()
-            # for each constituent in the u-component
-            for c in dsu.tmd.constituents:
-                # assert units between datasets are the same
-                if (dsu[c].attrs.get('units','') != dsv[c].attrs.get('units','')):
-                    raise ValueError(f'Incompatible units for {c} in u and v datasets')
-                # calculate ellipse parameters
-                major, minor, incl, phase = pyTMD.ellipse.ellipse(
-                    dsu[c].values, dsv[c].values)
-                # create xarray DataArray for ellipse parameters
-                dmajor[c] = xr.DataArray(major, dims=dsu[c].dims, coords=dsu.coords)
-                dminor[c] = xr.DataArray(minor, dims=dsu[c].dims, coords=dsu.coords)
-                dincl[c] = xr.DataArray(incl, dims=dsu[c].dims, coords=dsu.coords)
-                dphase[c] = xr.DataArray(phase, dims=dsu[c].dims, coords=dsu.coords)
-                # add attributes to each variable
-                dmajor[c].attrs['units'] = dsu[c].attrs.get('units', '')
-                dminor[c].attrs['units'] = dsu[c].attrs.get('units', '')
-                dincl[c].attrs['units'] = 'degrees'
-                dphase[c].attrs['units'] = 'degrees'
-            # add datasets to output datatree
-            dtree['major'] = dmajor
-            dtree['minor'] = dminor
-            dtree['incl'] = dincl
-            dtree['phase'] = dphase
-        if (kwargs['direction'].lower() == 'backward'):
-            # get ellipse parameters from datatree
-            dmajor = self._dtree['major'].to_dataset()
-            dminor = self._dtree['minor'].to_dataset()
-            dincl = self._dtree['incl'].to_dataset()
-            dphase = self._dtree['phase'].to_dataset()
-            # calculate currents for each constituent
-            dsu = xr.Dataset()
-            dsv = xr.Dataset()
-            # for each constituent in the major parameter
-            for c in dmajor.tmd.constituents:
-                # calculate ellipse parameters
-                u, v = pyTMD.ellipse.inverse(dmajor[c].values, dminor[c].values,
-                    dincl[c].values, dphase[c].values)
-                # create xarray DataArray for ellipse parameters
[…]
