### 2025-11-17 52d5565
test: parametrize crop
test: add back extend array checks
