### 2025-11-17 c4f1a8e
go back to no scales in databases
