### 2025-11-17 841fc6f
test: check new z, u, v database and JSON format
