+++ test/test_model.py
@@ -1,8 +1,9 @@
 """
-test_model.py (07/2025)
+test_model.py (11/2025)
 Tests the reading of model definition files
 
 UPDATE HISTORY:
+    Updated 11/2025: use new z, u, v database and JSON format
     Updated 07/2025: added GOT4.10_SAL subset of constituents
     Updated 06/2025: added function to check extra databases
     Updated 02/2025: added function to try to parse bathymetry files
@@ -36,22 +37,12 @@ def test_definition_CATS2008():
     # test read variables
     assert m.format == 'OTIS'
     assert m.name == 'CATS2008'
-    assert m.model_file == pathlib.Path('CATS2008/hf.CATS2008.out')
-    assert m.grid_file == pathlib.Path('CATS2008/grid_CATS2008')
+    assert m.z.grid_file == pathlib.Path('CATS2008/grid_CATS2008')
+    assert m.z.model_file == pathlib.Path('CATS2008/hf.CATS2008.out')
+    assert m.z.variable == 'tide_ocean'
     assert m.projection == {'datum': 'WGS84', 'lat_0': -90, 'lat_ts': -71,
         'lon_0': -70, 'proj': 'stere', 'type': 'crs', 'units': 'km',
         'x_0': 0, 'y_0': 0}
-    assert m.type == 'z'
-    assert m.variable == 'tide_ocean'
-    # test properties
-    assert m.atl03 == 'tide_ocean'
-    assert m.atl06 == 'tide_ocean'
-    assert m.atl07 == 'height_segment_ocean'
-    assert m.atl10 == 'height_segment_ocean'
-    assert m.atl11 == 'tide_ocean'
-    assert m.atl12 == 'tide_ocean_seg'
-    assert m.gla12 == 'd_ocElv'
-    assert m.long_name == 'ocean_tide_elevation'
 
 def test_definition_FES():
     """Tests the reading of the FES2014 model definition file
@@ -59,6 +50,7 @@ def test_definition_FES():
     # read definition file
     definition_file = 'model_FES2014.json'
     m = pyTMD.io.model().from_file(filepath.joinpath(definition_file))
+    m.parse_constituents(type='z')
     # model files and constituents
     model_files = ['fes2014/ocean_tide/2n2.nc.gz',
         'fes2014/ocean_tide/eps2.nc.gz', 'fes2014/ocean_tide/j1.nc.gz',
@@ -87,25 +79,13 @@ def test_definition_FES():
     assert m.name == 'FES2014'
     # assert that all model files are in the model definition
     for f in model_files:
-        assert pathlib.Path(f) in m.model_file
-    # assert that all constituents are in the model definition
-    assert m.constituents == constituents
-    assert m.type == 'z'
-    assert m.scale == 1.0/100.0
-    assert m.variable == 'tide_ocean'
+        assert pathlib.Path(f) in m['z'].model_file
+    assert m.z.units == 'cm'
+    assert m.z.variable == 'tide_ocean'
     assert m.compressed is True
     # check validity of parsed constituents
     parsed_constituents = [pyTMD.io.model.parse_file(f) for f in model_files]
     assert parsed_constituents == constituents
-    # test derived properties
-    assert m.atl03 == 'tide_ocean'
-    assert m.atl06 == 'tide_ocean'
-    assert m.atl07 == 'height_segment_ocean'
-    assert m.atl10 == 'height_segment_ocean'
-    assert m.atl11 == 'tide_ocean'
-    assert m.atl12 == 'tide_ocean_seg'
-    assert m.gla12 == 'd_ocElv'
-    assert m.long_name == 'ocean_tide_elevation'
 
 # PURPOSE: test glob file functionality
 def test_definition_FES_glob():
[…]
