### 2025-11-17 4efd5a5
refactor: replace io with xio functions
