+++ pyTMD/io/GOT.py
@@ -1,47 +1,20 @@
 #!/usr/bin/env python
 u"""
 GOT.py
-Written by Tyler Sutterley (10/2025)
+Written by Tyler Sutterley (11/2025)
 
-Reads files for Richard Ray's Goddard Ocean Tide (GOT) models and makes
-    initial calculations to run the tide program
-Includes functions to extract tidal harmonic constants out of a tidal
-    model for given locations
-
-INPUTS:
-    ilon: longitude to interpolate
-    ilat: latitude to interpolate
-    model_files: list of model files for each constituent
-
-OPTIONS:
-    method: interpolation method
-        bilinear: quick bilinear interpolation
-        spline: scipy bivariate spline interpolation
-        linear, nearest: scipy regular grid interpolations
-    extrapolate: extrapolate model using nearest-neighbors
-    cutoff: extrapolation cutoff in kilometers
-        set to np.inf to extrapolate for all points
-    compressed: input files are gzip compressed
-    scale: scaling factor for converting to output units
-
-OUTPUTS:
-    amplitude: amplitudes of tidal constituents
-    phase: phases of tidal constituents
-    constituents: list of model constituents
+Reads ascii and netCDF4 files from Richard Ray's Goddard Ocean Tide (GOT) model
+    https://earth.gsfc.nasa.gov/geo/data/ocean-tide-models
 
 PYTHON DEPENDENCIES:
     numpy: Scientific Computing Tools For Python
         https://numpy.org
         https://numpy.org/doc/stable/user/numpy-for-matlab-users.html
-    scipy: Scientific Tools for Python
-        https://docs.scipy.org/doc/
-    netCDF4: Python interface to the netCDF C library
-        https://unidata.github.io/netcdf4-python/netCDF4/index.html
-
-PROGRAM DEPENDENCIES:
-    interpolate.py: interpolation routines for spatial data
+    xarray: N-D labeled arrays and datasets in Python
+        https://docs.xarray.dev/en/stable/
 
 UPDATE HISTORY:
+    Updated 11/2025: near-complete rewrite of program to use xarray
     Updated 10/2025: simplify ascii read function to use masked_equal
     Updated 08/2025: use numpy degree to radian conversions
         added option to gap fill when reading constituent grids
@@ -89,491 +62,113 @@ UPDATE HISTORY:
 from __future__ import division, annotations
 
 import re
-import copy
 import gzip
-import uuid
-import logging
 import pathlib
 import datetime
-import warnings
 import numpy as np
+import xarray as xr
 import pyTMD.version
-import pyTMD.interpolate
 import pyTMD.io.constituents
-from pyTMD.utilities import import_dependency
-
-# attempt imports
-netCDF4 = import_dependency('netCDF4')
 
 __all__ = [
-    "extract_constants",
-    "read_constants",
-    "interpolate_constants",
-    "read_ascii_file",
-    "read_netcdf_file",
-    "output_netcdf_file",
-    "_extend_array",
-    "_extend_matrix",
-    "_crop",
-    "_shift"
+    'open_mfdataset',
+    'open_got_dataset',
+    'open_got_ascii',
+    'open_got_netcdf',
+    'GOTDataset',
 ]
 
-# PURPOSE: extract harmonic constants from tide models at coordinates
-def extract_constants(
-        ilon: np.ndarray,
-        ilat: np.ndarray,
-        model_files: str | pathlib.Path | list | None = None,
-        **kwargs
+# PURPOSE: read a list of GOT ASCII or netCDF4 files
+def open_mfdataset(
+        model_files: list[str] | list[pathlib.Path],
+        **kwargs,
     ):
     """
-    Reads files for Richard Ray's Goddard Ocean Tide (GOT) models
-
[…]
