+++ pyTMD/xio/__init__.py
@@ -1,8 +1,0 @@
-"""
-Experimental xarray-powered input/output functions
-"""
-from .ATLAS import *
-from .FES import *
-from .GOT import *
-from .OTIS import *
-from .IERS import *

[…]
