### 2025-11-17 852ec97
docs: update xarray recipes
