+++ doc/source/user_guide/zarr-to-s3-obstore.py
@@ -6,9 +6,11 @@ import obstore
 s3_bucket = 'pytmd-scratch'
 tide_model = 'FES2022'
 # setup tide model
-m = pyTMD.io.model(compressed=False)
-# read tide model and convert to xarray DataTree
-dtree = m.to_datatree(tide_model, gap_fill=True, N=100)
+m = pyTMD.io.model(verify=False).from_database(tide_model)
+# read tide model as an xarray DataTree
+dtree = m.open_datatree(tide_model)
+# gap-fill missing data with inpainting
+dtree = dtree.tmd.inpaint(N=100)
 
 # setup zarr store using obstore
 presigned_s3_url = f's3://{s3_bucket}/{m.name}.zarr'
