+++ doc/source/user_guide/setup-s3fs.py
@@ -6,7 +6,7 @@ import s3fs
 s3_bucket = 'pytmd-scratch'
 tide_model = 'FES2022'
 # setup tide model
-m = pyTMD.io.model().elevation(tide_model)
+m = pyTMD.io.model().from_database(tide_model)
 
 # setup zarr store using s3fs
 fs = s3fs.S3FileSystem(anon=False, asynchronous=True)
