+++ pyTMD/datasets/reduce_otis.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 reduce_otis.py
-Written by Tyler Sutterley (10/2025)
+Written by Tyler Sutterley (11/2025)
 Read OTIS-format tidal files and reduce to a regional subset
 
 COMMAND LINE OPTIONS:
@@ -30,6 +30,7 @@ PROGRAM DEPENDENCIES:
     crs.py: Coordinate Reference System (CRS) routines
 
 UPDATE HISTORY:
+    Updated 11/2025: use new xarray file access protocols for OTIS files
     Updated 10/2025: change default directory for tide models to cache
     Updated 09/2025: renamed module and function to reduce_otis
         made a callable function and added function docstrings
@@ -66,14 +67,11 @@ import logging
 import pathlib
 import argparse
 import traceback
+import xarray as xr
 import numpy as np
-import pyTMD.crs
 import pyTMD.io
 import pyTMD.utilities
 import timescale.time
-
-# attempt imports
-pyproj = pyTMD.utilities.import_dependency('pyproj')
 
 # default data directory for tide models
 _default_directory = pyTMD.utilities.get_cache_path()
@@ -111,10 +109,7 @@ def reduce_otis(MODEL: str,
         Permission mode of the output files
     """
     # get parameters for tide model grid
-    try:
-        model = pyTMD.io.model(directory=directory).elevation(MODEL)
-    except Exception as exc:
-        model = pyTMD.io.model(directory=directory).current(MODEL)
+    m = pyTMD.io.model(directory=directory).from_database(MODEL)
     # directionaries with input and output files
     model_file = {}
     new_model_file = {}
@@ -122,110 +117,42 @@ def reduce_otis(MODEL: str,
     # read the OTIS-format tide grid file
     if (model.format == 'ATLAS-compact'):
         # if reading a global solution with localized solutions
-        x0,y0,hz0,mz0,iob,dt,pmask,local = pyTMD.io.OTIS.read_atlas_grid(
-            model.grid_file)
-        xi,yi,hz = pyTMD.io.OTIS.combine_atlas_model(x0, y0, hz0, pmask,
-            local, variable='depth')
-        mz = pyTMD.io.OTIS.create_atlas_mask(x0, y0, mz0,
-            local, variable='depth')
+        dsg, dtg = pyTMD.io.OTIS.open_atlas_grid(m['z'].grid_file)
+        dsz, dtz = pyTMD.io.OTIS.open_atlas_elevation(m['z'].model_file)
+        dsu, dsv, dtu, dtv = pyTMD.io.OTIS.open_atlas_transport(m['u'].model_file)
+        # combine local solutions with global solution
+        dsg = dsg.compact.combine_local(dtg)
+        dsz = dsz.compact.combine_local(dtz)
+        dsu = dsu.compact.combine_local(dtu)
+        dsv = dsv.compact.combine_local(dtv)
     else:
         # if reading a pure global solution
-        xi,yi,hz,mz,iob,dt = pyTMD.io.OTIS.read_otis_grid(model.grid_file)
+        dsg = pyTMD.io.OTIS.open_otis_grid(m['z'].grid_file)
+        dsz = pyTMD.io.OTIS.open_otis_grid(m['z'].model_file)
+        dsu, dsv = pyTMD.io.OTIS.open_otis_transport(m['u'].model_file)
 
-    # converting bounds x,y from projection to latitude/longitude
-    crs1 = pyTMD.crs().from_input(projection)
-    crs2 = pyproj.CRS.from_epsg(4326)
-    transformer = pyproj.Transformer.from_crs(crs1, crs2, always_xy=True)
-    xbox = np.array([bounds[0], bounds[1], bounds[1], bounds[0], bounds[0]])
[…]

[…]
