### 2025-11-17 cf56b09
Update test_equilibrium_tide.py
