### 2025-11-17 7c04f87
docs: update io functions for xarray
chore: update pixi lock
