+++ doc/source/model_table.py
@@ -14,12 +14,15 @@ models_table = directory.joinpath('_assets', f'{model_type}-models.csv')
 fid = models_table.open(mode='w', encoding='utf8')
 # write to csv
 fid.write('Model,Directory\n')
-for model,parameters in models[model_type].items():
+for model,parameters in models.items():
+    # skip current models
+    if 'z' not in parameters:
+        continue
     # extract the model directory
-    if isinstance(parameters['model_file'], str):
-        model_directory = pathlib.Path(parameters['model_file']).parent
-    elif isinstance(parameters['model_file'], list):
-        model_directory = pathlib.Path(parameters['model_file'][0]).parent
+    if isinstance(parameters['z']['model_file'], str):
+        model_directory = pathlib.Path(parameters['z']['model_file']).parent
+    elif isinstance(parameters['z']['model_file'], list):
+        model_directory = pathlib.Path(parameters['z']['model_file'][0]).parent
     # extract the reference
     reference = parameters.get('reference', None)
     # write the model and directory to the csv file
@@ -33,15 +36,20 @@ models_table = directory.joinpath('_assets', f'{model_type}-models.csv')
 fid = models_table.open(mode='w', encoding='utf8')
 # write to csv
 fid.write('Model,U-Directory,V-Directory\n')
-for model,parameters in models[model_type].items():
+for model,parameters in models.items():
+    # skip elevation models
+    if 'u' not in parameters:
+        continue
     # extract the model directory
     model_directories = []
-    for t in parameters['model_file'].keys():
-        if isinstance(parameters['model_file'][t], str):
-            d = pathlib.Path(parameters['model_file'][t]).parent
-        elif isinstance(parameters['model_file'][t], list):
-            d = pathlib.Path(parameters['model_file'][t][0]).parent
+    for t in ('u', 'v'):
+        if isinstance(parameters[t]['model_file'], str):
+            d = pathlib.Path(parameters[t]['model_file']).parent
+        elif isinstance(parameters[t]['model_file'], list):
+            d = pathlib.Path(parameters[t]['model_file'][0]).parent
         model_directories.append(f'``<path_to_tide_models>/{d}``')
+    # only keep unique directories
+    model_directories = list(set(model_directories))
     # join the directories
     model_directory = ','.join(model_directories)
     # extract the reference
