### 2025-11-17 2797bef
reduce to only dataset predict (now time_series)
