+++ pyTMD/ellipse.py
@@ -46,7 +46,7 @@ __all__ = [
 
 def ellipse(u: np.ndarray, v: np.ndarray):
     """
-    Expresses the amplitudes and phases for the u and v components in terms of
+    Expresses the amplitudes and phases for the `u` and `v` components in terms of
     four ellipse parameters using Foreman's formula :cite:p:`Foreman:1989dt`
 
     Parameters
@@ -99,7 +99,7 @@ def inverse(
         phase: np.ndarray
     ):
     """
-    Calculates currents u, v using the four tidal ellipse
+    Calculates currents `u`, `v` using the four tidal ellipse
     parameters from Foreman's formula :cite:p:`Foreman:1989dt`
 
     Parameters
