### 2025-11-18 236a128
refactor: update compute for other tide functionals
