+++ pyTMD/math.py
@@ -24,6 +24,7 @@ UPDATE HISTORY:
 from __future__ import annotations
 
 import numpy as np
+import xarray as xr
 from scipy.special import factorial
 
 __all__ = [
@@ -168,7 +169,6 @@ def legendre(
         l: int,
         x: np.ndarray,
         m: int = 0,
-        deriv: bool = False
     ):
     """
     Computes associated Legendre functions and their first-derivatives
@@ -185,70 +185,76 @@ def legendre(
         Typically ``cos(theta)``, where ``theta`` is the colatitude in radians
     m: int, default 0
         order of the Legendre polynomials (0 to ``l``)
-    deriv: bool, default False
-        return the first derivative with respect to ``theta``
 
     Returns
     -------
     Plm: np.ndarray
         Legendre polynomials of degree ``l`` and order ``m``
+    dPlm: np.ndarray
+        first derivative of Legendre polynomials with respect to ``theta``
     """
+    # maximum degree and order
+    lmax = 4
     # verify values are integers
     l = np.int64(l)
     m = np.int64(m)
     # assert values
-    assert (l >= 0) and (l <= 4), 'Degree must be between 0 and 4'
+    assert (l >= 0) and (l <= lmax), f'Degree must be between 0 and {lmax}'
     assert (m >= 0) and (m <= l), 'Order must be between 0 and l'
-    # verify dimensions
-    singular_values = (np.ndim(x) == 0)
-    x = np.atleast_1d(x).flatten()
     # if x is the cos of colatitude, u is the sine
     u = np.sqrt(1.0 - x**2)
-    # size of the x array
-    nx = len(x)
-    # complete matrix of associated legendre functions
-    # up to degree and order 4
-    Plm = np.zeros((5, 5, nx), dtype=np.float64)
     # since tides only use low-degree harmonics:
     # functions are hard coded rather than using a recursion relation
-    if deriv:
-        # calculate first derivatives with respect to theta
-        Plm[1, 0, :] = -u
-        Plm[1, 1, :] = x
-        Plm[2, 0, :] = -3.0*u*x
-        Plm[2, 1, :] = 3.0*(1.0 - 2.0*u**2)
-        Plm[2, 2, :] = 6.0*u*x
-        Plm[3, 0, :] = u*(1.5 - 7.5*x**2)
-        Plm[3, 1, :] = -1.5*x*(10.0*u**2 - 5.0*x**2 + 1.0)
-        Plm[3, 2, :] = 15.0*u*(3.0*x**2 - 1.0)
-        Plm[3, 3, :] = 45.0*x*u**2
-        Plm[4, 0, :] = -2.5*(7.0*x**2 - 3.0)*u*x
-        Plm[4, 1, :] = 2.5*(28.0*x**4 - 27.0*x**2 + 3.0)
-        Plm[4, 2, :] = (105*x**2 - 105*u**2 - 15.0)*u*x
-        Plm[4, 3, :] = (420.0*x**3 - 105.0)*u**2
-        Plm[4, 4, :] = 420.0*x*u**3
-    else:
-        # calculate Legendre polynomials
-        Plm[0, 0, :] = 1.0
-        Plm[1, 0, :] = x
-        Plm[1, 1, :] = u
-        Plm[2, 0, :] = 0.5*(3.0*x**2 - 1.0)
-        Plm[2, 1, :] = 3.0*x*u
-        Plm[2, 2, :] = 3.0*u**2
-        Plm[3, 0, :] = 0.5*(5.0*x**2 - 3.0)*x
[…]
