+++ test/test_pole_tide.py
@@ -21,6 +21,7 @@ import inspect
 import pathlib
 import pytest
 import pyproj
+import xarray as xr
 import numpy as np
 import scipy.interpolate
 import pyTMD.compute
@@ -70,7 +71,7 @@ def test_load_pole_tide_displacements(TYPE):
     CONVENTION = '2018'
     FILL_VALUE = np.nan
 
-    # compute load pole tides using cartesian functions
+    # calculate load pole tide displacements
     test = pyTMD.compute.LPT_displacements(x, y, delta_time,
         EPSG=EPSG, EPOCH=EPOCH, TYPE=TYPE, TIME=TIME,
         ELLIPSOID=ELLIPSOID, CONVENTION=CONVENTION,
@@ -191,10 +192,8 @@ def test_load_pole_tide_displacements(TYPE):
     assert np.all(np.abs(Srad - test) < eps)
     assert np.all(np.abs(approx - test) < eps)
 
-# parameterize interpolation method
-@pytest.mark.parametrize("METHOD", ['spline','nearest','linear'])
 # test the interpolation of ocean pole tide values
-def test_read_ocean_pole(METHOD):
+def test_read_ocean_pole():
     # read ocean pole tide test file for header text
     ocean_pole_test_file = filepath.joinpath('opoleloadcmcor.test')
     with ocean_pole_test_file.open(mode='r', encoding='utf8') as f:
@@ -214,47 +213,21 @@ def test_read_ocean_pole(METHOD):
     header['latitude'] = np.float64(lat)
     # convert longitude from 0:360 to -180:180
     header['longitude'] = np.float64(lon) - 360.0
-
-    # number of points
-    npts = len(np.atleast_1d(header['longitude']))
+    # valid ocean pole tide coefficients
+    U_valid = {}
+    U_valid['R'] = 0.216192 + 0.285652j
+    U_valid['N'] = 0.012697 + 0.024930j
+    U_valid['E'] = 0.000868 + 0.010389j
     # read ocean pole tide map from Desai (2002)
-    Umap = {}
-    Umap['R'], Umap['N'], Umap['E'], ilon, ilat = \
-        pyTMD.io.IERS.read_binary_file()
-    # interpolate ocean pole tide map from Desai (2002)
-    Uint = {}
-    if (METHOD == 'spline'):
-        # use scipy bivariate splines to interpolate to output points
-        for key,val in Umap.items():
-            Uint[key] = np.zeros((npts), dtype=np.clongdouble)
-            f1 = scipy.interpolate.RectBivariateSpline(ilon, ilat[::-1],
-                val[:,::-1].real, kx=1, ky=1)
-            f2 = scipy.interpolate.RectBivariateSpline(ilon, ilat[::-1],
-                val[:,::-1].imag, kx=1, ky=1)
-            Uint[key].real = f1.ev(header['longitude'], header['latitude'])
-            Uint[key].imag = f2.ev(header['longitude'], header['latitude'])
-    else:
-        # use scipy regular grid to interpolate values for a given method
-        for key,val in Umap.items():
-            Uint[key] = np.zeros((npts), dtype=np.clongdouble)
-            r1 = scipy.interpolate.RegularGridInterpolator((ilon,ilat[::-1]),
-                val[:,::-1], bounds_error=False, method=METHOD)
-            points = np.c_[header['longitude'], header['latitude']]
-            Uint[key][:] = r1.__call__(points)
-
+    Umap = pyTMD.io.IERS.open_dataset()
[…]
