### 2025-11-20 d9b1854
refactor: xarray summations over constituents
