+++ pyTMD/predict.py
@@ -16,8 +16,8 @@ PYTHON DEPENDENCIES:
         https://pypi.org/project/timescale/
 
 PROGRAM DEPENDENCIES:
-    arguments.py: loads nodal corrections for tidal constituents
     astro.py: computes the basic astronomical mean longitudes
+    constituents.py: calculates constituent parameters and nodal arguments  
     spatial.py: utilities for working with geospatial data
 
 UPDATE HISTORY:
@@ -88,8 +88,8 @@ from __future__ import annotations
 
 import logging
 import numpy as np
-import pyTMD.arguments
 import pyTMD.astro
+import pyTMD.constituents
 import pyTMD.math
 import pyTMD.interpolate
 import pyTMD.spatial
@@ -154,37 +154,39 @@ def time_series(
     # list of constituents
     constituents = ds.tmd.constituents
     # load the nodal corrections
-    pu, pf, G = pyTMD.arguments.arguments(MJD,
+    pu, pf, G = pyTMD.constituents.arguments(MJD,
         constituents, **kwargs)
+    # calculate constituent phase angles
+    if kwargs['corrections'] in ('OTIS', 'ATLAS', 'TMD3'):
+        theta = np.zeros_like(pu)
+        for i, c in enumerate(constituents):
+            # load parameters for constituent
+            amp, ph, omega, alpha, species = \
+                pyTMD.constituents._constituent_parameters(c)
+            # phase angle from frequency and phase-0
+            theta[:,i] = omega*t*86400.0 + ph + pu[:,i]
+    else:
+        # phase angle from arguments
+        theta = np.radians(G) + pu
     # dataset of arguments
     arguments = xr.Dataset(
         data_vars=dict(
             u=(['time','constituent'], pu),
             f=(['time','constituent'], pf),
-            G=(['time','constituent'], G)),
+            G=(['time','constituent'], G),
+            theta=(['time','constituent'], np.exp(1j*theta))),
         coords=dict(time=np.atleast_1d(MJD), constituent=constituents),
     )
-    # sum over all tides
-    for i, c in enumerate(constituents):
-        # select argument for constituent
-        arg = arguments.sel(constituent=c)
-        if kwargs['corrections'] in ('OTIS', 'ATLAS', 'TMD3'):
-            # load parameters for constituent
-            amp, ph, omega, alpha, species = \
-                pyTMD.arguments._constituent_parameters(c)
-            th = omega*t*86400.0 + ph + arg.u
-        else:
-            th = np.radians(arg.G) + arg.u
-        # initialize or sum tidal elevations
-        if i == 0:
-            darr = arg.f*ds[c].real*np.cos(th) - \
-                arg.f*ds[c].imag*np.sin(th)
-            darr.attrs['units'] = ds[c].attrs.get('units', None)
-        else:
-            darr += arg.f*ds[c].real*np.cos(th) - \
-                arg.f*ds[c].imag*np.sin(th)
-    # return the predicted tidal elevations
-    return darr.drop_vars('constituent')
+    # convert Dataset to DataArray of complex tidal harmonics
+    darr = ds.tmd.to_dataarray(constituents=constituents)
+    # sum over tidal constituents
+    tpred = (arguments.f*darr.real*arguments.theta.real - 
+        arguments.f*darr.imag*arguments.theta.imag).sum(
+        dim='constituent', skipna=False)
+    # copy units attribute
+    tpred.attrs['units'] = ds[constituents[0]].attrs.get('units', None)
+    # return the predicted tides
+    return tpred
 
 # PURPOSE: infer the minor corrections from the major constituents
 def infer_minor(
@@ -227,18 +229,18 @@ def infer_minor(
     # list of minor constituents
[…]
