### 2025-11-20 0655713
docs: update sph harm plots
docs: update solid earth notebook
