+++ doc/source/background/sphharm34.py
@@ -1,5 +1,6 @@
 import numpy as np
 import pyTMD.math
+import xarray as xr
 import cartopy.crs as ccrs
 import matplotlib.pyplot as plt
 
@@ -9,8 +10,8 @@ lat = np.arange(-90 + dlat/2.0, 90 + dlat/2.0, dlat)
 lon = np.arange(0 + dlon/2.0, 360 + dlon/2.0, dlon)
 gridlon, gridlat = np.meshgrid(lon, lat)
 # colatitude and longitude in radians
-theta = np.radians(90.0 - gridlat)
-phi = np.radians(gridlon)
+theta = xr.DataArray(np.radians(90.0 - gridlat), dims=('y','x'))
+phi = xr.DataArray(np.radians(gridlon), dims=('y','x'))
 
 # minimum and maximum degree of spherical harmonics
 lmin, lmax = (3, 4)
@@ -29,7 +30,6 @@ for n, l in enumerate(range(lmin, lmax+1)):
         ax = fig.add_subplot(nrows, ncols, i, projection=projection)
         # calculate spherical harmonics (and derivatives w.r.t. colatitude)
         Y_lm, dY_lm = pyTMD.math.sph_harm(l, theta, phi, m=tau)
-        Y_lm = Y_lm.reshape(theta.shape)
         # plot the surface
         ax.pcolormesh(lon, lat, Y_lm.real,
             transform=ccrs.PlateCarree(),
