+++ doc/source/background/sphharm.py
@@ -1,5 +1,6 @@
 import numpy as np
 import pyTMD.math
+import xarray as xr
 import cartopy.crs as ccrs
 import matplotlib.pyplot as plt
 
@@ -9,8 +10,8 @@ lat = np.arange(-90 + dlat/2.0, 90 + dlat/2.0, dlat)
 lon = np.arange(0 + dlon/2.0, 360 + dlon/2.0, dlon)
 gridlon, gridlat = np.meshgrid(lon, lat)
 # colatitude and longitude in radians
-theta = np.radians(90.0 - gridlat)
-phi = np.radians(gridlon)
+theta = xr.DataArray(np.radians(90.0 - gridlat), dims=('y','x'))
+phi = xr.DataArray(np.radians(gridlon), dims=('y','x'))
 
 # spherical harmonic degree
 l = 2
@@ -30,7 +31,6 @@ for tau in range(0, l+1):
     ax2 = fig.add_subplot(nrows, ncols, ncols+tau+1, projection=proj2)
     # calculate spherical harmonics (and derivatives w.r.t. colatitude)
     Y_lm, dY_lm = pyTMD.math.sph_harm(l, theta, phi, m=tau)
-    Y_lm = Y_lm.reshape(theta.shape)
     # set the title
     ax1.set_title(f'{species[tau]}\n$l={l}, m={tau}$')
     # add projection text
