+++ pyTMD/interpolate.py
@@ -29,17 +29,12 @@ from __future__ import annotations
 import numpy as np
 import scipy.fftpack
 import scipy.spatial
-import scipy.interpolate
 import pyTMD.spatial
 
 __all__ = [
     "interp1d",
     "inpaint",
-    "bilinear",
-    "spline",
-    "regulargrid",
     "extrapolate",
-    "_distance"
 ]
 
 # PURPOSE: 1-dimensional linear interpolation on arrays
@@ -180,261 +175,6 @@ def inpaint(
     z0[W] = np.copy(zs[W])
     # return the inpainted grid
     return z0
-
-# PURPOSE: bilinear interpolation of input data to output data
-def bilinear(
-        xs: np.ndarray,
-        ys: np.ndarray,
-        zs: np.ndarray,
-        X: np.ndarray,
-        Y: np.ndarray,
-        fill_value: float = np.nan,
-        extrapolate: bool = False,
-        **kwargs
-    ):
-    """
-    Bilinear interpolation of input data to output coordinates
-
-    Parameters
-    ----------
-    xs: np.ndarray
-        x-coordinates of tidal model
-    ys: np.ndarray
-        y-coordinates of tidal model
-    zs: np.ndarray
-        tide model data
-    X: np.ndarray
-        output x-coordinates
-    Y: np.ndarray
-        output y-coordinates
-    fill_value: float, default np.nan
-        invalid value
-    extrapolate: bool, default False
-        allow linear extrapolation of points
-    dtype: np.dtype, default np.float64
-        output data type
-
-    Returns
-    -------
-    data: np.ndarray
-        interpolated data
-    """
-    # set default data type
-    dtype = kwargs.get('dtype', zs.dtype)
-    # verify that input data is masked array
-    if not isinstance(zs, np.ma.MaskedArray):
-        zs = np.ma.array(zs)
-        zs.mask = np.zeros_like(zs, dtype=bool)
-    # interpolate gridded data values to data
-    npts = len(X)
-    # allocate to output interpolated data array
-    data = np.ma.zeros((npts), dtype=dtype, fill_value=fill_value)
-    data.mask = np.ones((npts), dtype=bool)
-    # initially set all data to fill value
-    data.data[:] = data.fill_value
-    # for each point
-    for i in range(npts):
-        # calculating the indices for the original grid
-        ix = np.searchsorted(xs, X[i]) - 1
-        iy = np.searchsorted(ys, Y[i]) - 1
-        # check that all points are within valid bounds
-        bounds = (ix >= 0) & (iy >= 0) & (ix < len(xs)) & (iy < len(ys))
-        if not (extrapolate or bounds):
-            continue
-        # clip to handle extrapolation
-        ix = np.clip(ix, a_min=0, a_max=len(xs) - 2)
-        iy = np.clip(iy, a_min=0, a_max=len(ys) - 2)
-        # corner data values for adjacent grid cells
-        IM = np.ma.zeros((4), fill_value=fill_value, dtype=dtype)
-        IM.mask = np.ones((4), dtype=bool)
-        # corner weight values for adjacent grid cells
-        WM = np.zeros((4))
-        # build data and weight arrays
-        for j,XI,YI in zip([0,1,2,3],[ix,ix+1,ix,ix+1],[iy,iy,iy+1,iy+1]):
-            IM.data[j] = zs.data[YI,XI].astype(dtype)
-            IM.mask[j] = zs.mask[YI,XI]
-            WM[3-j] = np.abs(X[i]-xs[XI])*np.abs(Y[i]-ys[YI])
-        # if on corner value: use exact
-        if (np.isclose(Y[i],ys[iy]) & np.isclose(X[i],xs[ix])):
[…]
