+++ pyTMD/compute.py
@@ -40,8 +40,8 @@ PYTHON DEPENDENCIES:
         https://numpy.org/doc/stable/user/numpy-for-matlab-users.html
     scipy: Scientific Tools for Python
         https://docs.scipy.org/doc/
-    netCDF4: Python interface to the netCDF C library
-        https://unidata.github.io/netcdf4-python/netCDF4/index.html
+    h5netcdf: Pythonic interface to netCDF4 via h5py
+        https://h5netcdf.org/
     pyproj: Python interface to PROJ library
         https://pypi.org/project/pyproj/
     timescale: Python tools for time and astronomical calculations
@@ -171,16 +171,13 @@ __all__ = [
     "SET_displacements"
 ]
 
-# number of days between the Julian day epoch and MJD
-_jd_mjd = 2400000.5
-
 # default working data directory for tide models
 _default_directory = pyTMD.utilities.get_cache_path()
 
 # PURPOSE: wrapper function for computing values
 def corrections(
         x: np.ndarray, y: np.ndarray, delta_time: np.ndarray,
-        CORRECTION: str = 'ocean',
+        correction: str = 'ocean',
         **kwargs
     ):
     """
@@ -194,7 +191,7 @@ def corrections(
         y-coordinates in projection EPSG
     delta_time: np.ndarray
         seconds since EPOCH or datetime array
-    CORRECTION: str, default 'ocean'
+    correction: str, default 'ocean'
         Correction type to compute
 
             - ``'ocean'``: ocean tide from model constituents
@@ -211,41 +208,35 @@ def corrections(
     values: np.ndarray
         tidal correction at coordinates and time in meters
     """
-    if CORRECTION.lower() in ('ocean', 'load'):
+    if correction.lower() in ('ocean', 'load'):
         return tide_elevations(x, y, delta_time, **kwargs)
-    elif (CORRECTION.upper() == 'LPET'):
+    elif (correction.upper() == 'LPET'):
         return LPET_elevations(x, y, delta_time, **kwargs)
-    elif (CORRECTION.upper() == 'LPT'):
+    elif (correction.upper() == 'LPT'):
         return LPT_displacements(x, y, delta_time, **kwargs)
-    elif (CORRECTION.upper() == 'OPT'):
+    elif (correction.upper() == 'OPT'):
         return OPT_displacements(x, y, delta_time, **kwargs)
-    elif (CORRECTION.upper() == 'SET'):
+    elif (correction.upper() == 'SET'):
         return SET_displacements(x, y, delta_time, **kwargs)
     else:
-        raise ValueError(f'Unrecognized correction type: {CORRECTION}')
+        raise ValueError(f'Unrecognized correction type: {correction}')
 
 # PURPOSE: compute tides at points and times using tide model algorithms
 def tide_elevations(
         x: np.ndarray, y: np.ndarray, delta_time: np.ndarray,
-        DIRECTORY: str | pathlib.Path | None = _default_directory,
-        MODEL: str | None = None,
-        DEFINITION_FILE: str | pathlib.Path | IOBase | None = None,
-        CROP: bool = False,
-        BOUNDS: list | np.ndarray | None = None,
-        BUFFER: int | float = 0,
-        EPSG: str | int = 4326,
-        EPOCH: list | tuple = (2000, 1, 1, 0, 0, 0),
-        TYPE: str | None = 'drift',
-        TIME: str = 'UTC',
-        METHOD: str = 'linear',
-        EXTRAPOLATE: bool = False,
-        CUTOFF: int | float = 10.0,
-        CORRECTIONS: str | None = None,
-        CONSTITUENTS: list | None = None,
-        INFER_MINOR: bool = True,
[…]
