### 2025-11-20 e44ec3b
test: ignore pyproj warnings
