+++ pyTMD/astro.py
@@ -293,7 +293,7 @@ def planetary_longitudes(MJD: np.ndarray):
     jupiter_longitude = np.array([34.351519, 3036.3027748, 2.233e-4, 3.7e-8])
     LJu = polynomial_sum(jupiter_longitude, T)
     # mean longitudes of Saturn
-    saturn_longitude = np.array([50.077444, 1223.5110686, 5.1908-4, -3.0e-8])
+    saturn_longitude = np.array([50.077444, 1223.5110686, 5.1908e-4, -3.0e-8])
     LSa = polynomial_sum(saturn_longitude, T)
     # take the modulus of each
     LMe = normalize_angle(LMe)
@@ -303,17 +303,6 @@ def planetary_longitudes(MJD: np.ndarray):
     LSa = normalize_angle(LSa)
     # return as tuple
     return (LMe, LVe, LMa, LJu, LSa)
-
-# PURPOSE: computes the phase angles of astronomical means
-def phase_angles(MJD: np.ndarray):
-    # raise warning for deprecated function call
-    warnings.warn(("Deprecated. Please use "
-        "pyTMD.astro.doodson_arguments instead"),
-        DeprecationWarning)
-    # call updated function to not break current workflows
-    TAU, S, H, P, ZNS, PS = doodson_arguments(MJD)
-    # return as tuple
-    return (S, H, P, TAU, ZNS, PS)
 
 # PURPOSE: computes the phase angles of astronomical means
 def doodson_arguments(
@@ -358,8 +347,9 @@ def doodson_arguments(
     hour = np.mod(MJD, 1)*24.0
     # calculate Doodson phase angles
     # mean longitude of moon (degrees)
-    S = polynomial_sum(np.array([218.3164477, 481267.88123421,
-        -1.5786e-3, 1.855835e-6, -1.53388e-8]), T)
+    lunar_longitude = np.array([218.3164477, 481267.88123421,
+        -1.5786e-3, 1.855835e-6, -1.53388e-8])
+    S = polynomial_sum(lunar_longitude, T)
     # mean lunar time (degrees)
     if equinox:
         # create timescale from Modified Julian Day (MJD)
@@ -368,27 +358,33 @@ def doodson_arguments(
         # Equinox method converted to degrees
         TAU = 360.0*ts.st + 180.0 - S
     else:
-        LAMBDA = polynomial_sum(np.array([280.4606184,
-            36000.7700536, 3.8793e-4, -2.58e-8]), T)
+        lambda_coefficients = np.array([280.4606184,
+            36000.7700536, 3.8793e-4, -2.58e-8])
+        LAMBDA = polynomial_sum(lambda_coefficients, T)
         TAU = (hour*15.0) - S + LAMBDA
     # calculate correction for mean lunar longitude (degrees)
     if apply_correction:
-        PR = polynomial_sum(np.array([0.0, 1.396971278,
-            3.08889e-4, 2.1e-8, 7.0e-9]), T)
+        lunar_correction = np.array([0.0, 1.396971278,
+            3.08889e-4, 2.1e-8, 7.0e-9])
+        PR = polynomial_sum(lunar_correction, T)
         S += PR
     # mean longitude of sun (degrees)
-    H = polynomial_sum(np.array([280.46645, 36000.7697489,
[…]
