### 2025-11-20 1d068ec
Update test_coordinates.py
