### 2025-11-20 9b7932e
docs: run readthedocs in dev env
