### 2025-11-20 6c25d28
update xarray predict for new workflow
