### 2025-11-21 4efa5fd
Update project description in README.md
Removed 'powered by `xarray`' from the project description.
