### 2025-11-21 e6883af
test: add check for interp1d arrays
