### 2025-11-21 2488d57
Update CITATION.cff
