+++ CITATION.cff
@@ -3,7 +3,7 @@
 
 cff-version: 1.2.0
 title: pyTMD
-message: Python-based tidal prediction software
+abstract: Python-based tidal prediction software
 type: software
 keywords:
   - Ocean Tides
@@ -12,9 +12,9 @@ keywords:
   - Solid Earth Tides
   - Tidal Prediction
 license: MIT
-version: "2.2.9"
-date-released: "2025-11-21"
-message: If you use this software, please consider citing our library.
+version: "2.2.8"
+date-released: "2025-09-05"
+message: If you use this software, please consider citing library.
 
 authors:
   - given-names: Tyler
@@ -38,14 +38,22 @@ authors:
   - given-names: Matthew
     family-names: Siegfried
     affiliation: Colorado School of Mines
+
 identifiers:
   - type: doi
     value: 10.5281/zenodo.5555395
-    description: Zenodo Archive
-repository-code: 'https://github.com/pyTMD/pyTMD'
-url: 'https://pytmd.readthedocs.io'
-repository: 'https://pypi.org/project/pyTMD'
-repository-artifact: 'https://anaconda.org/conda-forge/pytmd'
+    description: Persistent identifier for pyTMD
+  - type: url
+    value: https://github.com/pyTMD/pyTMD
+    description: GitHub repository for pyTMD
+  - type: url
+    value: https://pypi.org/project/pyTMD
+    description: pyTMD on PyPI
+  - type: url
+    value: https://anaconda.org/conda-forge/pytmd
+    description: pyTMD on conda-forge
+repository-code: "https://github.com/pyTMD/pyTMD"
+url: "https://pytmd.readthedocs.io"
 doi: "10.5281/zenodo.5555395"
 
 references:
