### 2025-11-21 64f524b
allow parallel reading with dask
