### 2025-11-21 b7bf4ac
update some dataset attributes
