### 2025-11-22 b4e14d3
update optional dependency list
