### 2025-11-22 8750936
chore: bump version to `2.2.9`
