+++ doc/source/release_notes/release-v2.2.9.rst
@@ -1,0 +1,63 @@
+##################
+`Release v2.2.9`__
+##################
+
+* ``refactor``: generalize the calculation of body tides for degrees 3+ `(#467) <https://github.com/pyTMD/pyTMD/pull/467>`_
+* ``docs``: added IERS Conventions references for Love number calculations `(#467) <https://github.com/pyTMD/pyTMD/pull/467>`_
+* ``docs``: add a draft of a JOSS paper `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``ci``: add JOSS paper build `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: add non-antarctic pyTMD use cases for `#416 <https://github.com/pyTMD/pyTMD/issues/416>`_  `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: add cite to ESA cci+ velocities `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: add doi for Kantha and Clayson `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: mention ``PERTH`` is Fortran software `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``fix``: typo `(#466) <https://github.com/pyTMD/pyTMD/pull/466>`_ `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: add blurbs about the harmonic method `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: add tide potential catalog table `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``ci``: add environmental variable for paper path `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: add spherical harmonic plots `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: add table of max tide potential `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``chore``: add more ``ruff`` lint parameters `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``refactor``: functional access to downloading scripts for `#470 <https://github.com/pyTMD/pyTMD/issues/470>`_ `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: expand on the directory structure `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: update programs for downloading from ArcticData, GOT or AVISO for `#469 <https://github.com/pyTMD/pyTMD/issues/469>`_ `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: add admonitions noting models will need to be downloaded `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: admonition in ``xarray`` recipe `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``test``: add developer download routines for `#471 <https://github.com/pyTMD/pyTMD/issues/471>`_ `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``fix``: make ``pixi`` tasks associated with features for `#472 <https://github.com/pyTMD/pyTMD/issues/472>`_ `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
+* ``docs``: update data access for `#470 <https://github.com/pyTMD/pyTMD/issues/470>`_ `(#415) <https://github.com/pyTMD/pyTMD/pull/415>`_
[…]
