### 2025-11-22 7a9ba1a
chore: bump version to `2.2.9` (#484)
