### 2025-11-22 941388f
Update CITATION.cff
