+++ CITATION.cff
@@ -14,7 +14,7 @@ keywords:
 license: MIT
 version: "2.2.8"
 date-released: "2025-09-05"
-message: If you use this software, please consider citing library.
+message: If you use this software, please consider citing our library.
 
 authors:
   - given-names: Tyler
