### 2025-11-22 fdca549
Update CITATION.cff
