+++ CITATION.cff
@@ -12,8 +12,8 @@ keywords:
   - Solid Earth Tides
   - Tidal Prediction
 license: MIT
-version: "2.2.8"
-date-released: "2025-09-05"
+version: "2.2.9"
+date-released: "2025-11-21"
 message: If you use this software, please consider citing our library.
 
 authors:
