### 2025-11-22 e40ea03
feat: added function to reduce prediction stations to active
