### 2025-11-22 #8566 @cheginit
@tsutterley please address the DOI issues mentioned in the bot comment, then use `@editorialbot check references` command to check if they all have been address. Once, this is resolved, if necessary, please release a new version of the pyTMD with all the changes and modifications made to address issues raised in this review. If no new release is required, we can go ahead with the version that you submitted. The next step is to create a DOI for an archive of your submission using services such as Zenodo. Once I have these info, I will hand it over to EiC for final acceptance.
