### 2025-11-24 6fa1caf
minor changes to address copilot review
