### 2025-11-24 4d8f6d8
more copilot nitpicks
