+++ pyTMD/math.py
@@ -238,7 +238,7 @@ def legendre(
         Plm = 15.0*u**3
         dPlm = 45.0*x*u**2
     elif (l == 4) and (m == 0):
-        Plm = (35.0*x**4 - 30.0*x**2 + 3.0)/8.0
+        Plm = 0.125*(35.0*x**4 - 30.0*x**2 + 3.0)
         dPlm = -2.5*(7.0*x**2 - 3.0)*u*x
     elif (l == 4) and (m == 1):
         Plm = 2.5*(7.0*x**2 - 3.0)*u*x
