+++ pyTMD/io/model.py
@@ -132,7 +132,10 @@ class DataBase:
         return str(self.__dict__)
 
     def get(self, key, default=None):
-        return getattr(self, key, default) or default
+        if not hasattr(self, key) or getattr(self, key) is None:
+            return default
+        else:
+            return getattr(self, key, default)
 
     def __getitem__(self, key):
         return getattr(self, key)
@@ -786,7 +789,7 @@ class model:
         kwargs.setdefault('constituents', None)
         # model type
         mtype = kwargs['type'].lower()
-        assert mtype in ('z', 'u', 'v'), "Invalid model type"
+        assert mtype in ('z', 'u', 'v'), f"Invalid model type {mtype}"
         # extract model file
         model_file = self[mtype].get('model_file')
         # reduce constituents if specified
