### 2025-11-24 fe16aa0
feat: added string check to determine if is a valid URL
