+++ pyTMD/io/GOT.py
@@ -69,7 +69,7 @@ import numpy as np
 import xarray as xr
 import pyTMD.version
 import pyTMD.constituents
-from pyTMD.utilities import import_dependency
+from pyTMD.utilities import import_dependency, is_valid_url
 # attempt imports
 dask = import_dependency('dask')
 dask_available = xr.namedarray.utils.module_available('dask')
@@ -183,7 +183,9 @@ def open_got_ascii(
     # set default keyword arguments
     kwargs.setdefault('compressed', False)
     # tilde-expand input file
-    input_file = pathlib.Path(input_file).expanduser()
+    if not is_valid_url(input_file):
+        input_file = pathlib.Path(input_file).expanduser().absolute()
+        assert input_file.exists(), f'File not found: {input_file}'
     # read the ASCII-format tide elevation file
     if kwargs['compressed']:
         # read gzipped ascii file
@@ -283,7 +285,9 @@ def open_got_netcdf(
     # set default keyword arguments
     kwargs.setdefault('compressed', False)
     # tilde-expand input file
-    input_file = pathlib.Path(input_file).expanduser()
+    if not is_valid_url(input_file):
+        input_file = pathlib.Path(input_file).expanduser().absolute()
+        assert input_file.exists(), f'File not found: {input_file}'
     # read the netCDF4-format tide elevation file
     if kwargs['compressed']:
         # read gzipped netCDF4 file
@@ -339,7 +343,7 @@ class GOTDataset:
             additional keyword arguments for xarray netCDF4 writer
         """
             # tilde-expand output path
-        path = pathlib.Path(path).expanduser()       
+        path = pathlib.Path(path).expanduser().absolute() 
         # set default encoding
         kwargs.setdefault('encoding', dict(amplitude=encoding, phase=encoding))
         # coordinate remapping
