+++ pyTMD/io/FES.py
@@ -67,7 +67,7 @@ import datetime
 import numpy as np
 import xarray as xr
 import pyTMD.constituents
-from pyTMD.utilities import import_dependency
+from pyTMD.utilities import import_dependency, is_valid_url
 # attempt imports
 dask = import_dependency('dask')
 dask_available = xr.namedarray.utils.module_available('dask')
@@ -176,7 +176,9 @@ def open_fes_ascii(
     # set default keyword arguments
     kwargs.setdefault('compressed', False)
     # tilde-expand input file
-    input_file = pathlib.Path(input_file).expanduser()
+    if not is_valid_url(input_file):
+        input_file = pathlib.Path(input_file).expanduser().absolute()
+        assert input_file.exists(), f'File not found: {input_file}'
     # read the ASCII-format tide elevation file
     if kwargs['compressed']:
         # read gzipped ascii file
@@ -277,7 +279,9 @@ def open_fes_netcdf(
     kwargs.setdefault('type', 'z')
     kwargs.setdefault('compressed', False)
     # tilde-expand input file
-    input_file = pathlib.Path(input_file).expanduser()
+    if not is_valid_url(input_file):
+        input_file = pathlib.Path(input_file).expanduser().absolute()
+        assert input_file.exists(), f'File not found: {input_file}'
     # read the netCDF4-format tide elevation file
     if kwargs['compressed']:
         # read gzipped netCDF4 file
@@ -352,7 +356,7 @@ class FESDataset:
             additional keyword arguments for xarray netCDF4 writer
         """
         # tilde-expand output path
-        path = pathlib.Path(path).expanduser()       
+        path = pathlib.Path(path).expanduser().absolute() 
         # set variable names and units for type
         if (self._ds.attrs['type'] == 'z'):
             amp_key = 'amplitude'
