### 2025-11-24 547e5bd
fix: set `unsafe_hash` to allow mutable attributes for py3.11
