### 2025-11-24 728f382
drop `dataclass` dependency
seems no longer worth the trouble
