### 2025-11-24 dd160df
feat: added function to check if a dependency is available
