### 2025-11-24 a20e98c
Update NOAA.py
