+++ pyTMD/io/NOAA.py
@@ -21,7 +21,6 @@ import logging
 import traceback
 import numpy as np
 import pyTMD.constituents
-from pyTMD.pyTMD.io.ATLAS import dependency_available
 from pyTMD.utilities import (
     import_dependency,
     dependency_available
