### 2025-11-25 675fb03
create pathlib.path subclass for allow s3
