### 2025-11-25 2f631ad
Update utilities.py
