### 2025-11-25 a2bc823
Update utilities.py
