+++ doc/source/user_guide/Recipes.rst
@@ -7,9 +7,9 @@ Recipes
 .. grid:: 2 2 4 4
     :padding: 0
 
-    .. grid-item-card::  xarray Integration
+    .. grid-item-card:: Cloud Data Access
       :text-align: center
-      :link: ./xarray.html
+      :link: ./Cloud_Access.html
 
       :material-outlined:`insert_chart;5em`
 
@@ -18,4 +18,4 @@ Recipes
     :hidden:
     :caption: Recipes
 
-    ./xarray.rst
+    ./Cloud_Access.rst
