### 2025-11-25 b1e841e
Update Recipes.rst
