### 2025-11-25 5afb41c
feat: added detection functions for compression and model format
