+++ doc/source/api_reference/io/FES.rst
@@ -13,7 +13,7 @@ Calling Sequence
 .. code-block:: python
 
     import pyTMD.io
-    ds = pyTMD.io.FES.open_mfdataset(model_files, type='z', version=version)
+    ds = pyTMD.io.FES.open_mfdataset(model_files, type='z', format=format)
 
 `Source code`__
 
