### 2025-11-25 cfe2336
Update utilities.rst
