### 2025-11-26 33968da
docs: add `_parse_name` to constituents
