### 2025-11-26 9197b8a
Update constituents.py
