### 2025-12-01 7f8e3e1
Update predict.py
