### 2025-12-01 0fff475
Update astro.py
