+++ pyTMD/compute.py
@@ -40,8 +40,8 @@ PYTHON DEPENDENCIES:
         https://numpy.org/doc/stable/user/numpy-for-matlab-users.html
     scipy: Scientific Tools for Python
         https://docs.scipy.org/doc/
-    netCDF4: Python interface to the netCDF C library
-        https://unidata.github.io/netcdf4-python/netCDF4/index.html
+    h5netcdf: Pythonic interface to netCDF4 via h5py
+        https://h5netcdf.org/
     pyproj: Python interface to PROJ library
         https://pypi.org/project/pyproj/
     timescale: Python tools for time and astronomical calculations
@@ -51,9 +51,8 @@ PROGRAM DEPENDENCIES:
     time.py: utilities for calculating time operations
     spatial: utilities for reading, writing and operating on spatial data
     utilities.py: download and management utilities for syncing files
-    arguments.py: load the nodal corrections for tidal constituents
     astro.py: computes the basic astronomical mean longitudes
-    crs.py: Coordinate Reference System (CRS) routines
+    constituents.py: calculates constituent parameters and nodal arguments  
     predict.py: predict tide values using harmonic constants
     io/model.py: retrieves tide model parameters for named tide models
     io/OTIS.py: extract tidal harmonic constants from OTIS tide models
@@ -145,14 +144,11 @@ UPDATE HISTORY:
 """
 from __future__ import print_function, annotations
 
-import logging
 import pathlib
 import numpy as np
+import xarray as xr
 from io import IOBase
-import scipy.interpolate
-import pyTMD.crs
 import pyTMD.io
-import pyTMD.io.model
 import pyTMD.predict
 import pyTMD.spatial
 import pyTMD.utilities
@@ -172,16 +168,13 @@ __all__ = [
     "SET_displacements"
 ]
 
-# number of days between the Julian day epoch and MJD
-_jd_mjd = 2400000.5
-
 # default working data directory for tide models
 _default_directory = pyTMD.utilities.get_cache_path()
 
 # PURPOSE: wrapper function for computing values
 def corrections(
         x: np.ndarray, y: np.ndarray, delta_time: np.ndarray,
-        CORRECTION: str = 'ocean',
+        correction: str = 'ocean',
         **kwargs
     ):
     """
@@ -195,7 +188,7 @@ def corrections(
         y-coordinates in projection EPSG
     delta_time: np.ndarray
         seconds since EPOCH or datetime array
-    CORRECTION: str, default 'ocean'
+    correction: str, default 'ocean'
         Correction type to compute
 
             - ``'ocean'``: ocean tide from model constituents
@@ -212,43 +205,35 @@ def corrections(
     values: np.ndarray
         tidal correction at coordinates and time in meters
     """
-    if CORRECTION.lower() in ('ocean', 'load'):
+    if correction.lower() in ('ocean', 'load'):
         return tide_elevations(x, y, delta_time, **kwargs)
-    elif (CORRECTION.upper() == 'LPET'):
+    elif (correction.upper() == 'LPET'):
         return LPET_elevations(x, y, delta_time, **kwargs)
-    elif (CORRECTION.upper() == 'LPT'):
+    elif (correction.upper() == 'LPT'):
         return LPT_displacements(x, y, delta_time, **kwargs)
-    elif (CORRECTION.upper() == 'OPT'):
+    elif (correction.upper() == 'OPT'):
         return OPT_displacements(x, y, delta_time, **kwargs)
-    elif (CORRECTION.upper() == 'SET'):
+    elif (correction.upper() == 'SET'):
         return SET_displacements(x, y, delta_time, **kwargs)
     else:
[…]
