+++ pyTMD/io/dataset.py
@@ -17,7 +17,7 @@ PYTHON DEPENDENCIES:
         https://docs.xarray.dev/en/stable/
 
 UPDATE HISTORY:
-    Updated 11/2025: get crs directly using pyTMD.CRS.from_user_input
+    Updated 11/2025: get crs directly using pyproj.CRS.from_user_input
         set variable name to constituent for to_dataarray method
         added is_global property for models covering a global domain
         added pad function to pad global datasets along boundaries
@@ -26,22 +26,241 @@ UPDATE HISTORY:
         when converting to an xarray DataArray
     Written 08/2025
 """
+import warnings
 import numpy as np
+import xarray as xr
 from pyTMD.utilities import import_dependency
+# suppress warnings
+warnings.filterwarnings("ignore", category=UserWarning)
+
 # attempt imports
-xr = import_dependency('xarray')
 pyproj = import_dependency('pyproj')
+pint = import_dependency('pint')
 
 __all__ = [
+    'DataTree',
     'Dataset',
+    'DataArray'
 ]
+
+# pint unit registry
+__ureg__ = pint.UnitRegistry()
+# default units for pyTMD outputs
+_default_units = {
+    'elevation': 'm',
+    'current': 'cm/s',
+    'transport': 'm^2/s',
+}
+
+@xr.register_datatree_accessor('tmd')
+class DataTree:
+    """Accessor for extending an ``xarray.DataTree`` for tidal model data
+    """
+    def __init__(self, dtree):
+        # initialize DataTree
+        self._dtree = dtree
+
+    def crop(self, *args, **kwargs):
+        """
+        Crop ``DataTree`` to input bounding box
+        """
+        # create copy of datatree
+        dtree = self._dtree.copy()
+        # crop each dataset in the datatree
+        for key, ds in dtree.items():
+            ds = ds.to_dataset()
+            dtree[key] = ds.tmd.crop(*args, **kwargs)
+        # return the datatree
+        return dtree
+
+    def inpaint(self, **kwargs):
+        """
+        Inpaint over missing data in ``DataTree``
+        """
+        # create copy of datatree
+        dtree = self._dtree.copy()
+        # inpaint each dataset in the datatree
+        for key, ds in dtree.items():
+            ds = ds.to_dataset()
+            dtree[key] = ds.tmd.inpaint(**kwargs)
+        # return the datatree
+        return dtree
+
+    def interp(self, x, y, **kwargs):
+        """
+        Interpolate ``DataTree`` to input coordinates
+        
+        Parameters
+        ----------
+        x: np.ndarray
+            input x-coordinates
+        y: np.ndarray
+            input y-coordinates
+        """
+        # create copy of datatree
+        dtree = self._dtree.copy()
+        # interpolate each dataset in the datatree
+        for key, ds in dtree.items():
+            ds = ds.to_dataset()
+            dtree[key] = ds.tmd.interp(x, y, **kwargs)
+        # return the datatree
+        return dtree
+
+    def subset(self, c: str | list):
+        """
+        Reduce to a subset of constituents
+
+        Parameters
+        ----------
+        c: str or list
+            List of constituents names
+        """
+        # create copy of datatree
+        dtree = self._dtree.copy()
+        # subset each dataset in the datatree
+        for key, ds in dtree.items():
+            ds = ds.to_dataset()
+            dtree[key] = ds.tmd.subset(c)
+        # return the datatree
+        return dtree
+
+    def transform(self, i1, i2, crs=4326, **kwargs):
+        """
+        Transform coordinates to/from the datatree coordinate reference system
+
+        Parameters
+        ----------
+        i1: np.ndarray
+            Input x-coordinates
+        i2: np.ndarray
+            Input y-coordinates
[…]
