### 2025-12-01 dad626b
fix: infer minor admittance method
fix: add `m1a` and `m1b` to nodal corrections
fix: admittance method if singular points
refactor: simplify linear interpolation for minor inference

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
