+++ test/test_fes_predict.py
@@ -148,4 +148,4 @@ def test_extend_array(directory, CHUNKS):
     # check that longitude values are as expected
     dlon = 1.0/16.0
     lon = np.arange(-dlon, 360 + dlon, dlon)
-    assert np.all(np.isclose(lon, ds.x.values))
+    assert np.allclose(lon, ds.x.values)

[…]
