### 2025-12-02 e5843d1
docs: edits from Laurie
