+++ doc/source/_assets/paper.md
@@ -38,7 +38,7 @@ bibliography: pytmd-refs.bib
 
 # Summary
 
-`pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and Earth tides. It is not an ocean or load tide model, but a tool for using tidal constants provided by tide models to predict the height deflections or currents at particular locations and times. It is designed to be able to handle a wide range of model formats and can incorporate different physics schemes. This flexibility allows `pyTMD` to be tailored to fit specific use cases, while still allowing ease of use.
+`pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and Earth tides. It is not an ocean or load tide model, but a tool for using tidal constants provided by tide models to predict the height deflections or currents at particular locations and times. It is designed to handle a wide range of model formats and can incorporate different physics schemes. This flexibility allows `pyTMD` to be tailored to fit specific use cases, while still allowing ease of use.
 
 # Statement of need
 
@@ -52,7 +52,7 @@ There are several ocean tide prediction software options available. The OSU Tida
 
 ## Ocean and Load Tides
 
-With the harmonic method, tides are decomposed into harmonic constants, or constituents, associated with the relative positions of the sun, moon and Earth [@Padman:2018cv]. These constituents are typically classified into different "species" based on their approximate period: short-period, semi-diurnal, diurnal, and long-period. `pyTMD.io` contains routines for reading major constituent values (amplitude and phase) from commonly available tide models, which typically fall within a few general formats: `OTIS-binary` [@Egbert:2002ge; @Padman:2008ec], `OTIS-compact`, `OTIS-netcdf`, `TMD3-netcdf` [@Greene:2024di], `GOT-ascii` [@Ray:1999vm], `GOT-netcdf`, `FES-ascii` [@LeProvost:1994ie] and `FES-netcdf` [@Lyard:2025tr]. Information for each of the supported tide models is stored within a JSON database. For tidal predictions, `pyTMD.io` interpolates the amplitude and phase lag of tide model constituents to sets of spatial coordinates. 
+With the harmonic method, tides are decomposed into harmonic constants, or constituents, associated with the relative positions of the sun, moon and Earth [Doodson:1921kt; Cartwright:1999tj]. These constituents are typically classified into different "species" based on their approximate period: short-period, semi-diurnal, diurnal, and long-period. `pyTMD.io` contains routines for reading major constituent values (amplitude and phase) from commonly available tide models, which typically fall within a few general formats: `OTIS-binary` [@Egbert:2002ge; @Padman:2008ec], `OTIS-compact`, `OTIS-netcdf`, `TMD3-netcdf` [@Greene:2024di], `GOT-ascii` [@Ray:1999vm], `GOT-netcdf`, `FES-ascii` [@LeProvost:1994ie] and `FES-netcdf` [@Lyard:2025tr]. Information for each of the supported tide models is stored within a JSON database. For tidal predictions, `pyTMD.io` interpolates the amplitude and phase lag of tide model constituents to sets of spatial coordinates. 
 
[…]
