### 2025-12-02 6d32fd4
docs: update acknowledgements
docs: add grant numbers for Susan and Laurie
