+++ doc/source/_assets/paper.md
@@ -44,7 +44,7 @@ bibliography: pytmd-refs.bib
 
 There are several ocean tide prediction software options available. The OSU Tidal Inversion Software (`OTIS`) and OSU Tidal Prediction Software (`OTPS`) are Fortran programs developed for the [TPXO](https://www.tpxo.net/home) family of models [@Egbert:2002ge]. The Tidal Model Driver ([`TMD`](https://github.com/EarthAndSpaceResearch/TMD_Matlab_Toolbox_v2.5)) is a MATLAB Toolbox developed for the same family of models. An updated version of the MATLAB Toolbox ([`TMD3`](https://github.com/chadagreene/tide-model-driver)) was developed to use a custom consolidated netCDF4 format [@Greene:2024di]. [`pyFES`](https://cnes.github.io/aviso-fes/) is a Python library produced for the Finite Element Solution (FES) family of models [@Lyard:2025tr] funded by CNES. The NASA GSFC PREdict Tidal Heights (`PERTH3`) software is a Fortran program developed for the Goddard Ocean Tide (GOT) family of models. An updated and more versatile version of the NASA GSFC Fortran software ([`PERTH5`](https://codeberg.org/rray/perth5)) can read from multiple different tide model formats. These software options are typically created by or for the model providers, and, with the exception of `PERTH5`, singly support their specific model formats.  
 
-`pyTMD` is a generalized tide program that allows users to calculate both tide deflections and currents from a broad suite of models. Over 50 different models are presently supported within `pyTMD`, and additional model schemas can be defined with a JSON file. 
+`pyTMD` is a generalized tide program that allows users to calculate both tide deflections and currents from a broad suite of models. Over 50 different models are presently supported, and additional model schemas can be defined with a JSON file. 
 
 `pyTMD` was designed to be used by beginners and scientific researchers alike. The online documentation contains background information for both tidal modeling and prediction. The software has been used in a number of scientific publications for modeling regional tides [@Freer:2023bt; @Millan:2023ju; @Sutterley:2019cl], modeling global tides [@Gregg:2024ky; @Paprotny:2024et], and creating several Earth observation datasets [@ENVEO:2021jh; @Smith:2024fd; @Smith:2024cv]. It has also been leveraged within larger earth-observation software packages [@Fitzpatrick:2024iv; @BishopTaylor:2025dc]. 
 
@@ -68,7 +68,7 @@ The tidal deformation of the solid Earth can be modeled in `pyTMD` using one of 
 
 # Acknowledgements
 
-`pyTMD` was first supported through an appointment to the NASA Postdoctoral Program (NPP) at NASA Goddard Space Flight Center (GSFC), and currently supported under the NASA Cryospheric Sciences Program (Grant Number 80NSSC22K0379). The software was developed with the goal of supporting science applications for airborne and satellite altimetry in preparation for the launch of the NASA ICESat-2 mission. It was designed for scientific and technical purposes, and not for coastal navigation or applications risking life or property.
+Contributions to `pyTMD` were first supported through an appointment to the NASA Postdoctoral Program (NPP) at NASA Goddard Space Flight Center (GSFC), and currently supported by the NASA Cryospheric Sciences Program under NASA Awards 80NSSC22K0379 (TCS) and 80NSSC21K0911 (SLH and LP). It was originally developed to support the science applications of airborne and satellite altimetry in preparation for the NASA ICESat-2 mission. It was designed for scientific and technical purposes, and not for coastal navigation or applications risking life or property.
 
[…]
