### 2025-12-02 c535883
update affilitations
