### 2025-12-02 3f9fc28
docs: add grant numbers for SLH and LP (#487)
* docs: paper edits from Laurie
