+++ doc/source/_assets/paper.md
@@ -18,19 +18,21 @@ authors:
   - name: Laurie Padman
     orcid: 0000-0003-2010-642X
     equal-contrib: true
-    affiliation: 2
+    affiliation: 3
   - name: Matthew R. Siegfried
     orcid: 0000-0002-0868-4633
     equal-contrib: true
-    affiliation: 3
+    affiliation: 4
 
 affiliations:
- - name: Polar Science Center, Applied Physics Laboratory, University of Washington, United States
+ - name: Polar Science Center, Applied Physics Laboratory, University of Washington, Seattle, WA, USA
    index: 1
- - name: Earth & Space Research, United States
+ - name: Earth & Space Research, Seattle, WA, USA
    index: 2
- - name: Department of Geophysics, Colorado School of Mines, United States
+ - name: Earth & Space Research, Corvallis, OR, USA
    index: 3
+ - name: Hydrologic Sciences & Engineering Program, Department of Geophysics, Colorado School of Mines, Golden, CO, USA
+   index: 4
 
 date: 23 September 2025
 bibliography: pytmd-refs.bib
@@ -38,13 +40,13 @@ bibliography: pytmd-refs.bib
 
 # Summary
 
-`pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and Earth tides. It is not an ocean or load tide model, but a tool for using tidal constants provided by tide models to predict the height deflections or currents at particular locations and times. It is designed to be able to handle a wide range of model formats and can incorporate different physics schemes. This flexibility allows `pyTMD` to be tailored to fit specific use cases, while still allowing ease of use.
+`pyTMD` is an open-source tidal prediction software that aims to simplify the calculation of ocean and Earth tides. It is not an ocean or load tide model, but a tool for using tidal constants provided by tide models to predict the height deflections or currents at particular locations and times. It is designed to handle a wide range of model formats and can incorporate different physics schemes. This flexibility allows `pyTMD` to be tailored to fit specific use cases, while still allowing ease of use.
 
 # Statement of need
 
 There are several ocean tide prediction software options available. The OSU Tidal Inversion Software (`OTIS`) and OSU Tidal Prediction Software (`OTPS`) are Fortran programs developed for the [TPXO](https://www.tpxo.net/home) family of models [@Egbert:2002ge]. The Tidal Model Driver ([`TMD`](https://github.com/EarthAndSpaceResearch/TMD_Matlab_Toolbox_v2.5)) is a MATLAB Toolbox developed for the same family of models. An updated version of the MATLAB Toolbox ([`TMD3`](https://github.com/chadagreene/tide-model-driver)) was developed to use a custom consolidated netCDF4 format [@Greene:2024di]. [`pyFES`](https://cnes.github.io/aviso-fes/) is a Python library produced for the Finite Element Solution (FES) family of models [@Lyard:2025tr] funded by CNES. The NASA GSFC PREdict Tidal Heights (`PERTH3`) software is a Fortran program developed for the Goddard Ocean Tide (GOT) family of models. An updated and more versatile version of the NASA GSFC Fortran software ([`PERTH5`](https://codeberg.org/rray/perth5)) can read from multiple different tide model formats. These software options are typically created by or for the model providers, and, with the exception of `PERTH5`, singly support their specific model formats.  
 
-`pyTMD` is a generalized tide program that allows users to calculate both tide deflections and currents from a broad suite of models. Over 50 different models are presently supported within `pyTMD`, and additional model schemas can be defined with a JSON file. 
[…]
