### 2025-12-02 66adc38
refactor: make dataframe accessor inherit from Dataset
