### 2025-12-03 3e176a6
refactor: group over type for consistency
docs: add netcdf from s3 recipe
