+++ pyTMD/io/model.py
@@ -221,7 +221,7 @@ class model:
 
     def from_database(self,
             m: str,
-            type: tuple = ('z', 'u', 'v')
+            group: tuple = ('z', 'u', 'v')
         ):
         """
         Create a model object from known tidal models
@@ -230,7 +230,7 @@ class model:
         ----------
         m: str
             model name
-        type: tuple, default ('z', 'u', 'v')
+        group: tuple, default ('z', 'u', 'v')
             List of model types to extract
         """
         # set working data directory if unset
@@ -244,20 +244,20 @@ class model:
         except (ValueError, KeyError, AttributeError) as exc:
             raise ValueError(f"Unlisted tide model {m}") from exc
         # verify model types to extract
-        if isinstance(type, str):
-            type = (type,)
+        if isinstance(group, str):
+            group = (group,)
         # verify paths
-        for mtype in type:
-            # verify model type is valid
-            mtype = mtype.lower()
-            # skip if model type is unavailable
-            if not hasattr(self, mtype):
+        for g in group:
+            # verify model group is valid
+            g = g.lower()
+            # skip if model group is unavailable
+            if not hasattr(self, g):
                 continue
             # validate paths: grid file for OTIS, ATLAS models
-            if hasattr(self[mtype], 'grid_file'):
-                self[mtype].grid_file = self.pathfinder(self[mtype].grid_file)
+            if hasattr(self[g], 'grid_file'):
+                self[g].grid_file = self.pathfinder(self[g].grid_file)
             # validate paths: model constituent files
-            self[mtype].model_file = self.pathfinder(self[mtype].model_file)
+            self[g].model_file = self.pathfinder(self[g].model_file)
         # return the model parameters
         self.validate_format()
         # set dictionary of parameters
@@ -342,7 +342,7 @@ class model:
     @property
     def corrections(self) -> str:
         """
-        Returns the corrections type for the model
+        Returns the corrections group for the model
         """
         part1, _, part2 = self.format.partition('-')
         if self.format in ('GOT-ascii', ):
@@ -367,14 +367,14 @@ class model:
     def multifile(self) -> bool:
         """Returns if the model uses individual files for constituents
         """
-        # try to find a valid mode type
-        for mtype in ('z', 'u', 'v'):
-            # verify case of model type
-            mtype = mtype.lower()
-            # skip if model type is unavailable
-            if not hasattr(self, mtype):
+        # try to find a valid mode group
+        for g in ('z', 'u', 'v'):
+            # verify case of model group
+            g = g.lower()
+            # skip if model group is unavailable
+            if not hasattr(self, g):
                 continue
-            return isinstance(self[mtype].model_file, list)
+            return isinstance(self[g].model_file, list)
 
     @property
     def crs(self):
@@ -603,43 +603,43 @@ class model:
         # verify parameters for each model format
         if temp.format in ('OTIS','ATLAS-compact','ATLAS-netcdf',):
             # extract model files
-            for mtype in ('z', 'u', 'v'):
-                # check that model type is available
-                if not hasattr(temp, mtype):
+            for g in ('z', 'u', 'v'):
+                # check that model group is available
+                if not hasattr(temp, g):
                     continue
-                assert temp[mtype].grid_file
+                assert temp[g].grid_file
                 # check if grid file is relative or absolute
                 if (temp.directory is not None):
-                    temp[mtype].grid_file = \
-                        temp.directory.joinpath(temp[mtype].grid_file)
+                    temp[g].grid_file = \
+                        temp.directory.joinpath(temp[g].grid_file)
                 else:
-                    temp[mtype].grid_file = \
[…]
