### 2025-12-03 770d6cc
refactor: rewrite io routines to use `xarray` (#486)
* feat: unit-aware conversions with `pint`
* ci: don't run actions for `[WIP]` PRs
* refactor: restructure json files to `z`, `u`, `v` format
* refactor: reduce to only dataset predict (now `time_series`)
* refactor: update variable case for #100
* test: ignore `pyproj` warnings
* docs: run readthedocs in `dev` env
* test: add check for `interp1d` arrays
* feat: added function to reduce NOAA prediction stations to active
* feat: added function to check if a dependency is available
* feat: create `pathlib.Path` subclass for allow s3
* feat: a
[…]
