+++ test/test_solid_earth.py
@@ -18,6 +18,7 @@ UPDATE HISTORY:
 """
 import pytest
 import numpy as np
+import xarray as xr
 import pyTMD.astro
 import pyTMD.compute
 import pyTMD.predict
@@ -27,9 +28,15 @@ def test_out_of_phase_diurnal():
     """Test out-of-phase diurnal corrections with IERS outputs
     """
     # station locations and planetary ephemerides
-    XYZ = np.array([[4075578.385, 931852.890, 4801570.154]])
-    SXYZ = np.array([[137859926952.015, 54228127881.4350, 23509422341.6960]])
-    LXYZ = np.array([[-179996231.920342, -312468450.131567, -169288918.592160]])
+    XYZ = xr.Dataset(data_vars=dict(
+        X=4075578.385, Y=931852.890, Z=4801570.154
+    ))
+    SXYZ = xr.Dataset(data_vars=dict(
+        X=137859926952.015, Y=54228127881.4350, Z=23509422341.6960
+    ))
+    LXYZ = xr.Dataset(data_vars=dict(
+        X=-179996231.920342, Y=-312468450.131567, Z=-169288918.592160
+    ))
     # factors for sun and moon
     F2_solar = 0.163271964478954
     F2_lunar = 0.321989090026845
@@ -41,15 +48,23 @@ def test_out_of_phase_diurnal():
     dXYZ = pyTMD.predict._out_of_phase_diurnal(XYZ, SXYZ, LXYZ,
         F2_solar, F2_lunar)
     # assert matching
-    assert np.isclose(np.c_[dx_expected, dy_expected, dz_expected], dXYZ).all()
+    assert np.isclose(dx_expected, dXYZ['X'])
+    assert np.isclose(dy_expected, dXYZ['Y'])
+    assert np.isclose(dz_expected, dXYZ['Z'])
 
 def test_out_of_phase_semidiurnal():
     """Test out-of-phase semidiurnal corrections with IERS outputs
     """
     # station locations and planetary ephemerides
-    XYZ = np.array([[4075578.385, 931852.890, 4801570.154]])
-    SXYZ = np.array([[137859926952.015, 54228127881.4350, 23509422341.6960]])
-    LXYZ = np.array([[-179996231.920342, -312468450.131567, -169288918.592160]])
+    XYZ = xr.Dataset(data_vars=dict(
+        X=4075578.385, Y=931852.890, Z=4801570.154
+    ))
+    SXYZ = xr.Dataset(data_vars=dict(
+        X=137859926952.015, Y=54228127881.4350, Z=23509422341.6960
+    ))
+    LXYZ = xr.Dataset(data_vars=dict(
+        X=-179996231.920342, Y=-312468450.131567, Z=-169288918.592160
+    ))
     # factors for sun and moon
     F2_solar = 0.163271964478954
     F2_lunar = 0.321989090026845
@@ -61,15 +76,23 @@ def test_out_of_phase_semidiurnal():
     dXYZ = pyTMD.predict._out_of_phase_semidiurnal(XYZ, SXYZ, LXYZ,
         F2_solar, F2_lunar)
     # assert matching
-    assert np.isclose(np.c_[dx_expected, dy_expected, dz_expected], dXYZ).all()
+    assert np.isclose(dx_expected, dXYZ['X'])
+    assert np.isclose(dy_expected, dXYZ['Y'])
+    assert np.isclose(dz_expected, dXYZ['Z'])
 
 def test_latitude_dependence():
     """Test latitude dependence corrections with IERS outputs
     """
     # station locations and planetary ephemerides
-    XYZ = np.array([[4075578.385, 931852.890, 4801570.154]])
[…]
