### 2025-12-03 28b843f
refactor: merge minor constituent inference loop
