### 2025-12-03 7b030b3
docs: update module docstrings
