### 2025-12-03 96b4f8a
test: add tests against PERTH5
