+++ test/test_perth5_predict.py
@@ -1,0 +1,185 @@
+"""
+test_perth5_predict.py
+Tests predictions against outputs from the NASA PERTH5 program
+
+PYTHON DEPENDENCIES:
+    numpy: Scientific Computing Tools For Python
+        https://numpy.org
+        https://numpy.org/doc/stable/user/numpy-for-matlab-users.html
+    scipy: Scientific Tools for Python
+        https://docs.scipy.org/doc/
+    timescale: Python tools for time and astronomical calculations
+        https://pypi.org/project/timescale/
+    xarray: N-D labeled arrays and datasets in Python
+        https://docs.xarray.dev/en/stable/
+
+UPDATE HISTORY:
+    Updated 11/2025: using new xarray interface for tidal model data
+    Written 09/2024
+"""
+import pytest
+import pyTMD.io
+import numpy as np
+import xarray as xr
+import timescale.time
+
+# parametrize chunking
+@pytest.mark.parametrize("CHUNKS", ["auto"])
+def test_verify_GOT55(CHUNKS):
+    # model parameters for GOT5.5
+    m = pyTMD.io.model().from_database('GOT5.5')
+    # open dataset and keep units as centimeters
+    ds = m.open_dataset(group='z', chunks=CHUNKS, use_default_units=False)
+    # test point from perth5 validation data
+    MJD = np.array([45335.00000, 45335.04166667])
+    lat = 59.195
+    lon = -7.688
+    # convert time from MJD to timescale object
+    ts = timescale.time.Timescale(MJD=MJD)
+    # transform coordinates to model grid
+    X, Y = ds.tmd.transform(lon, lat, crs=4326)
+    # extract amplitude and phase from tide model
+    local = ds.tmd.interp(X, Y)
+    # check constituent values
+    assert np.isclose(local['q1'], 2.04 + 2.38j, atol=1e-2)
+    assert np.isclose(local['o1'], 7.72 - 1.23j, atol=1e-2)
+    assert np.isclose(local['p1'], -2.01 - 1.64j, atol=1e-2)
+    assert np.isclose(local['s1'], 0.48 - 0.49j, atol=1e-2)
+    assert np.isclose(local['k1'], -7.95 - 5.25j, atol=1e-2)
+    assert np.isclose(local['n2'], -18.64 - 3.01j, atol=1e-2)
+    assert np.isclose(local['m2'], -90.52 + 20.60j, atol=1e-2)
+    assert np.isclose(local['s2'], -24.58 + 25.66j, atol=1e-2)
+    assert np.isclose(local['k2'], -7.38 + 7.10j, atol=1e-2)
+    assert np.isclose(local['m4'], 0.26 - 0.14j, atol=1e-2)
+    assert np.isclose(local['ms4'], -0.06 + 0.14j, atol=1e-2)
+    assert np.isclose(local['2n2'], -2.24 - 1.32j, atol=1e-2)
+    assert np.isclose(local['mu2'], -2.65 - 2.41j, atol=1e-2)
+    assert np.isclose(local['j1'], -0.33 - 0.15j, atol=1e-2)
+    assert np.isclose(local['sigma1'], 0.03 + 0.70j, atol=1e-2)
+    assert np.isclose(local['oo1'], -0.22 + 0.18j, atol=1e-2)
+    # predict tidal elevations at time
+    tide = local.tmd.predict(ts.tide, deltat=ts.tt_ut1, corrections='GOT')
+    # infer semi-diurnal corrections
+    tide += pyTMD.predict._infer_semi_diurnal(ts.tide, local,
+        deltat=ts.tt_ut1, corrections='GOT')
+    # infer diurnal corrections
[…]
