### 2025-12-03 7ed6b99
test: copilot nitpicks
