### 2025-12-03 fa682ac
test: add tests against PERTH5 (#488)
* refactor: merge minor constituent inference loop
* docs: update module docstrings
