### 2025-12-04 533b4f2
feat: add functions to transform coordinates to DataArrays
will be formatted with appropriate dimensions
