+++ pyTMD/compute.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 compute.py
-Written by Tyler Sutterley (11/2025)
+Written by Tyler Sutterley (12/2025)
 Calculates tidal elevations for correcting elevation or imagery data
 Calculates tidal currents at locations and times
 
@@ -62,6 +62,7 @@ PROGRAM DEPENDENCIES:
     interpolate.py: interpolation routines for spatial data
 
 UPDATE HISTORY:
+    Updated 12/2025: use coords functions to convert x and y to DataArrays
     Updated 11/2025: use xarray DataArrays for input coordinates
         outputs from prediction functions will be also be DataArrays
     Updated 10/2025: change default directory for tide models to cache
@@ -356,25 +357,7 @@ def tide_elevations(
     assert type.lower() in ('grid', 'drift', 'time series')
     # convert coordinates to xarray DataArrays
     # in coordinate reference system of model
-    if (np.ndim(x) == 0) and (np.ndim(y) == 0):
-        X, Y = ds.tmd.transform(x, y, crs=crs)
-    elif (type.lower() == 'grid') and (np.size(x) != np.size(y)):
-        gridx, gridy = np.meshgrid(x, y)
-        mx, my = ds.tmd.transform(gridx, gridy, crs=crs)
-        X = xr.DataArray(mx, dims=('y','x'))
-        Y = xr.DataArray(my, dims=('y','x'))
-    elif (type.lower() == 'grid'):
-        mx, my = ds.tmd.transform(x, y, crs=crs)
-        X = xr.DataArray(mx, dims=('y','x'))
-        Y = xr.DataArray(my, dims=('y','x'))
-    elif (type.lower() == 'drift'):
-        mx, my = ds.tmd.transform(x, y, crs=crs)
-        X = xr.DataArray(mx, dims=('time'))
-        Y = xr.DataArray(my, dims=('time'))
-    elif (type.lower() == 'time series'):
-        mx, my = ds.tmd.transform(x, y, crs=crs)
-        X = xr.DataArray(mx, dims=('station'))
-        Y = xr.DataArray(my, dims=('station'))
+    X, Y = ds.tmd.coords_as(x, y, type=type, crs=crs)
 
     # crop tide model dataset to bounds
     if crop and bounds is None:
@@ -549,25 +532,7 @@ def tide_currents(
     assert type.lower() in ('grid', 'drift', 'time series')
     # convert coordinates to xarray DataArrays
     # in coordinate reference system of model
-    if (np.ndim(x) == 0) and (np.ndim(y) == 0):
-        X, Y = dtree.tmd.transform(x, y, crs=crs)
-    elif (type.lower() == 'grid') and (np.size(x) != np.size(y)):
-        gridx, gridy = np.meshgrid(x, y)
-        mx, my = dtree.tmd.transform(gridx, gridy, crs=crs)
-        X = xr.DataArray(mx, dims=('y','x'))
-        Y = xr.DataArray(my, dims=('y','x'))
-    elif (type.lower() == 'grid'):
-        mx, my = dtree.tmd.transform(x, y, crs=crs)
-        X = xr.DataArray(mx, dims=('y','x'))
-        Y = xr.DataArray(my, dims=('y','x'))
-    elif (type.lower() == 'drift'):
-        mx, my = dtree.tmd.transform(x, y, crs=crs)
-        X = xr.DataArray(mx, dims=('time'))
-        Y = xr.DataArray(my, dims=('time'))
-    elif (type.lower() == 'time series'):
-        mx, my = dtree.tmd.transform(x, y, crs=crs)
-        X = xr.DataArray(mx, dims=('station'))
[…]
