+++ pyTMD/io/dataset.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 u"""
 dataset.py
-Written by Tyler Sutterley (11/2025)
+Written by Tyler Sutterley (12/2025)
 An xarray.Dataset extension for tidal model data
 
 PYTHON DEPENDENCIES:
@@ -17,6 +17,7 @@ PYTHON DEPENDENCIES:
         https://docs.xarray.dev/en/stable/
 
 UPDATE HISTORY:
+    Updated 12/2025: add coords functions to transform coordinates
     Updated 11/2025: get crs directly using pyproj.CRS.from_user_input
         set variable name to constituent for to_dataarray method
         added is_global property for models covering a global domain
@@ -40,7 +41,9 @@ pint = import_dependency('pint')
 __all__ = [
     'DataTree',
     'Dataset',
-    'DataArray'
+    'DataArray',
+    '_transform',
+    '_coords'
 ]
 
 # pint unit registry
@@ -59,6 +62,42 @@ class DataTree:
     def __init__(self, dtree):
         # initialize DataTree
         self._dtree = dtree
+
+    def coords_as(self, 
+            x: np.ndarray,
+            y: np.ndarray,
+            crs: str | int | dict = 4326,
+            **kwargs
+        ):
+        """
+        Transform coordinates into DataArrays in the DataTree
+        coordinate reference system
+        
+        Parameters
+        ----------
+        x: np.ndarray
+            Input x-coordinates
+        y: np.ndarray
+            Input y-coordinates
+        crs: str, int, or dict, default 4326 (WGS84 Latitude/Longitude)
+            Coordinate reference system of input coordinates
+
+        Returns
+        -------
+        X: xarray.DataArray
+            Transformed x-coordinates
+        Y: xarray.DataArray
+            Transformed y-coordinates
+        """
+        # convert coordinate reference system to that of the datatree
+        # and format as xarray DataArray with appropriate dimensions
+        X, Y = _coords(x, y,
+            source_crs=crs,
+            target_crs=self.crs,
+            **kwargs
+        )
+        # return the transformed coordinates
+        return X, Y
 
     def crop(self, *args, **kwargs):
         """
@@ -124,15 +163,15 @@ class DataTree:
         # return the datatree
         return dtree
 
-    def transform(self, i1, i2, crs=4326, **kwargs):
+    def transform_as(self, x, y, crs=4326, **kwargs):
         """
         Transform coordinates to/from the datatree coordinate reference system
 
         Parameters
         ----------
-        i1: np.ndarray
+        x: np.ndarray
             Input x-coordinates
-        i2: np.ndarray
+        y: np.ndarray
             Input y-coordinates
         crs: str, int, or dict, default 4326 (WGS84 Latitude/Longitude)
             Coordinate reference system of input coordinates
@@ -144,21 +183,19 @@ class DataTree:
         
         Returns
         -------
-        o1: np.ndarray
+        X: np.ndarray
             Transformed x-coordinates
-        o2: np.ndarray
+        Y: np.ndarray
             Transformed y-coordinates
         """
-        # set the direction of the transformation
-        kwargs.setdefault('direction', 'FORWARD')
-        # get the coordinate reference system and transform
-        source_crs = pyproj.CRS.from_user_input(crs)
-        transformer = pyproj.Transformer.from_crs(
-            source_crs, self.crs, always_xy=True)
-        # convert coordinate reference system
-        o1, o2 = transformer.transform(i1, i2, **kwargs)
+        # convert coordinate reference system to that of the datatree
+        X, Y = _transform(x, y,
+            source_crs=crs,
+            target_crs=self.crs,
+            **kwargs
+        )
[…]

[…]
