### 2025-12-04 612629c
copilot nitpicks
