### 2025-12-04 3c05fc8
docs: add `.zenodo.json`
docs: update `CITATION.cff` to use 1.3.0
