### 2025-12-04 42e23e7
chore: bump version to `3.0.0`
🎉 🎉 🎉
