+++ doc/source/release_notes/release-v3.0.0.rst
@@ -1,0 +1,27 @@
+##################
+`Release v3.0.0`__
+##################
+
+* ``refactor``: rewrite ``io`` routines to use xarray `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``feat``: unit-aware conversions with ``pint`` `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``ci``: don't run actions for ``[WIP]`` PRs `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``refactor``: restructure json files to ``z``, ``u``, ``v`` format `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``refactor``: reduce to only dataset predict (now ``time_series``) `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``refactor``: update variable case for `#100 <https://github.com/pyTMD/pyTMD/issues/100>`_ `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``test``: ignore ``pyproj`` warnings `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``docs``: run readthedocs in ``dev`` env `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``test``: add check for ``interp1d`` arrays `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``feat``: added function to reduce NOAA prediction stations to active `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``feat``: added function to check if a dependency is available `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``feat``: create ``pathlib.Path`` subclass for allow s3 `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``feat``: added detection functions for compression and model format `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``fix``: add ``m1a`` and ``m1b`` to nodal corrections `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``docs``: add netcdf from s3 recipe `(#486) <https://github.com/pyTMD/pyTMD/pull/486>`_
+* ``docs``: add grant numbers for SLH and LP `(#487) <https://github.com/pyTMD/pyTMD/pull/487>`_
+* ``docs``: paper edits from Laurie Padman `(#487) <https://github.com/pyTMD/pyTMD/pull/487>`_
+* ``test``: add tests against ``PERTH5`` `(#488) <https://github.com/pyTMD/pyTMD/pull/488>`_
+* ``refactor``: merge minor constituent inference loop `(#488) <https://github.com/pyTMD/pyTMD/pull/488>`_
+* ``docs``: update module docstrings `(#488) <https://github.com/pyTMD/pyTMD/pull/488>`_
+* ``feat``: add functions to transform coordinates to ``DataArrays`` `(#489) <https://github.com/pyTMD/pyTMD/pull/489>`_
+
+.. __: https://github.com/pyTMD/pyTMD/releases/tag/3.0.0
