+++ doc/source/project/Citations.rst
@@ -13,7 +13,7 @@ It is currently supported under the NASA Cryospheric Sciences Program
 
 .. admonition:: Please consider citing our library
 
-    T. C. Sutterley, K. Alley, K. Brunt, S. Howard, L. Padman, and M. Siegfried,
+    T. C. Sutterley, S. L. Howard, L. Padman, and M. Siegfried,
     "pyTMD: Python-based tidal prediction software", (2017).
     `doi: 10.5281/zenodo.5555395 <https://doi.org/10.5281/zenodo.5555395>`_
 
