### 2025-12-04 574d577
chore: bump version to `3.0.0` (#490)
* docs: add `.zenodo.json` for authors and contributors
* docs: update `CITATION.cff` to use `1.3.0` schema
