+++ CITATION.cff
@@ -1,7 +1,7 @@
 # This CITATION.cff file was generated with cffinit.
 # Visit https://bit.ly/cffinit to generate yours today!
 
-cff-version: 1.2.0
+cff-version: 1.3.0
 title: pyTMD
 abstract: Python-based tidal prediction software
 type: software
@@ -12,23 +12,14 @@ keywords:
   - Solid Earth Tides
   - Tidal Prediction
 license: MIT
-version: "2.2.9"
-date-released: "2025-11-21"
+version: "3.0.0"
+date-released: "2025-12-04"
 message: If you use this software, please consider citing our library.
 
 authors:
   - given-names: Tyler
     family-names: Sutterley
     affiliation: University of Washington, Applied Physics Laboratory
-  - given-names: Karen
-    family-names: Alley
-    affiliation: University of Manitoba
-  - given-names: Robbi
-    family-names: Bishop-Taylor
-    affiliation: Geoscience Australia
-  - given-names: Kelly
-    family-names: Brunt
-    affiliation: NASA Goddard Space Flight Center
   - given-names: Susan
     family-names: Howard
     affiliation: Earth & Space Research
@@ -38,6 +29,17 @@ authors:
   - given-names: Matthew
     family-names: Siegfried
     affiliation: Colorado School of Mines
+
+contributors:
+  - given-names: Karen
+    family-names: Alley
+    affiliation: University of Manitoba
+  - given-names: Robbi
+    family-names: Bishop-Taylor
+    affiliation: Geoscience Australia
+  - given-names: Kelly
+    family-names: Brunt
+    affiliation: National Science Foundation
 
 identifiers:
   - type: doi
