### 2025-12-04 #8566 @tsutterley
Hopefully good to go! :)
https://zenodo.org/records/17822603
