### 2025-12-05 dd0d38a
fix: bump python versions to allow 3.14
