### 2025-12-05 6ec9598
chore: ruff format for #100
