+++ pyTMD/predict.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-u"""
+"""
 predict.py
 Written by Tyler Sutterley (11/2025)
 Prediction routines for ocean, load, equilibrium and solid earth tides
@@ -17,7 +17,7 @@ PYTHON DEPENDENCIES:
 
 PROGRAM DEPENDENCIES:
     astro.py: computes the basic astronomical mean longitudes
-    constituents.py: calculates constituent parameters and nodal arguments  
+    constituents.py: calculates constituent parameters and nodal arguments
     spatial.py: utilities for working with geospatial data
 
 UPDATE HISTORY:
@@ -55,7 +55,7 @@ UPDATE HISTORY:
     Updated 08/2024: minor nodal angle corrections in radians to match arguments
         include inference of eps2 and eta2 when predicting from GOT models
         add keyword argument to allow inferring specific minor constituents
-    	use nodal arguments for all non-OTIS model type cases
+        use nodal arguments for all non-OTIS model type cases
         add load pole tide function that exports in cartesian coordinates
         add ocean pole tide function that exports in cartesian coordinates
     Updated 07/2024: use normalize_angle from pyTMD astro module
@@ -84,6 +84,7 @@ UPDATE HISTORY:
     Updated 07/2018: added option to use GSFC GOT nodal corrections
     Updated 09/2017: Rewritten in Python
 """
+
 from __future__ import annotations
 
 import logging
@@ -94,7 +95,6 @@ import pyTMD.constituents
 import pyTMD.math
 import pyTMD.interpolate
 import pyTMD.spatial
-from pyTMD.utilities import import_dependency
 import timescale.eop
 
 __all__ = [
@@ -114,7 +114,7 @@ __all__ = [
     "_frequency_dependence_diurnal",
     "_frequency_dependence_long_period",
     "_free_to_mean",
-    "body_tide"
+    "body_tide",
 ]
 
 # number of days between the Julian day epoch and MJD
@@ -124,11 +124,8 @@ _mjd_tide = 48622.0
 # number of days between the Julian day epoch and the tide epoch
 _jd_tide = _jd_mjd + _mjd_tide
 
-def time_series(
-        t: float | np.ndarray,
-        ds: xr.Dataset,
-        **kwargs
-    ):
+
+def time_series(t: float | np.ndarray, ds: xr.Dataset, **kwargs):
     """
     Predict tides from ``Dataset`` at times
 
@@ -147,52 +144,51 @@ def time_series(
         predicted tides
     """
     # set default keyword arguments
-    kwargs.setdefault('corrections', 'OTIS')
+    kwargs.setdefault("corrections", "OTIS")
     # convert time to Modified Julian Days (MJD)
     MJD = t + _mjd_tide
     # list of constituents
     constituents = ds.tmd.constituents
     # load the nodal corrections
-    pu, pf, G = pyTMD.constituents.arguments(MJD,
-        constituents, **kwargs)
+    pu, pf, G = pyTMD.constituents.arguments(MJD, constituents, **kwargs)
     # calculate constituent phase angles
-    if kwargs['corrections'] in ('OTIS', 'ATLAS', 'TMD3'):
+    if kwargs["corrections"] in ("OTIS", "ATLAS", "TMD3"):
         theta = np.zeros_like(pu)
         for i, c in enumerate(constituents):
             # load parameters for constituent
[…]
