+++ pyTMD/datasets/fetch_test_data.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-u"""
+"""
 fetch_test_data.py
 Written by Tyler Sutterley (09/2025)
 Download files necessary to run the test suite
@@ -24,6 +24,7 @@ UPDATE HISTORY:
     Updated 10/2025: change default directory for tide models to cache
     Written 10/2025
 """
+
 import ssl
 import json
 import shutil
@@ -38,12 +39,13 @@ _default_directory = pyTMD.utilities.get_cache_path()
 # default ssl context
 _default_ssl_context = pyTMD.utilities._default_ssl_context
 
+
 def fetch_test_data(
-        directory: str | pathlib.Path = _default_directory,
-        provider: str = 'figshare',
-        mode: oct = 0o775,
-        **kwargs
-    ):
+    directory: str | pathlib.Path = _default_directory,
+    provider: str = "figshare",
+    mode: oct = 0o775,
+    **kwargs,
+):
     """
     Download files necessary to run the test suite
 
@@ -61,22 +63,23 @@ def fetch_test_data(
     directory.mkdir(parents=True, exist_ok=True, mode=mode)
     # create logger for verbosity level
     logger = pyTMD.utilities.build_logger(__name__, level=logging.INFO)
-    if (provider == 'figshare'):
+    if provider == "figshare":
         from_figshare(directory=directory, logger=logger, **kwargs)
     else:
-        raise ValueError(f'Unknown data provider: {provider}')
+        raise ValueError(f"Unknown data provider: {provider}")
+
 
 # PURPOSE: download data files from figshare
 def from_figshare(
-        directory: str | pathlib.Path = _default_directory,
-        article: str = '30260326',
-        timeout: int | None = None,
-        context: ssl.SSLContext = _default_ssl_context,
-        chunk: int | None = 16384,
-        logger: logging.Logger | None = None,
-        mode: oct = 0o775,
-        **kwargs
-    ):
+    directory: str | pathlib.Path = _default_directory,
+    article: str = "30260326",
+    timeout: int | None = None,
+    context: ssl.SSLContext = _default_ssl_context,
+    chunk: int | None = 16384,
+    logger: logging.Logger | None = None,
+    mode: oct = 0o775,
+    **kwargs,
+):
     """
     Download files necessary to run the test suite from figshare
 
@@ -100,31 +103,31 @@ def from_figshare(
         permissions mode of output local file
     """
     # figshare host for articles
-    HOST = ['https://api.figshare.com', 'v2', 'articles', article]
+    HOST = ["https://api.figshare.com", "v2", "articles", article]
     # Create and submit request
-    response = pyTMD.utilities.from_http(HOST,
-        timeout=timeout, context=context)
+    response = pyTMD.utilities.from_http(HOST, timeout=timeout, context=context)
     resp = json.loads(response.read())
     # for each file in the JSON response
-    for f in resp['files']:
+    for f in resp["files"]:
         # check if file already exists by matching MD5 checksums
-        local_file = directory.joinpath(f['name'])
+        local_file = directory.joinpath(f["name"])
         original_md5 = local_file.md5_hash
         # skip download if checksums match
-        if (original_md5 == f['supplied_md5']):
+        if original_md5 == f["supplied_md5"]:
             continue
         # output file information
         logger.info(f["download_url"])
         # get remote file as a byte-stream
-        remote_buffer = pyTMD.utilities.from_http(f['download_url'],
-            timeout=timeout, context=context)
+        remote_buffer = pyTMD.utilities.from_http(
+            f["download_url"], timeout=timeout, context=context
+        )
         # verify MD5 checksums
         computed_md5 = pyTMD.utilities.get_hash(remote_buffer)
         # raise exception if checksums do not match
-        if (computed_md5 != f['supplied_md5']):
[…]
