+++ pyTMD/__init__.py
@@ -11,6 +11,7 @@ matplotlib and cartopy
 
 Documentation is available at https://pytmd.readthedocs.io
 """
+
 import pyTMD.astro
 import pyTMD.compute
 import pyTMD.constituents

[…]
