### 2025-12-06 9608a54
ci: add linter actions
