+++ .github/workflows/linter_actions.yml
@@ -1,0 +1,15 @@
+name: Run ruff linters on pull requests
+
+on:
+  pull_request:
+    types: [opened, synchronize, reopened, ready_for_review]
+    branches:
+      - main
+
+jobs:
+  lint:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Lint and annotate PR
+        uses: astral-sh/ruff-action@v3
