### 2025-12-06 c737882
Delete linter_actions.yml
