### 2025-12-06 ac5ea4e
chore: `ruff` format for #100 (#493)
