+++ pyTMD/io/GOT.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-u"""
+"""
 GOT.py
 Written by Tyler Sutterley (11/2025)
 
@@ -59,6 +59,7 @@ UPDATE HISTORY:
     Updated 08/2018: added multivariate spline interpolation option
     Written 07/2018
 """
+
 from __future__ import division, annotations
 
 import re
@@ -70,24 +71,26 @@ import xarray as xr
 import pyTMD.version
 import pyTMD.constituents
 import pyTMD.utilities
+
 # attempt imports
-dask = pyTMD.utilities.import_dependency('dask')
-dask_available = pyTMD.utilities.dependency_available('dask')
+dask = pyTMD.utilities.import_dependency("dask")
+dask_available = pyTMD.utilities.dependency_available("dask")
 
 __all__ = [
-    'open_mfdataset',
-    'open_got_dataset',
-    'open_got_ascii',
-    'open_got_netcdf',
-    'GOTDataset',
+    "open_mfdataset",
+    "open_got_dataset",
+    "open_got_ascii",
+    "open_got_netcdf",
+    "GOTDataset",
 ]
+
 
 # PURPOSE: read a list of GOT ASCII or netCDF4 files
 def open_mfdataset(
-        model_files: list[str] | list[pathlib.Path],
-        parallel: bool = False,
-        **kwargs,
-    ):
+    model_files: list[str] | list[pathlib.Path],
+    parallel: bool = False,
+    **kwargs,
+):
     """
     Open multiple GOT model files
 
@@ -114,17 +117,15 @@ def open_mfdataset(
     d = [opener(f, **kwargs) for f in model_files]
     # read datasets as dask arrays
     if parallel and dask_available:
-        d, = dask.compute(d)
+        (d,) = dask.compute(d)
     # merge datasets
-    ds = xr.merge(d, compat='override')
+    ds = xr.merge(d, compat="override")
     # return xarray dataset
     return ds
 
+
 # PURPOSE: reads a GOT ASCII or netCDF4 file
-def open_got_dataset(
-        input_file: str | pathlib.Path,
-        **kwargs
-    ):
+def open_got_dataset(input_file: str | pathlib.Path, **kwargs):
     """
     Open GOT-formatted model files
 
@@ -139,36 +140,37 @@ def open_got_dataset(
             - ``'netcdf'``: GOT netCDF4 format
     **kwargs: dict
         additional keyword arguments for opening GOT files
-    
+
     Returns
     -------
     ds: xarray.Dataset
         GOT tide model data
     """
     # detect file format if not provided
-    if kwargs.get('format', None) is None:
-        kwargs['format'] = pyTMD.utilities.detect_format(input_file)
+    if kwargs.get("format", None) is None:
+        kwargs["format"] = pyTMD.utilities.detect_format(input_file)
     # detect if file is compressed if not provided
-    if kwargs.get('compressed', None) is None:
-        kwargs['compressed'] = pyTMD.utilities.detect_compression(input_file)
+    if kwargs.get("compressed", None) is None:
+        kwargs["compressed"] = pyTMD.utilities.detect_compression(input_file)
     # read constituent from elevation file
-    if kwargs['format'] == 'ascii':
+    if kwargs["format"] == "ascii":
         # GOT ascii constituent files
         ds = open_got_ascii(input_file, **kwargs)
-    elif kwargs['format'] == 'netcdf':
+    elif kwargs["format"] == "netcdf":
         # GOT netCDF4 constituent files
         ds = open_got_netcdf(input_file, **kwargs)
     else:
-        raise ValueError(f"Unrecognized file format: {kwargs['format']}") 
+        raise ValueError(f"Unrecognized file format: {kwargs['format']}")
[…]
