+++ pyTMD/tools.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-u"""
+"""
 tools.py
 Written by Tyler Sutterley (10/2025)
 Jupyter notebook, user interface and plotting tools
@@ -34,24 +34,26 @@ UPDATE HISTORY:
     Updated 02/2022: add leaflet map projections
     Written 09/2021
 """
+
 import copy
+import pyproj
 import datetime
 import numpy as np
 import pyTMD.io.model
 import pyTMD.utilities
 
 # attempt imports
-IPython = pyTMD.utilities.import_dependency('IPython')
-ipyleaflet = pyTMD.utilities.import_dependency('ipyleaflet')
-ipywidgets = pyTMD.utilities.import_dependency('ipywidgets')
-pyproj = pyTMD.utilities.import_dependency('pyproj')
+IPython = pyTMD.utilities.import_dependency("IPython")
+ipyleaflet = pyTMD.utilities.import_dependency("ipyleaflet")
+ipywidgets = pyTMD.utilities.import_dependency("ipywidgets")
+
 
 class widgets:
     def __init__(self, **kwargs):
         # set default keyword options
-        kwargs.setdefault('style', {})
+        kwargs.setdefault("style", {})
         # set style
-        self.style = copy.copy(kwargs['style'])
+        self.style = copy.copy(kwargs["style"])
         # pass through some ipywidgets objects
         self.HBox = ipywidgets.HBox
         self.VBox = ipywidgets.VBox
@@ -61,29 +63,39 @@ class widgets:
         # set the directory with tide models
         self.directory = ipywidgets.Text(
             value=str(default_directory),
-            description='Directory:',
-            disabled=False
+            description="Directory:",
+            disabled=False,
         )
 
         # dropdown menu for setting tide model
-        model_list = sorted(pyTMD.io.model.ocean_elevation() +
-            pyTMD.io.model.load_elevation())
+        model_list = sorted(
+            pyTMD.io.model.ocean_elevation() + pyTMD.io.model.load_elevation()
+        )
         self.model = ipywidgets.Dropdown(
             options=model_list,
-            value='GOT4.10',
-            description='Model:',
+            value="GOT4.10",
+            description="Model:",
             disabled=False,
             style=self.style,
         )
 
         # dropdown menu for setting model constituents
         constituents_list = [
-            'q1','o1','p1','k1','n2','m2','s2','k2','s1','m4'
+            "q1",
+            "o1",
+            "p1",
+            "k1",
+            "n2",
+            "m2",
+            "s2",
+            "k2",
+            "s1",
+            "m4",
         ]
         self.constituents = ipywidgets.Dropdown(
             options=constituents_list,
-            value='m2',
-            description='Constituents:',
+            value="m2",
+            description="Constituents:",
             disabled=False,
             style=self.style,
         )
@@ -91,112 +103,132 @@ class widgets:
         # checkbox for setting if tide files are compressed
         self.compress = ipywidgets.Checkbox(
             value=False,
-            description='Compressed?',
+            description="Compressed?",
             disabled=False,
             style=self.style,
         )
 
         # date picker widget for setting time
         self.datepick = ipywidgets.DatePicker(
-            description='Date:',
-            value = datetime.date.today(),
+            description="Date:",
+            value=datetime.date.today(),
             disabled=False,
             style=self.style,
         )
 
+
 # draw ipyleaflet map
 class leaflet:
-    def __init__(self, projection='Global', **kwargs):
+    def __init__(self, projection="Global", **kwargs):
         # set default keyword arguments
-        kwargs.setdefault('map',None)
-        kwargs.setdefault('attribution',True)
[…]

[…]
