### 2025-12-08 d3c7617
refactor: no longer subclassing `pathlib.Path`
that was a headache that required py3.12+
