+++ pyTMD/io/FES.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 """
 FES.py
-Written by Tyler Sutterley (11/2025)
+Written by Tyler Sutterley (12/2025)
 
 Reads ascii and netCDF4 files for FES tidal solutions provided by AVISO
     https://www.aviso.altimetry.fr/data/products/auxiliary-products/
@@ -15,6 +15,7 @@ PYTHON DEPENDENCIES:
         https://docs.xarray.dev/en/stable/
 
 UPDATE HISTORY:
+    Updated 12/2025: no longer subclassing pathlib.Path for working directories
     Updated 11/2025: near-complete rewrite of program to use xarray
     Updated 10/2025: simplify ascii read function to use masked_equal
     Updated 08/2025: use numpy degree to radian conversions
@@ -191,7 +192,7 @@ def open_fes_ascii(
     kwargs.setdefault("compressed", False)
     # tilde-expand input file
     input_file = pyTMD.utilities.Path(input_file).resolve()
-    if input_file.is_local() and not input_file.exists():
+    if isinstance(input_file, pathlib.Path) and not input_file.exists():
         raise FileNotFoundError(f"File not found: {input_file}")
     # read the ASCII-format tide elevation file
     if kwargs["compressed"]:
@@ -295,7 +296,7 @@ def open_fes_netcdf(
     kwargs.setdefault("compressed", False)
     # tilde-expand input file
     input_file = pyTMD.utilities.Path(input_file).resolve()
-    if input_file.is_local() and not input_file.exists():
+    if isinstance(input_file, pathlib.Path) and not input_file.exists():
         raise FileNotFoundError(f"File not found: {input_file}")
     # read the netCDF4-format tide elevation file
     if kwargs["compressed"]:
