+++ pyproject.toml
@@ -23,7 +23,7 @@ maintainers = [
 license = {file = "LICENSE"}
 readme = {file = "README.md", content-type = "text/markdown"}
 
-requires-python = "~=3.6"
+requires-python = "~=3.9"
 dependencies = [
     "h5netcdf",
     "lxml",
