### 2025-12-08 88099c3
refactor: no longer subclassing `pathlib.Path` (#494)
