+++ pyTMD/utilities.py
@@ -11,6 +11,8 @@ PYTHON DEPENDENCIES:
         https://pypi.org/project/platformdirs/
 
 UPDATE HISTORY:
+    Updated 12/2025: add URL class to build and operate on URLs
+        no longer subclassing pathlib.Path for working directories
     Updated 11/2025: added string check to determine if is a valid URL
         added function to check if a dependency is available
         added detection functions for compression and model format
@@ -94,9 +96,11 @@ __all__ = [
     "dependency_available",
     "is_valid_url",
     "Path",
+    "URL",
     "reify",
     "detect_format",
     "detect_compression",
+    "compressuser",
     "get_hash",
     "get_git_revision_hash",
     "get_git_status",
@@ -134,7 +138,7 @@ def get_data_path(relpath: list | str | pathlib.Path):
     """
     # current file path
     filename = inspect.getframeinfo(inspect.currentframe()).filename
-    filepath = Path(filename).absolute().parent
+    filepath = pathlib.Path(filename).absolute().parent
     if isinstance(relpath, list):
         # use *splat operator to extract from list
         return filepath.joinpath(*relpath)
@@ -163,7 +167,7 @@ def get_cache_path(
         filepath = filepath.joinpath(*relpath)
     elif isinstance(relpath, (str, pathlib.Path)):
         filepath = filepath.joinpath(relpath)
-    return Path(filepath)
+    return pathlib.Path(filepath)
 
 
 def import_dependency(
@@ -251,97 +255,126 @@ def is_valid_url(url: str) -> bool:
         return False
 
 
-class Path(pathlib.Path):
-    """``Pathlib.Path`` subclass with additional methods and properties"""
+def Path(filename: str | pathlib.Path, *args, **kwargs):
+    """
+    Create a ``URL`` or ``pathlib.Path`` object
 
-    def __init__(self, filename: str | pathlib.Path, *args, **kwargs):
-        if is_valid_url(filename):
-            self.filename = str(filename)
-            self._raw_paths = list(url_split(self.filename))
-        else:
-            super().__init__(filename)
-            self.filename = pathlib.Path(filename, *args, **kwargs)
+    Parameters
+    ----------
+    filename: str or pathlib.Path
+        file path or URL
+    """
+    if is_valid_url(filename):
+        return URL(filename, *args, **kwargs)
+    else:
+        return pathlib.Path(filename, *args, **kwargs)
 
-    def joinpath(self, *args):
-        """Join path components to the existing path"""
-        if self.is_url():
-            return Path("/".join([self.filename, *args]))
-        else:
-            return Path(self.filename.joinpath(*args))
 
-    # PURPOSE: platform independent file opener
-    def openfile(self):
-        """Platform independent file opener"""
-        if self.is_url():
-            raise RuntimeError("Cannot open URL paths")
-        if self.is_local() and (sys.platform == "win32"):
-            os.startfile(self.filename, "explore")
-        elif self.is_local() and (sys.platform == "darwin"):
-            subprocess.call(["open", self.filename])
-        else:
-            subprocess.call(["xdg-open", self.filename])
+class URL:
+    """Handles URLs similar to ``pathlib.Path`` objects"""
 
-    def compressuser(self):
-        """Tilde-compress a file to be relative to the home directory"""
-        if self.is_url():
-            raise RuntimeError("Cannot compress URL paths")
-        # attempt to compress filename relative to home directory
-        filename = pathlib.Path(self.filename).expanduser().absolute()
-        try:
-            relative_to = filename.relative_to(pathlib.Path().home())
-        except (ValueError, AttributeError) as exc:
-            return self.filename
-        else:
-            return Path("~").joinpath(relative_to)
+    def __init__(self, urlname: str | pathlib.Path, *args, **kwargs):
+        """Initialize a ``URL`` object"""
[…]
