### 2025-12-09 da34cef
chore: add grants to `.zenodo.json`
