### 2025-12-10 ccd8a37
fix: `expanduser` on local paths (#496)
refactor: use `notnull` to calculate masks
