### 2025-12-10 16ff857
docs: start adding transition guide
