### 2025-12-10 98bb6fe
Update Recipes.rst
