### 2025-12-10 9cb4580
docs: add plotting to transition guide
