### 2025-12-10 6767a9a
docs: add more blurbs for transition guide
