+++ doc/source/user_guide/Transition_Guide.rst
@@ -14,8 +14,8 @@ Key Changes
 ===========
 
 1. **xarray Integration**: utilizes ``xarray`` for reading, writing and operating on datasets. This improves our capability to handle multi-dimensional arrays.
-2. **Unit Support**: incorporates ``pint`` for unit management, allowing for more robust handling of physical quantities and unit conversion.
-3. **Database Restructuring**: The tidal model database has been restructured to flatten the hierarchy and (hopefully) improve accessibility.
+2. **Database Restructuring**: The tidal model database has been restructured to flatten the hierarchy and (hopefully) improve accessibility [see :ref:`restructure-json`].
+3. **Unit Support**: incorporates ``pint`` for unit management, allowing for more robust handling of physical quantities and unit conversion.
 4. **Function Renaming and Relocation**: Several functions have been renamed or moved to different modules.
 
 Examples
@@ -100,4 +100,63 @@ Examples
     ds.m2.tmd.phase.plot(ax=ax[1], cmap='hsv', vmin=0, vmax=360)
     # show plot
     plt.show()
+
+
+.. _restructure-json:
+
+Restructuring Definition Files
+==============================
+
+In version 2, the structure of the JSON database and definition files were separated into elevation and currents models.
+This was a legacy of the original design of the library (v1), which hardcoded the elevation and currents parameters.
+Model files in version 2 could be a string or list for elevation models, and a dictionary of strings or lists for currents models.
+In version 3, the database and example definition files have been flattened into a single structure  [see :ref:`definition-files`].
+In the new structure, each group (``'z'`` for elevation and ``'u'`` or ``'v'`` for currents) contains the model files and units.
+Units can also be extracted from files when reading the model data if stored as metadata attributes.
+
+Existing JSON files can be restructured to the new flattened format using a script provided in the ``test`` directory:
+
+.. code-block:: bash
+
+
+    python test/_restructure_json.py elevation_model.json
+
+    python test/_restructure_json.py elevation_model.json currents_model.json
+
+
+Units Support
+=============
+
+In version 2, units were only implicitly handled within ``pyTMD`` by setting scaling factors within the model definitions.
+This could lead to some confusion when working with different tide models that used different units.
+In version 3, units for different variables are explicitly handled using the ``pint`` library.
+When reading model data, units are extracted from the definition files or from attributes in the data files.
+The variables are then converted to default sets of units: 1) meters for elevation, 2) centimeters per second for currents and 3) meters squared per second for transports.
+Datasets can also be converted to different units using the ``to_units`` method.
+
+.. code-block:: python
+
+    # convert elevation dataset to millimeters
+    ds_z_mm = ds_z.tmd.to_units('mm')
+    # convert currents datasets to meters per second
+    ds_u_mps = ds_u.tmd.to_units('m/s')
+    ds_v_mps = ds_v.tmd.to_units('m/s')
+
+Removed Functions
+=================
+
+Each model format (``'ATLAS'``, ``'FES'``, ``'GOT'``, ``'OTIS'``, etc) previously had two functions for reading model constituents: 
+``extract_constants`` was used to read model data and interpolate to a specific location, and ``read_constants`` was used to read the full model grid.
+The ``constituents`` class stored the model constituents from ``read_constants`` and had some methods for calculating amplitude and phase.
+In version 3, these functions have been removed and replaced with a unified interface using ``xarray`` datasets.
+Model data can be read using the ``open_dataset`` method, which returns an ``xarray`` dataset containing all model constituents.
+Interpolation to specific locations is performed using the ``interp`` method.
+
+In version 2, tide prediction was be separated into ``drift``, ``map`` and ``time_series`` functions depending on the shape of the input data.
[…]
