### 2025-12-10 4c5e3ee
fix: copilot nitpicks
