### 2025-12-10 80c0a09
docs: add transition guide for v3 (#497)
