+++ doc/source/user_guide/Transition_Guide.rst
@@ -1,0 +1,162 @@
+=======================
+Transition Guide for v3
+=======================
+
+Overview
+========
+
+The transition from ``pyTMD`` version 2 to version 3 includes several changes to the
+API and functionality of the library. This document outlines the key differences
+between the two versions and provides guidance on how to adapt existing code to
+work with ``pyTMD`` v3.
+
+Key Changes
+===========
+
+1. **xarray Integration**: utilizes ``xarray`` for reading, writing and operating on datasets. This improves our capability to handle multi-dimensional arrays.
+2. **Database Restructuring**: The tidal model database has been restructured to flatten the hierarchy and (hopefully) improve accessibility [see :ref:`restructure-json`].
+3. **Unit Support**: incorporates ``pint`` for unit management, allowing for more robust handling of physical quantities and unit conversion.
+4. **Function Renaming and Relocation**: Several functions have been renamed or moved to different modules.
+
+Examples
+========
+
+**Example 1: Reading a Tidal Model and Predicting a Time Series**
+
+.. code-block:: python
+    :caption: Version 2
+
+    import pyTMD
+    # get model parameters for model
+    m = pyTMD.io.model(directory).elevation(model)
+    # read model and interpolate to location
+    amp, ph, c = m.extract_constants(lon, lat, extrapolate=True)
+    # calculate complex phase in radians for Euler's
+    cph = -1j*ph*np.pi/180.0
+    # calculate constituent oscillation
+    hc = amp*np.exp(cph)
+    # calculate tide values for time series
+    tide = pyTMD.predict.time_series(time, hc, c,
+        deltat=deltat, corrections=m.corrections)
+    # infer minor corrections and add to prediction
+    tide += pyTMD.predict.infer_minor(time, hc, c,
+        deltat=deltat, corrections=m.corrections)
+
+.. code-block:: python
+    :caption: Version 3
+
+    import pyTMD
+    # get model parameters for model
+    m = pyTMD.io.model(directory).from_database(model)
+    # read model as xarray Dataset
+    ds = m.open_dataset(group='z')
+    # interpolate to location
+    local = ds.tmd.interp(lon, lat, extrapolate=True)
+    # calculate tide values for time series
+    tide = local.tmd.predict(time, deltat=deltat,
+        corrections=m.corrections)
+    # infer minor corrections and add to prediction
+    tide += local.tmd.infer(time, deltat=deltat,
+        corrections=m.corrections)
+
+**Example 2: Plotting a Tidal Model**
+
+.. code-block:: python
+    :caption: Version 2
+
+    import pyTMD
+    import matplotlib.pyplot as plt
+    # get model parameters for model
+    m = pyTMD.io.model(directory).elevation(model)
+    # read model constituents
+    c = m.read_constants()
+    # get spatial bounds of model grid
+    xmin, xmax = c.coords.x.min(), c.coords.x.max()
+    ymin, ymax = c.coords.y.min(), c.coords.y.max()
+    extent = [xmin, xmax, ymin, ymax]
+    # plot model amplitude and phase
+    fig, ax = plt.subplots(ncols=2)
+    ax[0].imshow(c.amplitude('m2'), origin='lower',
+        extent=extent, interpolation='nearest',
+        cmap='viridis', vmin=0)
+    ax[1].imshow(c.phase('m2'), origin='lower',
+        extent=extent, interpolation='nearest',
+        cmap='hsv', vmin=0, vmax=360)
+    # show plot
+    plt.show()
+
+.. code-block:: python
+    :caption: Version 3
+
+    import pyTMD
+    import matplotlib.pyplot as plt
[…]
