### 2025-12-11 9ccdbf4
docs: add reference to v3 release notes
