+++ doc/source/release_notes/release-v2.0.0.rst
@@ -1,3 +1,5 @@
+.. _release-v2.0.0:
+
 ##################
 `Release v2.0.0`__
 ##################
