+++ doc/source/release_notes/release-v1.0.8.rst
@@ -1,3 +1,5 @@
+.. _release-v1.0.8:
+
 ##################
 `Release v1.0.8`__
 ##################
