+++ doc/source/release_notes/release-v1.0.6.rst
@@ -1,3 +1,5 @@
+.. _release-v1.0.6:
+
 ##################
 `Release v1.0.6`__
 ##################
