+++ doc/source/release_notes/release-v1.0.2.13.rst
@@ -1,3 +1,5 @@
+.. _release-v1.0.2.13:
+
 #####################
 `Release v1.0.2.13`__
 #####################
