+++ doc/source/release_notes/release-v3.0.0.rst
@@ -1,3 +1,5 @@
+.. _release-v3.0.0:
+
 ##################
 `Release v3.0.0`__
 ##################
