+++ doc/source/release_notes/release-v2.1.8.rst
@@ -1,3 +1,5 @@
+.. _release-v2.1.8:
+
 ##################
 `Release v2.1.8`__
 ##################
