### 2025-12-11 a02a5fd
docs: use moderate resolution coastilnes
