### 2025-12-11 94d3121
docs: add reference to v3 release notes (#498)
* docs: only use moderate resolution coastlines
