+++ doc/source/release_notes/release-v2.2.9.rst
@@ -1,3 +1,5 @@
+.. _release-v2.2.9:
+
 ##################
 `Release v2.2.9`__
 ##################
