+++ doc/source/release_notes/release-v2.1.6.rst
@@ -1,3 +1,5 @@
+.. _release-v2.1.6:
+
 ##################
 `Release v2.1.6`__
 ##################
