+++ doc/source/release_notes/release-v1.1.1.rst
@@ -1,3 +1,5 @@
+.. _release-v1.1.1:
+
 ##################
 `Release v1.1.1`__
 ##################
