+++ doc/source/release_notes/release-v2.2.6.rst
@@ -1,3 +1,5 @@
+.. _release-v2.2.6:
+
 ##################
 `Release v2.2.6`__
 ##################
