+++ doc/source/release_notes/release-v2.0.5.rst
@@ -1,3 +1,5 @@
+.. _release-v2.0.5:
+
 ##################
 `Release v2.0.5`__
 ##################
