### 2025-12-11 baca8e8
docs: add currents to transition guide
