### 2025-12-11 6de7c27
docs: add currents to transition guide (#499)
