### 2025-12-12 7e4e3bb
feat: add name, stem and parent to `URL` class
