### 2025-12-12 #8566 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
