### 2025-12-12 de44161
Update utilities.py
