### 2025-12-12 84c8359
Update utilities.py
