### 2025-12-12 a4807a0
feat: add name, stem and parent to `URL` class (#500)
