### 2025-12-12 f850547
docs: add reference to TMD2.5 (#501)
docs: fix line about interpolation of amplitude and phase to just tidal constituent
docs: standardize author bibtex entries
