+++ doc/source/_assets/paper.md
@@ -44,7 +44,7 @@ bibliography: pytmd-refs.bib
 
 # Statement of need
 
-There are several ocean tide prediction software options available. The OSU Tidal Inversion Software (`OTIS`) and OSU Tidal Prediction Software (`OTPS`) are Fortran programs developed for the [TPXO](https://www.tpxo.net/home) family of models [@Egbert:2002ge]. The Tidal Model Driver ([`TMD`](https://github.com/EarthAndSpaceResearch/TMD_Matlab_Toolbox_v2.5)) is a MATLAB Toolbox developed for the same family of models. An updated version of the MATLAB Toolbox ([`TMD3`](https://github.com/chadagreene/tide-model-driver)) was developed to use a custom consolidated netCDF4 format [@Greene:2024di]. [`pyFES`](https://cnes.github.io/aviso-fes/) is a Python library produced for the Finite Element Solution (FES) family of models [@Lyard:2025tr] funded by CNES. The NASA GSFC PREdict Tidal Heights (`PERTH3`) software is a Fortran program developed for the Goddard Ocean Tide (GOT) family of models. An updated and more versatile version of the NASA GSFC Fortran software ([`PERTH5`](https://codeberg.org/rray/perth5)) can read from multiple different tide model formats. These software options are typically created by or for the model providers, and, with the exception of `PERTH5`, singly support their specific model formats.  
+There are several ocean tide prediction software options available. The OSU Tidal Inversion Software (`OTIS`) and OSU Tidal Prediction Software (`OTPS`) are Fortran programs developed for the [TPXO](https://www.tpxo.net/home) family of models [@Egbert:2002ge]. The Tidal Model Driver ([`TMD`](https://github.com/EarthAndSpaceResearch/TMD_Matlab_Toolbox_v2.5)) is a MATLAB Toolbox developed for the same family of models [@Padman:2022fh]. An updated version of the MATLAB Toolbox ([`TMD3`](https://github.com/chadagreene/tide-model-driver)) was developed to use a custom consolidated netCDF4 format [@Greene:2024di]. [`pyFES`](https://cnes.github.io/aviso-fes/) is a Python library produced for the Finite Element Solution (FES) family of models [@Lyard:2025tr] funded by CNES. The NASA GSFC PREdict Tidal Heights (`PERTH3`) software is a Fortran program developed for the Goddard Ocean Tide (GOT) family of models. An updated and more versatile version of the NASA GSFC Fortran software ([`PERTH5`](https://codeberg.org/rray/perth5)) can read from multiple different tide model formats. These software options are typically created by or for the model providers, and, with the exception of `PERTH5`, singly support their specific model formats.  
 
 `pyTMD` is a generalized tide program that allows users to calculate both tide deflections and currents from a broad suite of models. Over 50 different models are presently supported, and additional model schemas can be defined with a JSON file. 
 
@@ -54,7 +54,7 @@ There are several ocean tide prediction software options available. The OSU Tida
 
 ## Ocean and Load Tides
 
[…]
