### 2025-12-12 59ceb5c
chore: update release notes for beta-3
