+++ CITATION.cff
@@ -13,7 +13,7 @@ keywords:
   - Tidal Prediction
 license: MIT
 version: "3.0.0"
-date-released: "2025-12-04"
+date-released: "2025-12-12"
 message: If you use this software, please consider citing our library.
 
 authors:
