### 2025-12-12 #8566 @kthyng
@tsutterley Also a couple of references aren't compiling at the top of page 2. Can you take a look?
