### 2025-12-13 1e6fc5a
docs: fixes for proof stage
thanks @kthyng!
