+++ doc/source/_assets/pytmd-refs.bib
@@ -37,7 +37,7 @@ pages = {4479},
 publisher = {MDPI AG},
 }
 @article{BishopTaylor:2025dc,
-author = {Bishop-Taylor, Robbi and Phillips, Claire and Sagar, Stephen and Newey, Vanessa and Sutterley, Tyler},
+author = {Bishop-Taylor, R and Phillips, C and Sagar, S and Newey, V and Sutterley, T},
 title = {{eo-tides: Tide modelling tools for large-scale satellite Earth observation analysis}},
 journal = {Journal of Open Source Software},
 year = {2025},
@@ -300,7 +300,7 @@ pages = {183--204},
 publisher = {American Meteorological Society},
 }
 @manual{ENVEO:2021jh,
-author = {{ENVEO} and Wuite, Jan and Hetzenecker, Markus and Nagler, Thomas and Scheiblauer, Stefan},
+author = {{ENVEO} and Wuite, J and Hetzenecker, M and Nagler, T and Scheiblauer, S},
 title = {{ESA Antarctic Ice Sheet Climate Change Initiative (Antarctic\_Ice\_Sheet\_cci): Antarctic Ice Sheet monthly velocity from 2017 to 2020, derived from Sentinel-1}},
 organization = {NERC EDS Centre for Environmental Data Analysis},
 year = {2021},
@@ -309,7 +309,7 @@ url = {https://doi.org/10.5285/00fe090efc58446e8980992a617f632f},
 doi = {10.5285/00fe090efc58446e8980992a617f632f},
 }
 @article{Fitzpatrick:2024iv,
-author = {Fitzpatrick, Sharon and Buscombe, Daniel and Warrick, Jonathan A and Lundine, Mark A and Vos, Kilian},
+author = {Fitzpatrick, S and Buscombe, D and Warrick, J A and Lundine, M A and Vos, K},
 title = {{CoastSeg: an accessible and extendable hub for satellite-derived-shoreline (SDS) detection and mapping}},
 journal = {Journal of Open Source Software},
 year = {2024},
@@ -395,7 +395,7 @@ publisher = {The Open Journal},
 url = {https://doi.org/10.21105/joss.06018},
 }
 @article{Gregg:2024ky,
-author = {Gregg, David E and Penna, Nigel T and Jones, Christopher and Maqueda, Miguel A},
+author = {Gregg, D E and Penna, N T and Jones, C and Maqueda, M A},
 title = {{Accuracy assessment of recent global ocean tide models in coastal waters of the European North West Shelf}},
 journal = {Ocean Modelling},
 year = {2024},
@@ -739,8 +739,16 @@ doi = {10.1002/2016RG000546},
 pages = {142--184},
 publisher = {American Geophysical Union (AGU)},
 }
+@software{Padman:2022fh,
+author = {Padman, L and Erofeeva, S and Howard, S L},
+title = {{Tide Model Driver (TMD) version 2.5, Toolbox for MATLAB (MATrix LABoratory) [Computer software]}},
+year = {2022},
+url = {https://arcticdata.io/catalog/view/doi:10.18739/A21Z41V08},
+doi = {10.18739/A21Z41V08},
+publisher = {NSF Arctic Data Center},
+}
 @article{Paprotny:2024et,
-author = {Paprotny, Dominik and Rhein, Belinda and Vousdoukas, Michalis I and Terefenko, Pawe{\l} and Dottori, Francesco and Treu, Simon and {\'S}ledziowski, Jakub and Feyen, Luc and Kreibich, Heidi},
[…]

[…]
