### 2025-12-15 e482387
docs: add citation to JOSS manuscript
