### 2025-12-15 #8566 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
